brms model wrapper

Allows users to specify a model using the bf wrapper from

brms_model(
  y = NULL,
  samples = NULL,
  horizon = NULL,
  model = NULL,
  n_cores = 1,
  n_chains = 4,
  n_iter = 2000,
  ...
)

Arguments

y
Numeric vector of time points to forecast

samples
Numeric, number of samples to take.

horizon
Numeric, the time horizon over which to predict.

model
A brms model wrapped in the bf function

n_cores
Numeric, the number of cores to use

n_chains
Numeric, the number of chains to use

n_iter
Numeric, the number of iterations in the sampler to use

...
additional arguments passed to brms (e.g. priors or family)

Value

A dataframe of predictions (with columns representing the time horizon and rows representing samples).

Examples

## Used on its own
library(brms)

#> Loading required package: Rcpp

#> Loading 'brms' package (version 2.12.0). Useful instructions
#> can be found by typing help('brms'). A more detailed introduction
#> to the package is available through vignette('brms_overview').

#> 
#> Attaching package: ‘brms’

#> The following object is masked from ‘package:stats’:
#> 
#>     ar

brms_model(y = EpiSoon::example_obs_rts[1:10, ]$rt,
          model = brms::bf(y ~ gp(time)),
          samples = 10, horizon = 7, n_iter = 40, n_chains = 1)

#> Compiling the C++ model

#> Start sampling

#> 
#> SAMPLING FOR MODEL '53361083d65bf2d26ff65f0961f9653c' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 6.9e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.69 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 15
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 40 [  2%]  (Warmup)
#> Chain 1: Iteration:  4 / 40 [ 10%]  (Warmup)
#> Chain 1: Iteration:  8 / 40 [ 20%]  (Warmup)
#> Chain 1: Iteration: 12 / 40 [ 30%]  (Warmup)
#> Chain 1: Iteration: 16 / 40 [ 40%]  (Warmup)
#> Chain 1: Iteration: 20 / 40 [ 50%]  (Warmup)
#> Chain 1: Iteration: 21 / 40 [ 52%]  (Sampling)
#> Chain 1: Iteration: 24 / 40 [ 60%]  (Sampling)
#> Chain 1: Iteration: 28 / 40 [ 70%]  (Sampling)
#> Chain 1: Iteration: 32 / 40 [ 80%]  (Sampling)
#> Chain 1: Iteration: 36 / 40 [ 90%]  (Sampling)
#> Chain 1: Iteration: 40 / 40 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.050044 seconds (Warm-up)
#> Chain 1:                0.001141 seconds (Sampling)
#> Chain 1:                0.051185 seconds (Total)
#> Chain 1: 

#> Warning: There were 15 divergent transitions after warmup. Increasing adapt_delta above 0.8 may help. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: The largest R-hat is 2.22, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#>          11       12       13       14       15       16       17
#> 1  2.017867 2.021494 2.036544 2.061383 2.093139 2.128197 2.163587
#> 2  2.016829 2.018320 2.030719 2.052307 2.081767 2.116560 2.154692
#> 3  2.016531 2.017850 2.029124 2.048271 2.073274 2.101734 2.131743
#> 4  2.035013 2.039439 2.057232 2.087508 2.127966 2.176985 2.230809
#> 5  2.033995 2.036189 2.049532 2.071446 2.099231 2.129988 2.160843
#> 6  2.028196 2.030763 2.044194 2.067588 2.099588 2.139005 2.184113
#> 7  2.024960 2.025681 2.035607 2.052955 2.076042 2.103622 2.134588
#> 8  2.029830 2.032020 2.044693 2.065442 2.092254 2.122041 2.152500
#> 9  2.030989 2.035468 2.051701 2.078502 2.113949 2.156632 2.204567
#> 10 2.028276 2.030172 2.043711 2.067194 2.099926 2.139172 2.183723

## Used for forecasting
## Note that the timeout parameter has been increased to allow
## for the time for the code to be compiled
forecast_rt(EpiSoon::example_obs_rts[1:10, ],
            model = function(...){
              brms_model(model = brms::bf(y ~ gp(time)), n_iter = 40, n_chains = 1, ...)},
                   horizon = 7, samples = 10, timeout = 300)

#> Compiling the C++ model

#> recompiling to avoid crashing R session

#> Start sampling

#> 
#> SAMPLING FOR MODEL '53361083d65bf2d26ff65f0961f9653c' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 4.9e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.49 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 15
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 40 [  2%]  (Warmup)
#> Chain 1: Iteration:  4 / 40 [ 10%]  (Warmup)
#> Chain 1: Iteration:  8 / 40 [ 20%]  (Warmup)
#> Chain 1: Iteration: 12 / 40 [ 30%]  (Warmup)
#> Chain 1: Iteration: 16 / 40 [ 40%]  (Warmup)
#> Chain 1: Iteration: 20 / 40 [ 50%]  (Warmup)
#> Chain 1: Iteration: 21 / 40 [ 52%]  (Sampling)
#> Chain 1: Iteration: 24 / 40 [ 60%]  (Sampling)
#> Chain 1: Iteration: 28 / 40 [ 70%]  (Sampling)
#> Chain 1: Iteration: 32 / 40 [ 80%]  (Sampling)
#> Chain 1: Iteration: 36 / 40 [ 90%]  (Sampling)
#> Chain 1: Iteration: 40 / 40 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.06833 seconds (Warm-up)
#> Chain 1:                0.066964 seconds (Sampling)
#> Chain 1:                0.135294 seconds (Total)
#> Chain 1: 

#> Warning: There were 7 divergent transitions after warmup. Increasing adapt_delta above 0.8 may help. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: The largest R-hat is 1.51, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> # A tibble: 70 x 4
#>    sample date          rt horizon
#>     <int> <date>     <dbl>   <int>
#>  1      1 2020-03-11  2.00       1
#>  2      2 2020-03-11  2.02       1
#>  3      3 2020-03-11  2.04       1
#>  4      4 2020-03-11  2.01       1
#>  5      5 2020-03-11  2.01       1
#>  6      6 2020-03-11  2.04       1
#>  7      7 2020-03-11  2.07       1
#>  8      8 2020-03-11  2.12       1
#>  9      9 2020-03-11  2.03       1
#> 10     10 2020-03-11  2.09       1
#> # … with 60 more rows