BSTS model wrapper

BSTS model wrapper

bsts_model(y = NULL, samples = NULL, horizon = NULL, model = NULL)

Arguments

y
Numeric vector of time points to forecast

samples
Numeric, number of samples to take.

horizon
Numeric, the time horizon over which to predict.

model
A bsts model object wrapped in a function with an ss and y argument.

Value

A dataframe of predictions (with columns representing the time horizon and rows representing samples).

Examples

## Used on its own
bsts_model(y = EpiSoon::example_obs_rts[1:10, ]$rt,
           model = function(ss, y){
           bsts::AddAr(ss, y = y, lags = 2)},
           samples = 10, horizon = 7)

#>          V1       V2       V3       V4       V5       V6       V7
#> 1  1.978836 1.912334 1.855109 1.768865 1.753945 1.707922 1.686209
#> 2  1.927865 1.950344 1.859923 1.777695 1.741134 1.721496 1.667161
#> 3  1.962695 1.889590 1.838024 1.815644 1.757994 1.721021 1.691455
#> 4  1.981115 1.894634 1.849228 1.779481 1.745365 1.675597 1.680467
#> 5  1.930705 1.921930 1.843386 1.803190 1.784410 1.706881 1.657857
#> 6  1.930400 1.887915 1.810533 1.769233 1.759128 1.689534 1.668091
#> 7  1.962564 1.858163 1.857333 1.798905 1.720588 1.718295 1.669056
#> 8  1.933069 1.861758 1.857774 1.746918 1.765273 1.725499 1.715009
#> 9  1.971430 1.882372 1.861836 1.811739 1.763136 1.724746 1.660507
#> 10 1.907581 1.895718 1.856984 1.795226 1.721012 1.689042 1.651312

## Used for forecasting
 forecast_rt(EpiSoon::example_obs_rts[1:10, ],
                    model = function(...){EpiSoon::bsts_model(model =
                      function(ss, y){
                        bsts::AddAr(ss, y = y, lags = 3)}, ...)},
                    horizon = 7, samples = 10)

#> # A tibble: 70 x 4
#>    sample date          rt horizon
#>     <int> <date>     <dbl>   <int>
#>  1      1 2020-03-11  1.94       1
#>  2      2 2020-03-11  1.98       1
#>  3      3 2020-03-11  1.98       1
#>  4      4 2020-03-11  1.95       1
#>  5      5 2020-03-11  1.97       1
#>  6      6 2020-03-11  1.93       1
#>  7      7 2020-03-11  1.91       1
#>  8      8 2020-03-11  2.03       1
#>  9      9 2020-03-11  1.97       1
#> 10     10 2020-03-11  1.93       1
#> # … with 60 more rows