Compare forecasting models

Compare forecasting models

compare_models(
  observations = NULL,
  models = NULL,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE
)

Arguments

observations
A dataframe of observations against which to score. Should contain a date and rt column.

models
A list of bsts models. An example configuration is given in the examples. Each bsts model needs to be wrapped in a function that takes a ss and y argument (i.e. function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}).

horizon
Numeric, the time horizon over which to predict

samples
Numeric, the number of samples to from the posterior of the model fit. Note that twice this number of MCMC steps will be taken with half used as burn in.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

Value

A list of dataframes as produced by evaluate model but with an additional model column.

Examples

## Dummy data
observations <- data.frame(rt = 1:20,
                            date = as.Date("2020-01-01")
                           + lubridate::days(1:20))

## List of forecasting bsts models wrapped in functions.
models <- list("Sparse AR" = function(ss, y){bsts::AddAutoAr(ss, y = y, lags = 7)},
               "Semi-local linear trend" = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)})

## Compare models
evaluations <- compare_models(observations, models, horizon = 7, samples = 10)

## Example evaluation plot for comparing forecasts
## with actuals for a range of models and time horizons.
plot_forecast_evaluation(evaluations$forecast, observations, c(1, 3, 7)) +
  ggplot2::facet_grid(model ~ horizon) +
  cowplot::panel_border()

#> Error in ggplot2::geom_line(data = observations, ggplot2::aes(y = rt),     size = 1.1): object 'observations' not found