Compare timeseries and forecast models

Compare timeseries and forecast models

compare_timeseries(
  timeseries = NULL,
  models = NULL,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE
)

Arguments

timeseries
A list of named timeseries with each one having a rt vector and a date vector.

models
A list of bsts models. An example configuration is given in the examples. Each bsts model needs to be wrapped in a function that takes a ss and y argument (i.e. function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}).

horizon
Numeric, the time horizon over which to predict

samples
Numeric, the number of samples to from the posterior of the model fit. Note that twice this number of MCMC steps will be taken with half used as burn in.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

Value

A list of dataframes as produced by evaluate model but with an additional model column.

Examples

## Dummy data
observations <- data.frame(rt = 1:20,
                            date = as.Date("2020-01-01")
                           + lubridate::days(1:20))

timeseries <- list(observations, observations)
names(timeseries) <- c("Region 1", "Region 2")

## List of forecasting bsts models wrapped in functions.
models <- list("Sparse AR" = function(ss, y){bsts::AddAutoAr(ss, y = y, lags = 7)},
               "Semi-local linear trend" = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)})

## Compare models
evaluations <- compare_timeseries(timeseries, models,
                                  horizon = 7, samples = 10)

evaluations

#> $forecasts
#> # A tibble: 504 x 12
#>    region model forecast_date date       horizon bottom lower median  mean upper
#>    <chr>  <chr> <chr>         <date>       <int>  <dbl> <dbl>  <dbl> <dbl> <dbl>
#>  1 Regio… Spar… 2020-01-04    2020-01-05       1      0 0      0.222 1.32  2.20 
#>  2 Regio… Spar… 2020-01-04    2020-01-06       2      0 0      1.50  2.43  3.65 
#>  3 Regio… Spar… 2020-01-04    2020-01-07       3      0 0      0.185 1.36  1.45 
#>  4 Regio… Spar… 2020-01-04    2020-01-08       4      0 0      0.139 0.690 1.05 
#>  5 Regio… Spar… 2020-01-04    2020-01-09       5      0 0      1.18  4.39  1.81 
#>  6 Regio… Spar… 2020-01-04    2020-01-10       6      0 0.130  1.37  2.92  5.49 
#>  7 Regio… Spar… 2020-01-04    2020-01-11       7      0 0      0     0.928 0.978
#>  8 Regio… Spar… 2020-01-05    2020-01-06       1      0 0      0.257 1.75  2.61 
#>  9 Regio… Spar… 2020-01-05    2020-01-07       2      0 0      1.65  1.97  3.56 
#> 10 Regio… Spar… 2020-01-05    2020-01-08       3      0 0      0.687 2.01  2.95 
#> # … with 494 more rows, and 2 more variables: top <dbl>, sd <dbl>
#> 
#> $scores
#> # A tibble: 392 x 10
#>    region model forecast_date date       horizon   dss  crps  logs  bias
#>    <chr>  <chr> <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>
#>  1 Regio… Spar… 2020-01-04    2020-01-05       1  3.49  2.02  2.63 0.100
#>  2 Regio… Spar… 2020-01-04    2020-01-06       2  2.90  1.99  2.75 0.200
#>  3 Regio… Spar… 2020-01-04    2020-01-07       3  5.16  4.14  9.91 0.100
#>  4 Regio… Spar… 2020-01-04    2020-01-08       4 42.1   5.84 29.9  0    
#>  5 Regio… Spar… 2020-01-04    2020-01-09       5  4.64  5.03  6.37 0.100
#>  6 Regio… Spar… 2020-01-04    2020-01-10       6  5.55  4.36  3.15 0.100
#>  7 Regio… Spar… 2020-01-04    2020-01-11       7 25.1   8.33 33.0  0    
#>  8 Regio… Spar… 2020-01-05    2020-01-06       1  3.44  2.78  3.58 0.100
#>  9 Regio… Spar… 2020-01-05    2020-01-07       2  5.38  2.92  2.86 0    
#> 10 Regio… Spar… 2020-01-05    2020-01-08       3  5.07  4.18  4.86 0.100
#> # … with 382 more rows, and 1 more variable: sharpness <dbl>
#> 

## Example evaluation plot for comparing forecasts
## with actuals for a range of models and timeseries.
plot_forecast_evaluation(evaluations$forecast, observations, c(7)) +
  ggplot2::facet_grid(model ~ region) +
  cowplot::panel_border()

#> Error in ggplot2::geom_line(data = observations, ggplot2::aes(y = rt),     size = 1.1): object 'observations' not found