Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  observations = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE
)

Arguments

observations
A dataframe of observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

model
An uninitialised bsts model

horizon
Numeric, the time horizon over which to predict

samples
Numeric, the number of samples to from the posterior of the model fit. Note that twice this number of MCMC steps will be taken with half used as burn in.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

Value

Examples

## Observed data
observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Evaluate a model
evaluate_model(observations,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10)

#> $forecasts
#> # A tibble: 126 x 10
#>    forecast_date date       horizon bottom lower median  mean upper   top    sd
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 2020-01-04    2020-01-05       1   2.76  3.12   3.20  3.31  3.57  3.91 0.384
#>  2 2020-01-04    2020-01-06       2   2.22  3.36   3.85  3.64  4.20  4.43 0.768
#>  3 2020-01-04    2020-01-07       3   2.97  3.86   4.72  4.62  5.10  6.68 1.19 
#>  4 2020-01-04    2020-01-08       4   3.85  4.65   5.36  5.41  5.85  7.57 1.19 
#>  5 2020-01-04    2020-01-09       5   3.70  5.54   6.19  6.10  6.97  8.31 1.48 
#>  6 2020-01-04    2020-01-10       6   4.36  6.03   6.80  7.03  7.99  9.92 1.75 
#>  7 2020-01-04    2020-01-11       7   4.39  6.56   7.53  7.67  9.20 10.7  2.11 
#>  8 2020-01-05    2020-01-06       1   4.71  4.86   4.95  5.05  5.24  5.52 0.277
#>  9 2020-01-05    2020-01-07       2   5.73  6.16   6.38  6.27  6.49  6.57 0.298
#> 10 2020-01-05    2020-01-08       3   6.59  6.95   7.26  7.25  7.58  7.89 0.447
#> # … with 116 more rows
#> 
#> $scores
#> # A tibble: 98 x 8
#>    forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-01-04    2020-01-05       1  1.62 0.497   1.05  0         0.166
#>  2 2020-01-04    2020-01-06       2  2.84 0.965   2.32  0         0.640
#>  3 2020-01-04    2020-01-07       3  1.74 0.985   2.22  0.100     0.827
#>  4 2020-01-04    2020-01-08       4  2.21 1.16    2.58  0.100     0.985
#>  5 2020-01-04    2020-01-09       5  2.51 1.23    2.07  0.100     1.27 
#>  6 2020-01-04    2020-01-10       6  2.43 1.34    2.29  0.100     1.45 
#>  7 2020-01-04    2020-01-11       7  2.74 1.38    2.15  0.100     2.41 
#>  8 2020-01-05    2020-01-06       1 -2.64 0.0717 -0.187 0.4       0.265
#>  9 2020-01-05    2020-01-07       2 -1.65 0.206   0.631 0.8       0.254
#> 10 2020-01-05    2020-01-08       3 -1.35 0.161   0.406 0.7       0.535
#> # … with 88 more rows
#> 

## Samples of observed data
sampled_obs <- observations %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(observations %>%
       dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10)

#> $forecasts
#> # A tibble: 126 x 10
#>    forecast_date date       horizon bottom lower median  mean upper   top    sd
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 2020-01-04    2020-01-05       1   2.98  3.44   3.72  3.81  4.10  4.75 0.517
#>  2 2020-01-04    2020-01-06       2   4.15  4.47   4.70  4.89  5.24  6.13 0.604
#>  3 2020-01-04    2020-01-07       3   4.82  5.46   5.78  5.93  6.06  7.78 0.832
#>  4 2020-01-04    2020-01-08       4   5.90  6.37   6.72  6.94  6.99  9.03 0.921
#>  5 2020-01-04    2020-01-09       5   6.80  7.14   7.41  7.85  8.23 10.3  1.01 
#>  6 2020-01-04    2020-01-10       6   7.45  7.99   8.58  8.90  9.50 11.4  1.17 
#>  7 2020-01-04    2020-01-11       7   8.18  8.88   9.24  9.80 10.5  12.5  1.27 
#>  8 2020-01-05    2020-01-06       1   4.50  4.84   5.04  5.24  5.65  6.39 0.626
#>  9 2020-01-05    2020-01-07       2   4.96  5.78   6.20  6.16  6.57  7.10 0.631
#> 10 2020-01-05    2020-01-08       3   5.39  6.76   7.10  7.29  7.71  9.29 1.10 
#> # … with 116 more rows
#> 
#> $scores
#> # A tibble: 196 x 9
#>    sample forecast_date date       horizon     dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int>   <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-01-04    2020-01-05       1 -0.821  0.274 1.26  0.200     0.309
#>  2 1      2020-01-04    2020-01-06       2 -0.655  0.302 2.01  0.200     0.285
#>  3 1      2020-01-04    2020-01-07       3  0.0435 0.252 1.08  0.3       1.13 
#>  4 1      2020-01-04    2020-01-08       4  0.112  0.235 1.06  0.3       0.870
#>  5 1      2020-01-04    2020-01-09       5  0.184  0.304 1.18  0.5       1.09 
#>  6 1      2020-01-04    2020-01-10       6  0.482  0.387 1.46  0.5       1.54 
#>  7 1      2020-01-04    2020-01-11       7  0.516  0.410 1.51  0.5       1.44 
#>  8 1      2020-01-05    2020-01-06       1 -0.844  0.165 0.626 0.4       0.470
#>  9 1      2020-01-05    2020-01-07       2 -0.846  0.193 0.719 0.3       0.628
#> 10 1      2020-01-05    2020-01-08       3  0.419  0.303 1.15  0.5       0.830
#> # … with 186 more rows
#>