Fit and Forecast using a BSTS Model

Fit and Forecast using a BSTS Model

forecast_rt(
  rts,
  model = model,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE,
  timeout = 100
)

Arguments

rts
A dataframe of containing two variables rt and date with rt being numeric and date being a date.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

Value

A dataframe of samples containing the following variables: sample, date, rt, and horizon.

Examples

forecast_rt(EpiSoon::example_obs_rts[1:10, ],
            model = function(...){
            EpiSoon::bsts_model(model = function(ss, y){
            bsts::AddAutoAr(ss, y = y, lags = 10)}, ...)
            },
            horizon = 7, samples = 10)

#> # A tibble: 70 x 4
#>    sample date          rt horizon
#>     <int> <date>     <dbl>   <int>
#>  1      1 2020-03-11  0          1
#>  2      2 2020-03-11  2.03       1
#>  3      3 2020-03-11  2.13       1
#>  4      4 2020-03-11  1.91       1
#>  5      5 2020-03-11  1.98       1
#>  6      6 2020-03-11  1.97       1
#>  7      7 2020-03-11  1.97       1
#>  8      8 2020-03-11  2.05       1
#>  9      9 2020-03-11  2.10       1
#> 10     10 2020-03-11  1.91       1
#> # … with 60 more rows