Iteratively forecast cases using a iterative Rt forecasts

Iteratively forecast cases using a iterative Rt forecasts

iteractive_case_forecast(
  it_fit_samples = NULL,
  cases = NULL,
  serial_interval,
  rdist = NULL
)

Arguments

it_fit_samples
Dataframe of iterative forecasts as produced by iterative_rt_forecast.

cases
A dataframe containing date and cases variables

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

Value

A dataframe of iterative case forecasts

Examples

## Iterative Rt forecast
it_forecast <-
  iterative_rt_forecast(EpiSoon::example_obs_rts,
                        model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                        horizon = 7, samples = 10)

## Iterative case forecast
iteractive_case_forecast(it_fit_samples = it_forecast,
                         cases = EpiSoon::example_obs_cases,
                         serial_interval = EpiSoon::example_serial_interval)

#> # A tibble: 1,330 x 5
#>    forecast_date sample date       cases horizon
#>    <chr>          <dbl> <date>     <int>   <int>
#>  1 2020-03-04         1 2020-03-05    60       1
#>  2 2020-03-04         1 2020-03-06    73       2
#>  3 2020-03-04         1 2020-03-07   101       3
#>  4 2020-03-04         1 2020-03-08   106       4
#>  5 2020-03-04         1 2020-03-09   119       5
#>  6 2020-03-04         1 2020-03-10   172       6
#>  7 2020-03-04         1 2020-03-11   155       7
#>  8 2020-03-04         2 2020-03-05    64       1
#>  9 2020-03-04         2 2020-03-06    57       2
#> 10 2020-03-04         2 2020-03-07    80       3
#> # … with 1,320 more rows