Predict cases for Rts based on observed data

Predict cases for Rts based on observed data

predict_current_cases(
  cases = NULL,
  rts = NULL,
  serial_interval = NULL,
  rdist = NULL
)

Arguments

cases
A dataframe containing date and cases variables

rts
A dataframe of containing two variables rt and date with rt being numeric and date being a date.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

Value

Forecast cases for the current timestep

Examples

predict_current_cases(cases = EpiSoon::example_obs_cases,
                      rts = EpiSoon::example_obs_rts,
                      serial_interval = EpiSoon::example_serial_interval)

#>          rt       date infectiousness cases
#> 1  2.490547 2020-03-01         14.145    30
#> 2  2.442588 2020-03-02         17.355    40
#> 3  2.402473 2020-03-03         21.145    54
#> 4  2.335572 2020-03-04         25.515    65
#> 5  2.266551 2020-03-05         30.875    81
#> 6  2.192293 2020-03-06         37.220    87
#> 7  2.146429 2020-03-07         44.195    98
#> 8  2.104371 2020-03-08         52.205   118
#> 9  2.059281 2020-03-09         62.125   137
#> 10 2.027134 2020-03-10         73.060   166
#> 11 2.014678 2020-03-11         84.975   162
#> 12 1.998946 2020-03-12        100.695   196
#> 13 1.968350 2020-03-13        119.065   247
#> 14 1.947376 2020-03-14        138.675   258
#> 15 1.906984 2020-03-15        158.155   277
#> 16 1.812842 2020-03-16        183.825   350
#> 17 1.718532 2020-03-17        208.025   352
#> 18 1.665646 2020-03-18        223.985   390
#> 19 1.639927 2020-03-19        246.015   415
#> 20 1.633795 2020-03-20        276.575   405
#> 21 1.682025 2020-03-21        309.540   519
#> 22 1.561653 2020-03-22        349.995   564