Score a case forecast

Score a case forecast

score_case_forecast(pred_cases, obs_cases, scores = "all")

Arguments

pred_cases
Dataframe of predicted cases with the following variables: sample, date, cases and forecast horizon. As produced by forecast_cases.

obs_cases
Dataframe of observed cases with the following variables: date and cases.

scores
Character vector defaulting to "all". Select which scores to return, default is all scores but any subset can be returned.

Value

A dataframe containing the following scores per forecast timepoint: dss, crps, logs, bias, and sharpness as well as the forecast date and time horizon.

Examples

## Fit a model (using a subset of observations)
samples <- forecast_rt(EpiSoon::example_obs_rts[1:10, ],
                     model = function(...) {EpiSoon::bsts_model(model =
                     function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
                     horizon = 7, samples = 10)

pred_cases <- forecast_cases(EpiSoon::example_obs_cases,
                             samples, EpiSoon::example_serial_interval)

## Score the model fit (with observations during the time horizon of the forecast)
score_case_forecast(pred_cases, EpiSoon::example_obs_cases)

#> # A tibble: 7 x 11
#>   date       horizon   dss  crps  logs   bias sharpness calibration median   iqr
#>   <date>       <int> <dbl> <dbl> <dbl>  <dbl>     <dbl>       <dbl>  <dbl> <dbl>
#> 1 2020-03-11       1  4.56  1.59  2.92  0.3        5.19   0.0000857      1   8  
#> 2 2020-03-12       2  4.42  3     3.40 -0.200     11.1    0.0000857      8  13.5
#> 3 2020-03-13       3  5.92  7.49  4.17  0.200     19.3    0.0000857     28  29  
#> 4 2020-03-14       4  6.19  7.14  4.39  0.200     28.2    0.0000857     13  40.8
#> 5 2020-03-15       5  4.99  3.67  3.70 -0.200     11.1    0.0000857      6  21  
#> 6 2020-03-16       6  9.43 33.8   5.29  1         42.3    0.0000857    116 152. 
#> 7 2020-03-17       7 10.2  46.4   5.63  1         43.0    0.0000857    139 207. 
#> # … with 1 more variable: ci <dbl>

## Score the model fit (with observations during the time horizon of the forecast)
score_case_forecast(pred_cases, EpiSoon::example_obs_cases, scores = c("crps", "sharpness", "bias"))

#> # A tibble: 7 x 5
#>   date       horizon  crps   bias sharpness
#>   <date>       <int> <dbl>  <dbl>     <dbl>
#> 1 2020-03-11       1  1.59  0.3        5.19
#> 2 2020-03-12       2  3    -0.200     11.1 
#> 3 2020-03-13       3  7.49  0.200     19.3 
#> 4 2020-03-14       4  7.14  0.200     28.2 
#> 5 2020-03-15       5  3.67 -0.200     11.1 
#> 6 2020-03-16       6 33.8   1         42.3 
#> 7 2020-03-17       7 46.4   1         43.0