Score a Model Fit

Score a Model Fit

score_forecast(fit_samples, observations)

Arguments

fit_samples
A dataframe as produced by EpiSoon::forecast.

observations
A dataframe of observations against which to score. Should contain a date and rt column.

Value

A dataframe containing the following scores per forecast timepoint: dss, crps, logs, bias, and sharpness as well as the forecast date and time horizon.

Examples

## Fit a model (using a subset of observations)
samples <- forecast_rt(EpiSoon::example_obs_rts[1:10, ],
                     model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                     horizon = 7, samples = 10)

## Score the model fit (with observations during the time horizon of the forecast)
score_forecast(samples, EpiSoon::example_obs_rts)

#> # A tibble: 7 x 7
#>   date       horizon   dss   crps   logs  bias sharpness
#>   <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#> 1 2020-03-11       1 -5.62 0.0260 -1.41  0.100    0.0254
#> 2 2020-03-12       2 -2.88 0.0407 -1.25  0        0.0395
#> 3 2020-03-13       3 -4.35 0.0440 -1.03  0.200    0.0385
#> 4 2020-03-14       4 -2.11 0.0617 -0.805 0        0.0412
#> 5 2020-03-15       5 -3.49 0.0571 -1.09  0.100    0.0699
#> 6 2020-03-16       6 -5.12 0.0250 -1.61  0.200    0.0502
#> 7 2020-03-17       7 -4.79 0.0223 -1.22  0.4      0.108