Score a Model Fit

Score a Model Fit

score_model(fit_samples, observations)

Arguments

fit_samples
A dataframe as produced by EpiSoon::fit_model.

observations
A dataframe of observations against which to score. Should contain a date and rt column.

Value

A dataframe containing the following scores per forecast timepoint: dss, crps, logs, bias, and sharpness as well as the forecast date and time horizon.

Examples

## Observed data
observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Fit a model (using a subset of observations)
samples <- fit_model(observations[1:10, ],
                     model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                     horizon = 7, samples = 10)

## Score the model fit (with observations during the time horizon of the forecast)
score_model(samples, observations)

#> # A tibble: 7 x 7
#>   date       horizon    dss   crps    logs  bias sharpness
#>   <date>       <int>  <dbl>  <dbl>   <dbl> <dbl>     <dbl>
#> 1 2020-01-12       1 -1.98  0.109  -0.0200 0.3       0.270
#> 2 2020-01-13       2 -2.40  0.102  -0.224  0.100     0.168
#> 3 2020-01-14       3 -2.12  0.0976 -0.127  0.3       0.216
#> 4 2020-01-15       4 -2.14  0.159   0.321  0.100     0.216
#> 5 2020-01-16       5 -1.30  0.127   0.176  0.200     0.150
#> 6 2020-01-17       6 -1.35  0.166   0.429  0.3       0.335
#> 7 2020-01-18       7 -0.751 0.232   0.715  0.3       0.170