Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 18 x 9
#>    score   model      bottom    lower   median     mean   upper     top       sd
#>    <chr>   <chr>       <dbl>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1 bias    AR 3     -1.00e+0 -8.00e-1 -6.00e-1  -0.559  -0.400    0       0.259 
#>  2 bias    Semi-lo… -1.00e+0 -8.00e-1 -4.00e-1  -0.211   0.200    1       0.681 
#>  3 calibr… AR 3      8.57e-5  8.57e-5  6.43e-3   0.0313  0.0234   0.190   0.0766
#>  4 calibr… Semi-lo…  8.57e-5  8.57e-5  1.20e-4   0.0990  0.144    0.907   0.209 
#>  5 ci      AR 3      1.91e+0  2.84e+0  3.81e+0   5.21    4.94    22.7     5.79  
#>  6 ci      Semi-lo…  6.82e-2  2.07e-1  3.81e-1   0.988   1.07     4.79    1.31  
#>  7 crps    AR 3      3.87e-1  6.90e-1  8.38e-1   0.890   1.08     1.59    0.298 
#>  8 crps    Semi-lo…  8.54e-3  2.24e-2  4.63e-2   0.0612  0.0883   0.164   0.0468
#>  9 dss     AR 3      5.10e-1  9.79e-1  1.29e+0   2.03    1.77     9.91    3.28  
#> 10 dss     Semi-lo… -7.11e+0 -4.96e+0 -3.70e+0  -2.23   -1.73    12.8     5.58  
#> 11 iqr     AR 3      1.69e+0  2.38e+0  3.39e+0   3.84    5.09     8.06    1.82  
#> 12 iqr     Semi-lo…  3.39e-2  1.06e-1  2.07e-1   0.269   0.391    0.723   0.212 
#> 13 logs    AR 3      1.29e+0  1.75e+0  1.94e+0 Inf       2.42   Inf     Inf     
#> 14 logs    Semi-lo… -2.47e+0 -1.54e+0 -9.33e-1   0.634  -0.0848  10.9     6.88  
#> 15 median  AR 3      4.16e-1  2.42e+0  3.27e+0   2.95    3.76     4.29    1.04  
#> 16 median  Semi-lo…  9.08e-3  5.22e-2  1.32e-1   0.160   0.247    0.411   0.129 
#> 17 sharpn… AR 3      0.       0.       2.88e-1   0.509   0.900    2.07    0.610 
#> 18 sharpn… Semi-lo…  1.46e-2  3.49e-2  5.53e-2   0.0708  0.0889   0.261   0.0574

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model           bottom  lower median   mean  upper   top     sd
#>      <int> <chr> <chr>            <dbl>  <dbl>  <dbl>  <dbl>  <dbl> <dbl>  <dbl>
#>  1       1 crps  AR 3           0.423   0.709  0.890  0.971  1.31   1.67  0.359 
#>  2       1 crps  Semi-local li… 0.00517 0.0122 0.0213 0.0292 0.0309 0.103 0.0271
#>  3       2 crps  AR 3           0.401   0.670  0.838  0.870  1.08   1.37  0.290 
#>  4       2 crps  Semi-local li… 0.00910 0.0176 0.0288 0.0420 0.0598 0.120 0.0324
#>  5       3 crps  AR 3           0.365   0.676  0.824  0.902  1.10   1.67  0.351 
#>  6       3 crps  Semi-local li… 0.0132  0.0188 0.0457 0.0599 0.0920 0.144 0.0443
#>  7       4 crps  AR 3           0.376   0.711  0.809  0.891  1.09   1.55  0.328 
#>  8       4 crps  Semi-local li… 0.0197  0.0370 0.0579 0.0742 0.105  0.178 0.0517
#>  9       5 crps  AR 3           0.423   0.712  0.869  0.857  1.04   1.22  0.237 
#> 10       5 crps  Semi-local li… 0.0212  0.0441 0.0828 0.0847 0.117  0.176 0.0479
#> 11       6 crps  AR 3           0.577   0.722  0.909  0.875  1.01   1.13  0.173 
#> 12       6 crps  Semi-local li… 0.0209  0.0387 0.0589 0.0762 0.104  0.167 0.0460
#> 13       7 crps  AR 3           0.440   0.626  0.784  0.832  1.06   1.32  0.271 
#> 14       7 crps  Semi-local li… 0.0213  0.0360 0.0656 0.0780 0.107  0.187 0.0518

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model            bottom lower median   mean upper   top    sd
#>   <chr>      <chr> <chr>             <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 3               5.67  6.57   7.10 Inf     8.72 Inf   Inf  
#> 2 Region 1   logs  Semi-local line…   3.47  4.27   5.01   8.86  8.16  34.9  14.6
#> 3 Region 2   logs  AR 3               5.86  6.83   7.34 Inf     8.63 Inf   Inf  
#> 4 Region 2   logs  Semi-local line…   3.29  4.00   4.68   8.27  6.77  38.5  11.3