BSTS model wrapper

BSTS model wrapper

bsts_model(y = NULL, samples = NULL, horizon = NULL, model = NULL)

Arguments

y
Numeric vector of time points to forecast

samples
Numeric, number of samples to take.

horizon
Numeric, the time horizon over which to predict.

model
A bsts model object wrapped in a function with an ss and y argument.

Value

A dataframe of predictions (with columns representing the time horizon and rows representing samples).

Examples

## Used on its own
bsts_model(y = EpiSoon::example_obs_rts[1:10, ]$rt,
           model = function(ss, y){
           bsts::AddAr(ss, y = y, lags = 2)},
           samples = 10, horizon = 7)

#>          V1       V2       V3       V4       V5       V6       V7
#> 1  1.939753 1.886331 1.830687 1.813617 1.773919 1.748181 1.664778
#> 2  1.950571 1.914041 1.856184 1.815465 1.775030 1.741579 1.694284
#> 3  1.928800 1.939337 1.828777 1.751670 1.779120 1.686248 1.661204
#> 4  1.923604 1.895852 1.874197 1.799498 1.786664 1.737384 1.672588
#> 5  1.946035 1.960296 1.865433 1.830409 1.756605 1.738047 1.737828
#> 6  1.957777 1.908921 1.865521 1.811248 1.786993 1.753433 1.703787
#> 7  1.950992 1.895699 1.857558 1.810138 1.767719 1.719569 1.688735
#> 8  1.958121 1.909698 1.869564 1.819371 1.773408 1.719430 1.667307
#> 9  1.959800 1.906915 1.881088 1.832100 1.786182 1.727494 1.666008
#> 10 1.957397 1.926289 1.894480 1.783016 1.765466 1.733608 1.692860

## Used for forecasting
 forecast_rt(EpiSoon::example_obs_rts[1:10, ],
                    model = function(...){EpiSoon::bsts_model(model =
                      function(ss, y){
                        bsts::AddAr(ss, y = y, lags = 3)}, ...)},
                    horizon = 7, samples = 10)

#> # A tibble: 70 x 4
#>    sample date          rt horizon
#>     <int> <date>     <dbl>   <int>
#>  1      1 2020-03-11  1.96       1
#>  2      2 2020-03-11  1.98       1
#>  3      3 2020-03-11  1.97       1
#>  4      4 2020-03-11  1.98       1
#>  5      5 2020-03-11  1.96       1
#>  6      6 2020-03-11  1.97       1
#>  7      7 2020-03-11  1.94       1
#>  8      8 2020-03-11  1.96       1
#>  9      9 2020-03-11  1.96       1
#> 10     10 2020-03-11  1.94       1
#> # … with 60 more rows