Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.30 0.0489   2.25  2.27  2.29  2.41
#>  2 2020-03-04    2020-03-06       2   2.23  2.22 0.0237   2.19  2.23  2.24  2.26
#>  3 2020-03-04    2020-03-07       3   2.17  2.19 0.0710   2.10  2.17  2.22  2.35
#>  4 2020-03-04    2020-03-08       4   2.13  2.14 0.0780   2.02  2.11  2.16  2.32
#>  5 2020-03-04    2020-03-09       5   2.09  2.11 0.0951   1.98  2.03  2.12  2.31
#>  6 2020-03-04    2020-03-10       6   2.01  2.05 0.0989   1.94  1.98  2.03  2.29
#>  7 2020-03-04    2020-03-11       7   1.96  2.00 0.124    1.86  1.89  1.97  2.28
#>  8 2020-03-05    2020-03-06       1   2.23  2.24 0.0276   2.19  2.21  2.25  2.28
#>  9 2020-03-05    2020-03-07       2   2.19  2.19 0.0205   2.15  2.19  2.20  2.23
#> 10 2020-03-05    2020-03-08       3   2.14  2.14 0.0410   2.07  2.13  2.18  2.20
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss   crps   logs   bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl>  <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -5.47 0.0169 -2.27   0.8      0.0255
#>  2 2020-03-04    2020-03-06       2 -5.48 0.0218 -2.08   0.6      0.0128
#>  3 2020-03-04    2020-03-07       3 -4.88 0.0254 -1.74   0.6      0.0511
#>  4 2020-03-04    2020-03-08       4 -4.95 0.0205 -1.86   0.6      0.0496
#>  5 2020-03-04    2020-03-09       5 -4.51 0.0237 -1.47   0.4      0.0702
#>  6 2020-03-04    2020-03-10       6 -4.69 0.0155 -1.82  -0.400    0.0409
#>  7 2020-03-04    2020-03-11       7 -4.26 0.0401 -0.974 -0.400    0.0738
#>  8 2020-03-05    2020-03-06       1 -4.52 0.0295 -1.65   0.8      0.0272
#>  9 2020-03-05    2020-03-07       2 -3.61 0.0299 -1.66   1        0.0131
#> 10 2020-03-05    2020-03-08       3 -5.55 0.0231 -1.73   0.6      0.0548
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   69    68.9  8.36     55    66    74    81
#>  2 2020-03-04    2020-03-06       2   83    83.8  9.39     70    79    91   100
#>  3 2020-03-04    2020-03-07       3   99.5  96.8 11.4      73    99   109   109
#>  4 2020-03-04    2020-03-08       4  126   125.   7.07    115   124   132   135
#>  5 2020-03-04    2020-03-09       5  134.  142.  16.4     128   128   135   180
#>  6 2020-03-04    2020-03-10       6  167   169.  15.9     147   162   177   200
#>  7 2020-03-04    2020-03-11       7  196.  198.  32.7     157   191   215   274
#>  8 2020-03-05    2020-03-06       1   80    83.6  8.97     72    72    81   103
#>  9 2020-03-05    2020-03-07       2   98    96.2  8.57     81    95   103   111
#> 10 2020-03-05    2020-03-08       3  111   111.  12.4      94   103   113   134
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.69  4.07  3.59   0.6      5.93
#>  2 1      2020-03-04    2020-03-06       2  5.84  6.56  3.73   0.6      8.90
#>  3 1      2020-03-04    2020-03-07       3  5.42  6.1   3.92   0.6     11.1 
#>  4 1      2020-03-04    2020-03-08       4 15.4  19.0   7.73   1        6.67
#>  5 1      2020-03-04    2020-03-09       5  8.17 17.6   4.88   1        9.64
#>  6 1      2020-03-04    2020-03-10       6  8.77 19.2   4.89   1       11.9 
#>  7 1      2020-03-04    2020-03-11       7  7.90 18.6   4.86   0.6     19.3 
#>  8 1      2020-03-05    2020-03-06       1  5.83  6.26  3.56   0.8      8.15
#>  9 1      2020-03-05    2020-03-07       2  5.21  5.52  3.81   0.6      5.93
#> 10 1      2020-03-05    2020-03-08       3  5.50  4.89  3.60   0.6     10.4 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.30 0.0505   2.22  2.27  2.32  2.36
#>  2 2020-03-04    2020-03-06       2   2.25  2.25 0.0750   2.12  2.19  2.28  2.33
#>  3 2020-03-04    2020-03-07       3   2.19  2.20 0.0794   2.02  2.15  2.20  2.26
#>  4 2020-03-04    2020-03-08       4   2.18  2.17 0.100    1.98  2.15  2.26  2.29
#>  5 2020-03-04    2020-03-09       5   2.10  2.13 0.142    1.91  1.95  2.12  2.30
#>  6 2020-03-04    2020-03-10       6   2.07  2.09 0.149    1.83  1.94  2.11  2.28
#>  7 2020-03-04    2020-03-11       7   2.01  2.03 0.178    1.73  1.87  2.07  2.29
#>  8 2020-03-05    2020-03-06       1   2.23  2.24 0.0577   2.14  2.20  2.25  2.32
#>  9 2020-03-05    2020-03-07       2   2.20  2.20 0.0707   2.14  2.14  2.21  2.36
#> 10 2020-03-05    2020-03-08       3   2.15  2.15 0.125    1.95  2.09  2.19  2.37
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -5.36 0.0141 -1.90  0.200    0.0558
#>  2 1      2020-03-04    2020-03-06       2 -4.44 0.0205 -1.45  0.4      0.0691
#>  3 1      2020-03-04    2020-03-07       3 -4.38 0.0266 -1.60  0.6      0.0553
#>  4 1      2020-03-04    2020-03-08       4 -3.79 0.0448 -0.890 0.200    0.133 
#>  5 1      2020-03-04    2020-03-09       5 -3.21 0.0671 -0.470 0.200    0.218 
#>  6 1      2020-03-04    2020-03-10       6 -3.28 0.0551 -0.511 0        0.212 
#>  7 1      2020-03-04    2020-03-11       7 -3.02 0.0678 -0.277 0        0.260 
#>  8 1      2020-03-05    2020-03-06       1 -4.55 0.0359 -1.35  0.8      0.0197
#>  9 1      2020-03-05    2020-03-07       2 -4.68 0.0304 -1.74  0.8      0.0403
#> 10 1      2020-03-05    2020-03-08       3 -4.81 0.0297 -1.70  0.8      0.0530
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   73    71.6  9.28     54    67    79    84
#>  2 2020-03-04    2020-03-06       2   86    86.4 11.4      67    83    92   102
#>  3 2020-03-04    2020-03-07       3  102.  101.  12.6      76   102   113   116
#>  4 2020-03-04    2020-03-08       4  120.  122.  14.1      98   112   128   142
#>  5 2020-03-04    2020-03-09       5  138.  143.  18.7     117   125   143   172
#>  6 2020-03-04    2020-03-10       6  172.  173.  26.3     128   155   174   229
#>  7 2020-03-04    2020-03-11       7  192   200.  39.7     144   171   210   266
#>  8 2020-03-05    2020-03-06       1   80.5  80    9.77     65    78    86    89
#>  9 2020-03-05    2020-03-07       2  102.  100.   9.16     86   101   112   112
#> 10 2020-03-05    2020-03-08       3  115   114.  16.0      90    97   117   136
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.74  7.29  3.98   0.6     11.9 
#>  2 1      2020-03-04    2020-03-06       2  6.40  7.75  3.78   1       10.4 
#>  3 1      2020-03-04    2020-03-07       3  5.84  7.22  4.10   0.4     11.1 
#>  4 1      2020-03-04    2020-03-08       4  8.35 18.1   4.82   0.8     17.0 
#>  5 1      2020-03-04    2020-03-09       5  7.75 15.7   4.45   1       17.8 
#>  6 1      2020-03-04    2020-03-10       6  7.94 17.9   4.94   0.6     17.8 
#>  7 1      2020-03-04    2020-03-11       7  8.31 17.6   4.97   0.4     52.6 
#>  8 1      2020-03-05    2020-03-06       1  4.40  3.52  3.38   0.4      9.64
#>  9 1      2020-03-05    2020-03-07       2  6.00  6.68  3.67   0.8      8.90
#> 10 1      2020-03-05    2020-03-08       3  6.21  8.7  10.7    0.6      4.45
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>