Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>  * <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.27  2.27 0.0349   2.24  2.24  2.27  2.33
#>  2 2020-03-04    2020-03-06       2   2.21  2.21 0.0674   2.10  2.14  2.23  2.31
#>  3 2020-03-04    2020-03-07       3   2.13  2.16 0.101    2.05  2.05  2.14  2.33
#>  4 2020-03-04    2020-03-08       4   2.07  2.11 0.127    1.99  1.99  2.09  2.34
#>  5 2020-03-04    2020-03-09       5   1.99  2.05 0.149    1.90  1.90  2.02  2.35
#>  6 2020-03-04    2020-03-10       6   1.99  2.02 0.170    1.84  1.84  2.06  2.32
#>  7 2020-03-04    2020-03-11       7   1.96  1.98 0.168    1.77  1.77  2.00  2.31
#>  8 2020-03-05    2020-03-06       1   2.22  2.21 0.0349   2.15  2.21  2.25  2.27
#>  9 2020-03-05    2020-03-07       2   2.17  2.17 0.0362   2.12  2.13  2.19  2.22
#> 10 2020-03-05    2020-03-08       3   2.12  2.12 0.0550   2.02  2.07  2.14  2.21
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss    crps   logs   bias sharpness
#>    <chr>         <date>       <int> <dbl>   <dbl>  <dbl>  <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -6.76 0.00886 -2.27   0        0.0378
#>  2 2020-03-04    2020-03-06       2 -5.42 0.0173  -1.64   0.200    0.0640
#>  3 2020-03-04    2020-03-07       3 -4.68 0.0298  -1.11  -0.200    0.112 
#>  4 2020-03-04    2020-03-08       4 -4.23 0.0414  -0.839 -0.200    0.107 
#>  5 2020-03-04    2020-03-09       5 -3.91 0.0502  -0.686 -0.200    0.102 
#>  6 2020-03-04    2020-03-10       6 -3.65 0.0572  -0.506  0        0.214 
#>  7 2020-03-04    2020-03-11       7 -3.63 0.0495  -0.610 -0.200    0.201 
#>  8 2020-03-05    2020-03-06       1 -6.36 0.0156  -2.04   0.6      0.0359
#>  9 2020-03-05    2020-03-07       2 -6.29 0.0140  -2.08   0.200    0.0489
#> 10 2020-03-05    2020-03-08       3 -5.76 0.0169  -1.80   0.4      0.0561
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>  * <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   70    71.8  5.18     65    68    71    82
#>  2 2020-03-04    2020-03-06       2   74.5  75.5 10.7      63    64    78    95
#>  3 2020-03-04    2020-03-07       3   99.5  96.9 10.2      77    99   108   108
#>  4 2020-03-04    2020-03-08       4  112.  113.  12.2      90   111   127   127
#>  5 2020-03-04    2020-03-09       5  139   139.  12.9     124   130   144   167
#>  6 2020-03-04    2020-03-10       6  167   163.  26.4     126   164   195   195
#>  7 2020-03-04    2020-03-11       7  188.  188.  35.3     139   156   201   238
#>  8 2020-03-05    2020-03-06       1   84.5  83.9  5.97     75    83    91    91
#>  9 2020-03-05    2020-03-07       2   93    95.5  8.51     88    88    95   116
#> 10 2020-03-05    2020-03-08       3  120.  122.  13.5     103   103   122   140
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>  * <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  6.39  6.18  3.85 1          2.22
#>  2 1      2020-03-04    2020-03-06       2  4.70  3.17  3.52 0         13.3 
#>  3 1      2020-03-04    2020-03-07       3  5.39  6.39  3.94 0.6        8.90
#>  4 1      2020-03-04    2020-03-08       4  5.77  6.78  3.89 0.8       14.1 
#>  5 1      2020-03-04    2020-03-09       5  8.66 16.7   5.05 1         10.4 
#>  6 1      2020-03-04    2020-03-10       6  7.15 12.8   4.62 0.5       36.3 
#>  7 1      2020-03-04    2020-03-11       7  7.43 13.1   4.76 0.200     45.2 
#>  8 1      2020-03-05    2020-03-06       1  7.17  7.69  3.84 1          8.15
#>  9 1      2020-03-05    2020-03-07       2  5.04  3.41  3.09 0.8        7.41
#> 10 1      2020-03-05    2020-03-08       3  7.58 12.9   4.29 1         19.3 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>  * <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0385   2.22  2.26  2.31  2.34
#>  2 2020-03-04    2020-03-06       2   2.25  2.24 0.0498   2.12  2.24  2.26  2.30
#>  3 2020-03-04    2020-03-07       3   2.20  2.19 0.0815   2.03  2.16  2.23  2.32
#>  4 2020-03-04    2020-03-08       4   2.17  2.16 0.128    1.95  2.11  2.20  2.34
#>  5 2020-03-04    2020-03-09       5   2.13  2.13 0.198    1.83  2.03  2.18  2.52
#>  6 2020-03-04    2020-03-10       6   2.08  2.09 0.251    1.59  1.96  2.15  2.44
#>  7 2020-03-04    2020-03-11       7   2.04  2.06 0.307    1.48  1.94  2.10  2.51
#>  8 2020-03-05    2020-03-06       1   2.21  2.20 0.0374   2.15  2.19  2.22  2.26
#>  9 2020-03-05    2020-03-07       2   2.16  2.14 0.0771   2.03  2.15  2.18  2.21
#> 10 2020-03-05    2020-03-08       3   2.09  2.07 0.103    2.00  2.07  2.10  2.21
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss    crps    logs   bias
#>  * <chr>  <chr>         <date>       <int> <dbl>   <dbl>   <dbl>  <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -6.24 0.0154  -1.80    0.200
#>  2 1      2020-03-04    2020-03-06       2 -5.15 0.0347  -0.827   0.6  
#>  3 1      2020-03-04    2020-03-07       3 -4.39 0.0444  -0.776   0.6  
#>  4 1      2020-03-04    2020-03-08       4 -3.61 0.0574  -0.599   0.4  
#>  5 1      2020-03-04    2020-03-09       5 -2.75 0.0807  -0.267   0.4  
#>  6 1      2020-03-04    2020-03-10       6 -2.23 0.0808  -0.0300  0.200
#>  7 1      2020-03-04    2020-03-11       7 -1.88 0.0841   0.105   0.200
#>  8 1      2020-03-05    2020-03-06       1 -7.56 0.00757 -2.84    0.6  
#>  9 1      2020-03-05    2020-03-07       2 -6.91 0.0133  -2.18    0.6  
#> 10 1      2020-03-05    2020-03-08       3 -6.00 0.00822 -2.58   -0.400
#> # … with 200 more rows, and 5 more variables: sharpness <dbl>,
#> #   calibration <dbl>, median <dbl>, iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>  * <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   68    68.7 11.8      42    64    75    79
#>  2 2020-03-04    2020-03-06       2   84    81.4 11.5      64    77    90    99
#>  3 2020-03-04    2020-03-07       3   98.5  98.0 13.2      81    87   103   121
#>  4 2020-03-04    2020-03-08       4  121   121.  14.9      95   110   127   147
#>  5 2020-03-04    2020-03-09       5  144.  144.  22.5     117   139   151   202
#>  6 2020-03-04    2020-03-10       6  164   167.  35.2      96   151   176   213
#>  7 2020-03-04    2020-03-11       7  197   206.  60.7      95   179   214   276
#>  8 2020-03-05    2020-03-06       1   77.5  79.6  8.74     69    69    81    94
#>  9 2020-03-05    2020-03-07       2   96    96.4  8.29     84    92   103   110
#> 10 2020-03-05    2020-03-08       3  118.  115.  11.7     100   116   128   131
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>  * <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.24  2.08  3.19 0.3        6.67
#>  2 1      2020-03-04    2020-03-06       2  5.55  6.63  3.89 0.6       10.4 
#>  3 1      2020-03-04    2020-03-07       3  5.76  5.17  3.98 0.3       18.5 
#>  4 1      2020-03-04    2020-03-08       4  6.45  9.14  4.14 0.6       11.9 
#>  5 1      2020-03-04    2020-03-09       5  7.72 18.6   5.02 0.8       25.2 
#>  6 1      2020-03-04    2020-03-10       6  7.86 13.3   4.84 0.4       44.5 
#>  7 1      2020-03-04    2020-03-11       7  9.00 25.2   5.51 0.200    104.  
#>  8 1      2020-03-05    2020-03-06       1  4.79  4.09  3.38 0.7        7.41
#>  9 1      2020-03-05    2020-03-07       2  6.68  7.82  3.87 1          7.41
#> 10 1      2020-03-05    2020-03-08       3  8.36 13.0   4.40 1          8.90
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>