Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.30 0.0227   2.27  2.29  2.31  2.34
#>  2 2020-03-04    2020-03-06       2   2.26  2.25 0.0245   2.21  2.24  2.27  2.30
#>  3 2020-03-04    2020-03-07       3   2.21  2.22 0.0381   2.17  2.20  2.22  2.31
#>  4 2020-03-04    2020-03-08       4   2.16  2.17 0.0282   2.12  2.16  2.18  2.22
#>  5 2020-03-04    2020-03-09       5   2.13  2.12 0.0387   2.05  2.10  2.13  2.17
#>  6 2020-03-04    2020-03-10       6   2.08  2.08 0.0442   2.00  2.07  2.09  2.15
#>  7 2020-03-04    2020-03-11       7   2.03  2.03 0.0630   1.91  2.01  2.05  2.15
#>  8 2020-03-05    2020-03-06       1   2.23  2.22 0.0740   2.02  2.21  2.26  2.28
#>  9 2020-03-05    2020-03-07       2   2.19  2.19 0.0454   2.12  2.16  2.21  2.28
#> 10 2020-03-05    2020-03-08       3   2.14  2.14 0.0811   2.04  2.08  2.16  2.32
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon    dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int>  <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -4.56  0.0263 -1.57    0.8    0.0190
#>  2 2020-03-04    2020-03-06       2 -0.549 0.0485 -0.191   1      0.0211
#>  3 2020-03-04    2020-03-07       3 -3.00  0.0517  0.663   1      0.0212
#>  4 2020-03-04    2020-03-08       4 -1.68  0.0483 -0.208   1      0.0270
#>  5 2020-03-04    2020-03-09       5 -4.30  0.0385 -1.52    0.6    0.0320
#>  6 2020-03-04    2020-03-10       6 -4.98  0.0351 -0.363   0.6    0.0151
#>  7 2020-03-04    2020-03-11       7 -5.58  0.0130 -2.13    0.4    0.0271
#>  8 2020-03-05    2020-03-06       1 -5.20  0.0254 -1.63    0.6    0.0443
#>  9 2020-03-05    2020-03-07       2 -5.04  0.0298 -1.66    0.8    0.0329
#> 10 2020-03-05    2020-03-08       3 -4.86  0.0252 -1.51    0.4    0.0607
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1    71   70.8  7.58     54    67    73    83
#>  2 2020-03-04    2020-03-06       2    81   79.6 10.2      67    70    81   100
#>  3 2020-03-04    2020-03-07       3   103  103    8.94     91    96   106   118
#>  4 2020-03-04    2020-03-08       4   123  121.   6.38    111   121   125   131
#>  5 2020-03-04    2020-03-09       5   142. 141.  11.5     125   132   147   158
#>  6 2020-03-04    2020-03-10       6   176. 177   13.6     157   169   184   204
#>  7 2020-03-04    2020-03-11       7   194. 198.  25.2     167   167   195   239
#>  8 2020-03-05    2020-03-06       1    81   82.8  7.80     68    79    88    95
#>  9 2020-03-05    2020-03-07       2    99   96.7  8.17     83    95   102   108
#> 10 2020-03-05    2020-03-08       3   110. 111.  12.9      91   108   120   136
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.12  5.82  4.39 0.8        5.19
#>  2 1      2020-03-04    2020-03-06       2  5.01  3.9   3.46 0.200     13.3 
#>  3 1      2020-03-04    2020-03-07       3  7.40 10.3   4.25 1         10.4 
#>  4 1      2020-03-04    2020-03-08       4 13.3  15.6   7.31 1          2.97
#>  5 1      2020-03-04    2020-03-09       5 10.1  18.9   5.34 1         14.8 
#>  6 1      2020-03-04    2020-03-10       6 12.7  28.4   7.40 1         11.1 
#>  7 1      2020-03-04    2020-03-11       7  8.01 17.3   4.64 0.8       34.8 
#>  8 1      2020-03-05    2020-03-06       1  5.76  6.74  3.89 0.8        7.41
#>  9 1      2020-03-05    2020-03-07       2  5.36  6.53  4.60 0.6        5.93
#> 10 1      2020-03-05    2020-03-08       3  5.50  5.42  3.78 0.4       12.6 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.28 0.0234   2.26  2.29  2.31  2.32
#>  2 2020-03-04    2020-03-06       2   2.23  2.23 0.0328   2.20  2.21  2.24  2.30
#>  3 2020-03-04    2020-03-07       3   2.18  2.18 0.0353   2.13  2.17  2.21  2.26
#>  4 2020-03-04    2020-03-08       4   2.13  2.12 0.0550   2.05  2.10  2.16  2.23
#>  5 2020-03-04    2020-03-09       5   2.08  2.07 0.0437   1.98  2.05  2.09  2.13
#>  6 2020-03-04    2020-03-10       6   2.00  2.01 0.0668   1.92  1.97  2.03  2.17
#>  7 2020-03-04    2020-03-11       7   1.97  1.97 0.0710   1.83  1.96  2.01  2.02
#>  8 2020-03-05    2020-03-06       1   2.22  2.22 0.0164   2.19  2.21  2.23  2.24
#>  9 2020-03-05    2020-03-07       2   2.16  2.17 0.0313   2.11  2.15  2.19  2.20
#> 10 2020-03-05    2020-03-08       3   2.10  2.10 0.0336   2.04  2.09  2.13  2.16
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss    crps  logs   bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>   <dbl> <dbl>  <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -6.53 0.0154  -2.24  0.6      0.0132
#>  2 1      2020-03-04    2020-03-06       2 -2.56 0.0328  -1.10  1        0.0280
#>  3 1      2020-03-04    2020-03-07       3 -4.71 0.0245  -1.94  1        0.0306
#>  4 1      2020-03-04    2020-03-08       4 -6.64 0.0153  -2.02  0.4      0.0273
#>  5 1      2020-03-04    2020-03-09       5 -6.42 0.0147  -2.01  0.4      0.0378
#>  6 1      2020-03-04    2020-03-10       6 -6.60 0.0113  -2.21 -0.200    0.0379
#>  7 1      2020-03-04    2020-03-11       7 -5.27 0.0231  -1.98 -0.8      0.0321
#>  8 1      2020-03-05    2020-03-06       1 -5.17 0.0202  -2.03  1        0.0189
#>  9 1      2020-03-05    2020-03-07       2 -6.13 0.0167  -2.12  0.8      0.0321
#> 10 1      2020-03-05    2020-03-08       3 -6.80 0.00907 -2.42  0.200    0.0274
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   70.5  69.6  8.91     56    63    75    79
#>  2 2020-03-04    2020-03-06       2   83.5  82.2  8.74     68    75    87    97
#>  3 2020-03-04    2020-03-07       3   95.5  96.8  9.90     81    91   101   114
#>  4 2020-03-04    2020-03-08       4  114.  115.  10.1     100   110   120   133
#>  5 2020-03-04    2020-03-09       5  143   143.  16.8     120   127   146   174
#>  6 2020-03-04    2020-03-10       6  160   163   15.7     145   145   160   198
#>  7 2020-03-04    2020-03-11       7  184.  185.  19.1     157   163   186   214
#>  8 2020-03-05    2020-03-06       1   81    81.7  9.01     72    72    82   100
#>  9 2020-03-05    2020-03-07       2   94.5  93.5  9.68     76    92   103   105
#> 10 2020-03-05    2020-03-08       3  116.  116.  13.7      88   116   126   129
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.43  5.73  3.78   0.5      8.90
#>  2 1      2020-03-04    2020-03-06       2  5.11  4.24  3.37   0.6      8.15
#>  3 1      2020-03-04    2020-03-07       3  5.43  5.25  3.47   0.8      5.19
#>  4 1      2020-03-04    2020-03-08       4  5.96  7.26  4.15   0.8      7.41
#>  5 1      2020-03-04    2020-03-09       5  8.96 19.3   4.93   1       14.8 
#>  6 1      2020-03-04    2020-03-10       6  9.17 14.8   5.04   1       10.4 
#>  7 1      2020-03-04    2020-03-11       7  6.74 11.2   4.41   0.4     25.2 
#>  8 1      2020-03-05    2020-03-06       1  4.51  3.77  3.48   0.6      5.93
#>  9 1      2020-03-05    2020-03-07       2  5.36  5.43  3.75   0.6      8.15
#> 10 1      2020-03-05    2020-03-08       3  6.43 10.7   5.79   0.6      7.41
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>