Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0171   2.27  2.29  2.30  2.33
#>  2 2020-03-04    2020-03-06       2   2.24  2.24 0.0238   2.22  2.23  2.24  2.29
#>  3 2020-03-04    2020-03-07       3   2.20  2.20 0.0183   2.17  2.20  2.22  2.22
#>  4 2020-03-04    2020-03-08       4   2.15  2.16 0.0360   2.12  2.12  2.16  2.21
#>  5 2020-03-04    2020-03-09       5   2.12  2.12 0.0266   2.08  2.10  2.13  2.17
#>  6 2020-03-04    2020-03-10       6   2.06  2.06 0.0320   2.01  2.05  2.07  2.11
#>  7 2020-03-04    2020-03-11       7   2.00  2.01 0.0338   1.95  1.99  2.01  2.07
#>  8 2020-03-05    2020-03-06       1   2.23  2.21 0.0970   2.00  2.18  2.26  2.33
#>  9 2020-03-05    2020-03-07       2   2.20  2.19 0.107    2.01  2.15  2.27  2.38
#> 10 2020-03-05    2020-03-08       3   2.13  2.14 0.131    1.92  2.12  2.26  2.35
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon    dss    crps   logs   bias sharpness
#>    <chr>         <date>       <int>  <dbl>   <dbl>  <dbl>  <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -5.90  0.0159  -2.36   1        0.0178
#>  2 2020-03-04    2020-03-06       2 -2.68  0.0384   5.50   1        0.0132
#>  3 2020-03-04    2020-03-07       3  0.270 0.0406  -0.194  1        0.0190
#>  4 2020-03-04    2020-03-08       4 -3.92  0.0382  -1.51   1        0.0439
#>  5 2020-03-04    2020-03-09       5 -1.60  0.0465  -0.406  1        0.0264
#>  6 2020-03-04    2020-03-10       6 -6.02  0.0222   0.766  0.6      0.0198
#>  7 2020-03-04    2020-03-11       7 -6.83  0.00938 -2.42  -0.400    0.0175
#>  8 2020-03-05    2020-03-06       1 -4.71  0.0258  -1.40   0.4      0.0606
#>  9 2020-03-05    2020-03-07       2 -4.37  0.0334  -1.07   0.4      0.0979
#> 10 2020-03-05    2020-03-08       3 -4.10  0.0356  -0.938  0.4      0.134 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   65    66.7  8.72     54    59    66    84
#>  2 2020-03-04    2020-03-06       2   85.5  87.5  6.67     81    81    86   101
#>  3 2020-03-04    2020-03-07       3  100   101.  12.1      85    85   101   117
#>  4 2020-03-04    2020-03-08       4  116   119.   8.97    110   110   117   139
#>  5 2020-03-04    2020-03-09       5  146   147.  10.8     132   139   154   166
#>  6 2020-03-04    2020-03-10       6  169   169   14.9     151   158   175   200
#>  7 2020-03-04    2020-03-11       7  198   200.  19.8     172   189   212   239
#>  8 2020-03-05    2020-03-06       1   83.5  83.4 11.5      67    72    85   105
#>  9 2020-03-05    2020-03-07       2   97.5  98.4  9.56     86    93    99   119
#> 10 2020-03-05    2020-03-08       3  113   115.  12.4     103   107   115   144
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.43  2.09  2.79   0.4      5.19
#>  2 1      2020-03-04    2020-03-06       2  8.94 11.0   5.14   1        5.93
#>  3 1      2020-03-04    2020-03-07       3  6.19  7.45  3.84   0.6     16.3 
#>  4 1      2020-03-04    2020-03-08       4  8.37 12.7   5.18   1        8.15
#>  5 1      2020-03-04    2020-03-09       5 13.6  24.8   7.32   1       11.1 
#>  6 1      2020-03-04    2020-03-10       6  9.09 19.7   5.16   1       14.8 
#>  7 1      2020-03-04    2020-03-11       7  8.87 22.1   5.02   1       17.0 
#>  8 1      2020-03-05    2020-03-06       1  5.69  5.68  3.72   0.5     13.3 
#>  9 1      2020-03-05    2020-03-07       2  5.72  6.    3.57   0.8      5.19
#> 10 1      2020-03-05    2020-03-08       3  6.21  7.39  3.73   1        8.15
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.28  2.29 0.0537   2.20  2.25  2.30  2.39
#>  2 2020-03-04    2020-03-06       2   2.24  2.25 0.0752   2.16  2.23  2.28  2.40
#>  3 2020-03-04    2020-03-07       3   2.22  2.22 0.0928   2.14  2.15  2.22  2.40
#>  4 2020-03-04    2020-03-08       4   2.18  2.19 0.117    2.08  2.08  2.18  2.43
#>  5 2020-03-04    2020-03-09       5   2.14  2.15 0.136    1.85  1.99  2.15  2.31
#>  6 2020-03-04    2020-03-10       6   2.09  2.10 0.157    1.90  2.00  2.13  2.43
#>  7 2020-03-04    2020-03-11       7   2.07  2.06 0.181    1.84  1.93  2.11  2.43
#>  8 2020-03-05    2020-03-06       1   2.24  2.24 0.0381   2.19  2.23  2.26  2.29
#>  9 2020-03-05    2020-03-07       2   2.19  2.19 0.0571   2.12  2.12  2.19  2.28
#> 10 2020-03-05    2020-03-08       3   2.13  2.13 0.0850   2.00  2.04  2.16  2.26
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -5.64 0.0128 -1.95  0        0.0451
#>  2 1      2020-03-04    2020-03-06       2 -4.23 0.0386 -1.49  0.8      0.0564
#>  3 1      2020-03-04    2020-03-07       3 -3.49 0.0573 -1.06  0.8      0.0838
#>  4 1      2020-03-04    2020-03-08       4 -3.39 0.0637 -0.885 0.6      0.113 
#>  5 1      2020-03-04    2020-03-09       5 -3.26 0.0759 -0.667 0.6      0.0922
#>  6 1      2020-03-04    2020-03-10       6 -3.60 0.0583 -0.977 0.6      0.145 
#>  7 1      2020-03-04    2020-03-11       7 -3.66 0.0641 -0.566 0.6      0.115 
#>  8 1      2020-03-05    2020-03-06       1 -5.71 0.0192 -2.17  0.4      0.0220
#>  9 1      2020-03-05    2020-03-07       2 -5.97 0.0124 -2.12  0.200    0.0377
#> 10 1      2020-03-05    2020-03-08       3 -5.45 0.0210 -1.52  0        0.0638
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   70    71.4  7.54     62    65    72    88
#>  2 2020-03-04    2020-03-06       2   84.5  84.4  7.28     73    81    91    95
#>  3 2020-03-04    2020-03-07       3  103   103.  13.5      80    92   106   121
#>  4 2020-03-04    2020-03-08       4  128.  125.  15.0     108   118   137   150
#>  5 2020-03-04    2020-03-09       5  144.  146.  22.7     117   130   162   186
#>  6 2020-03-04    2020-03-10       6  180.  179.  25.1     135   166   196   215
#>  7 2020-03-04    2020-03-11       7  219   215.  38.1     152   207   245   254
#>  8 2020-03-05    2020-03-06       1   79.5  82.4  8.62     73    73    83    98
#>  9 2020-03-05    2020-03-07       2  101   102.   8.77     88    95   105   119
#> 10 2020-03-05    2020-03-08       3  116.  116.  10.4     101   107   119   133
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.87  5.78  3.53   0.8      7.41
#>  2 1      2020-03-04    2020-03-06       2  5.71  6.07  3.61   0.8      7.41
#>  3 1      2020-03-04    2020-03-07       3  9.32 15.6   4.88   1       11.9 
#>  4 1      2020-03-04    2020-03-08       4  7.19 13.3   4.37   0.8     11.9 
#>  5 1      2020-03-04    2020-03-09       5  9.38 25.3   5.00   1       17.8 
#>  6 1      2020-03-04    2020-03-10       6  9.59 29.8   5.26   1       28.2 
#>  7 1      2020-03-04    2020-03-11       7 10.4  38.1   5.33   1       24.5 
#>  8 1      2020-03-05    2020-03-06       1  5.37  5.43  3.63   0.8      8.90
#>  9 1      2020-03-05    2020-03-07       2  6.64  8.74  3.94   0.9      8.90
#> 10 1      2020-03-05    2020-03-08       3  6.53 10.4   4.38   0.8     13.3 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>