Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.30 0.0648   2.20  2.24  2.32  2.41
#>  2 2020-03-04    2020-03-06       2   2.25  2.26 0.0711   2.13  2.20  2.30  2.36
#>  3 2020-03-04    2020-03-07       3   2.22  2.24 0.105    2.06  2.21  2.35  2.37
#>  4 2020-03-04    2020-03-08       4   2.16  2.17 0.124    1.98  2.15  2.31  2.31
#>  5 2020-03-04    2020-03-09       5   2.11  2.14 0.147    1.91  2.10  2.30  2.30
#>  6 2020-03-04    2020-03-10       6   2.09  2.10 0.178    1.78  2.09  2.32  2.32
#>  7 2020-03-04    2020-03-11       7   2.05  2.06 0.201    1.69  1.99  2.27  2.34
#>  8 2020-03-05    2020-03-06       1   2.23  2.22 0.0560   2.12  2.21  2.26  2.31
#>  9 2020-03-05    2020-03-07       2   2.19  2.17 0.0633   2.04  2.19  2.24  2.24
#> 10 2020-03-05    2020-03-08       3   2.13  2.11 0.0589   2.00  2.09  2.16  2.19
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -5.25 0.0230 -1.62  0.4      0.0638
#>  2 2020-03-04    2020-03-06       2 -4.43 0.0395 -1.29  0.8      0.0765
#>  3 2020-03-04    2020-03-07       3 -3.81 0.0587 -0.871 0.6      0.153 
#>  4 2020-03-04    2020-03-08       4 -3.93 0.0506 -0.804 0.4      0.196 
#>  5 2020-03-04    2020-03-09       5 -3.64 0.0541 -0.655 0.6      0.247 
#>  6 2020-03-04    2020-03-10       6 -3.39 0.0594 -0.500 0.200    0.235 
#>  7 2020-03-04    2020-03-11       7 -3.25 0.0521 -0.470 0.200    0.220 
#>  8 2020-03-05    2020-03-06       1 -5.60 0.0227 -1.44  0.4      0.0365
#>  9 2020-03-05    2020-03-07       2 -5.53 0.0242 -1.47  0.200    0.0567
#> 10 2020-03-05    2020-03-08       3 -5.75 0.0188 -1.59  0.200    0.0511
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   73.5  72.7  3.77     65    70    75    78
#>  2 2020-03-04    2020-03-06       2   86    86.4  9.90     70    76    90   100
#>  3 2020-03-04    2020-03-07       3  106.  102.  15.3      69   103   119   119
#>  4 2020-03-04    2020-03-08       4  118.  116.   8.42    100   116   122   128
#>  5 2020-03-04    2020-03-09       5  150   148.  22.3     114   149   175   175
#>  6 2020-03-04    2020-03-10       6  187   176.  29.6     127   160   195   221
#>  7 2020-03-04    2020-03-11       7  210.  214.  40.0     154   195   253   271
#>  8 2020-03-05    2020-03-06       1   83    83.2  9.59     72    72    83   102
#>  9 2020-03-05    2020-03-07       2   97.5  97.3  9.97     79    94   105   111
#> 10 2020-03-05    2020-03-08       3  119   118    4.78    109   115   120   126
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  9.90  7.73  4.40   1        3.71
#>  2 1      2020-03-04    2020-03-06       2  6.51  8.64  3.96   0.8     11.9 
#>  3 1      2020-03-04    2020-03-07       3  6.23  9.46  4.32   0.7     15.6 
#>  4 1      2020-03-04    2020-03-08       4  7.32 10.4   3.76   0.8      4.45
#>  5 1      2020-03-04    2020-03-09       5  8.35 20.2   4.77   0.8     31.1 
#>  6 1      2020-03-04    2020-03-10       6  8.15 22.4   4.97   0.6     25.9 
#>  7 1      2020-03-04    2020-03-11       7  8.81 28.2   5.13   0.8     56.3 
#>  8 1      2020-03-05    2020-03-06       1  5.67  5.36  3.55   0.7     11.9 
#>  9 1      2020-03-05    2020-03-07       2  5.46  6.43  4.04   0.6      8.15
#> 10 1      2020-03-05    2020-03-08       3 15.5  13.5   8.89   1        3.71
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0277   2.23  2.28  2.30  2.31
#>  2 2020-03-04    2020-03-06       2   2.22  2.23 0.0464   2.16  2.18  2.23  2.32
#>  3 2020-03-04    2020-03-07       3   2.19  2.18 0.0674   2.04  2.19  2.26  2.26
#>  4 2020-03-04    2020-03-08       4   2.12  2.11 0.0913   1.96  2.12  2.21  2.21
#>  5 2020-03-04    2020-03-09       5   2.08  2.06 0.103    1.86  2.07  2.17  2.17
#>  6 2020-03-04    2020-03-10       6   2.01  2.00 0.130    1.78  1.96  2.11  2.18
#>  7 2020-03-04    2020-03-11       7   1.94  1.93 0.154    1.69  1.89  2.04  2.18
#>  8 2020-03-05    2020-03-06       1   2.22  2.23 0.0623   2.12  2.18  2.24  2.32
#>  9 2020-03-05    2020-03-07       2   2.15  2.17 0.107    2.01  2.12  2.19  2.25
#> 10 2020-03-05    2020-03-08       3   2.14  2.13 0.173    1.87  2.07  2.20  2.24
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps   logs   bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>  <dbl>  <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -3.40 0.0206 -1.63   1        0.0129
#>  2 1      2020-03-04    2020-03-06       2 -5.33 0.0211 -1.74   0.6      0.0523
#>  3 1      2020-03-04    2020-03-07       3 -4.88 0.0290 -1.14   0.200    0.0886
#>  4 1      2020-03-04    2020-03-08       4 -4.38 0.0380 -0.981  0.200    0.0693
#>  5 1      2020-03-04    2020-03-09       5 -4.19 0.0421 -0.868  0.200    0.0782
#>  6 1      2020-03-04    2020-03-10       6 -3.77 0.0417 -0.769  0        0.142 
#>  7 1      2020-03-04    2020-03-11       7 -3.12 0.0494 -0.581 -0.200    0.160 
#>  8 1      2020-03-05    2020-03-06       1 -5.15 0.0122 -2.11   0.200    0.0428
#>  9 1      2020-03-05    2020-03-07       2 -3.91 0.0201 -1.77  -0.400    0.0526
#> 10 1      2020-03-05    2020-03-08       3 -2.95 0.0390 -0.961 -0.200    0.127 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   70    69.9  6.85     61    64    73    80
#>  2 2020-03-04    2020-03-06       2   84    81.2  9.14     68    79    88    95
#>  3 2020-03-04    2020-03-07       3  100    97   11.5      81    93   108   114
#>  4 2020-03-04    2020-03-08       4  114.  117.  12.4     101   101   116   134
#>  5 2020-03-04    2020-03-09       5  143   140.  12.4     123   140   157   161
#>  6 2020-03-04    2020-03-10       6  166   162.  18.2     137   166   188   188
#>  7 2020-03-04    2020-03-11       7  197   188.  22.2     161   196   217   217
#>  8 2020-03-05    2020-03-06       1   82    82.2 11.7      64    73    83    97
#>  9 2020-03-05    2020-03-07       2   98.5  98.0 11.5      81    85   100   113
#> 10 2020-03-05    2020-03-08       3  111   114.  15.0      96   100   116   128
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.92  3.87  3.20 0.6        9.64
#>  2 1      2020-03-04    2020-03-06       2  5.52  7.95  4.68 0.6        3.71
#>  3 1      2020-03-04    2020-03-07       3  5.10  2.9   3.39 0.4        8.15
#>  4 1      2020-03-04    2020-03-08       4  6.75  9.57  4.05 0.6       10.4 
#>  5 1      2020-03-04    2020-03-09       5  7.80 13.6   4.38 1         17.0 
#>  6 1      2020-03-04    2020-03-10       6  6.95 11.4   4.32 0.6       23.0 
#>  7 1      2020-03-04    2020-03-11       7  6.98 11.8   4.61 0.8       27.4 
#>  8 1      2020-03-05    2020-03-06       1  4.82  4.21  3.76 0.6        6.67
#>  9 1      2020-03-05    2020-03-07       2  5.57  4.16  3.52 0.400     11.1 
#> 10 1      2020-03-05    2020-03-08       3  6.28  5.78  3.73 0.5       13.3 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>