Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.32  2.34 0.0467   2.26  2.32  2.36  2.43
#>  2 2020-03-04    2020-03-06       2   2.25  2.23 0.0741   2.11  2.23  2.32  2.32
#>  3 2020-03-04    2020-03-07       3   2.17  2.20 0.0955   2.06  2.06  2.18  2.36
#>  4 2020-03-04    2020-03-08       4   2.18  2.17 0.107    1.99  2.07  2.23  2.32
#>  5 2020-03-04    2020-03-09       5   2.13  2.11 0.120    1.87  2.01  2.14  2.27
#>  6 2020-03-04    2020-03-10       6   2.02  2.03 0.114    1.80  1.95  2.08  2.19
#>  7 2020-03-04    2020-03-11       7   2.02  2.02 0.114    1.79  2.01  2.13  2.18
#>  8 2020-03-05    2020-03-06       1   2.23  2.23 0.0143   2.21  2.22  2.24  2.26
#>  9 2020-03-05    2020-03-07       2   2.18  2.17 0.0273   2.10  2.18  2.19  2.20
#> 10 2020-03-05    2020-03-08       3   2.12  2.12 0.0314   2.07  2.11  2.15  2.17
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 8
#>    forecast_date date       horizon    dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int>  <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -3.77  0.0473 -0.970   0.9    0.0432
#>  2 2020-03-04    2020-03-06       2 -4.97  0.0306 -1.24    0.7    0.0856
#>  3 2020-03-04    2020-03-07       3 -4.50  0.0288 -1.33    0.7    0.0898
#>  4 2020-03-04    2020-03-08       4 -4.10  0.0431 -0.952   0.7    0.133 
#>  5 2020-03-04    2020-03-09       5 -4.14  0.0400 -0.946   0.7    0.126 
#>  6 2020-03-04    2020-03-10       6 -4.44  0.0249 -1.14    0.4    0.0928
#>  7 2020-03-04    2020-03-11       7 -4.44  0.0239 -1.17    0.6    0.112 
#>  8 2020-03-05    2020-03-06       1  0.887 0.0344 -0.288   1      0.0132
#>  9 2020-03-05    2020-03-07       2 -6.27  0.0229  0.207   0.9    0.0122
#> 10 2020-03-05    2020-03-08       3 -6.80  0.0112 -2.24    0.7    0.0356
#> # … with 95 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   63.5  65    7.56     58    58    64    83
#>  2 2020-03-04    2020-03-06       2   68    68.3  7.66     54    62    69    83
#>  3 2020-03-04    2020-03-07       3   83.5  79.4 11.1      61    83    91    91
#>  4 2020-03-04    2020-03-08       4   90    93.1 14.4      72    78    95   114
#>  5 2020-03-04    2020-03-09       5  108.  106.  12.2      76   107   117   118
#>  6 2020-03-04    2020-03-10       6  115   121.  18.8     102   102   118   160
#>  7 2020-03-04    2020-03-11       7  134   140.  12.6     123   123   134   157
#>  8 2020-03-05    2020-03-06       1   73    73    5.64     65    72    78    83
#>  9 2020-03-05    2020-03-07       2   76    78.4  7.11     71    71    77    90
#> 10 2020-03-05    2020-03-08       3   97    95.5  8.40     79    91   102   107
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.02  1.34  2.73 0.5        5.93
#>  2 1      2020-03-04    2020-03-06       2  4.38  2.99  3.21 0.200      5.19
#>  3 1      2020-03-04    2020-03-07       3  5.38  3.96  3.47 0.200     11.1 
#>  4 1      2020-03-04    2020-03-08       4  5.65  7.03  4.09 0.3       13.3 
#>  5 1      2020-03-04    2020-03-09       5  5.63  4.67  3.38 0.200      8.90
#>  6 1      2020-03-04    2020-03-10       6  7.05 14.2   4.65 0.100     17.8 
#>  7 1      2020-03-04    2020-03-11       7 10.2  21.0   5.04 0          9.64
#>  8 1      2020-03-05    2020-03-06       1  3.35  1.38  2.85 0.5        6.67
#>  9 1      2020-03-05    2020-03-07       2  5.84  6.52  3.70 0.200      5.93
#> 10 1      2020-03-05    2020-03-08       3  4.82  3.11  3.28 0.100      8.15
#> # … with 95 more rows
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0407   2.24  2.26  2.30  2.34
#>  2 2020-03-04    2020-03-06       2   2.24  2.25 0.0483   2.15  2.23  2.27  2.32
#>  3 2020-03-04    2020-03-07       3   2.19  2.19 0.0692   2.11  2.15  2.20  2.33
#>  4 2020-03-04    2020-03-08       4   2.15  2.14 0.0813   2.06  2.12  2.19  2.28
#>  5 2020-03-04    2020-03-09       5   2.10  2.09 0.111    1.97  2.07  2.15  2.27
#>  6 2020-03-04    2020-03-10       6   2.05  2.04 0.125    1.90  2.03  2.12  2.21
#>  7 2020-03-04    2020-03-11       7   2.02  1.99 0.154    1.85  1.96  2.04  2.22
#>  8 2020-03-05    2020-03-06       1   2.23  2.22 0.0438   2.15  2.21  2.25  2.28
#>  9 2020-03-05    2020-03-07       2   2.17  2.17 0.0488   2.11  2.14  2.18  2.25
#> 10 2020-03-05    2020-03-08       3   2.11  2.11 0.0788   2.03  2.09  2.16  2.28
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss    crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>   <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -7.04 0.00595 -2.71    0.6    0.0184
#>  2 1      2020-03-04    2020-03-06       2 -6.26 0.0192  -1.71    0.7    0.0209
#>  3 1      2020-03-04    2020-03-07       3 -5.65 0.0139  -2.08    0.8    0.0306
#>  4 1      2020-03-04    2020-03-08       4 -5.13 0.0125  -2.06    0.6    0.0443
#>  5 1      2020-03-04    2020-03-09       5 -4.32 0.0176  -1.71    0.5    0.0596
#>  6 1      2020-03-04    2020-03-10       6 -3.84 0.0230  -1.67    0.3    0.0577
#>  7 1      2020-03-04    2020-03-11       7 -3.22 0.0448  -1.25    0.3    0.0816
#>  8 1      2020-03-05    2020-03-06       1 -5.00 0.0349  -0.660   0.9    0.0308
#>  9 1      2020-03-05    2020-03-07       2 -5.51 0.0227  -1.73    0.7    0.0528
#> 10 1      2020-03-05    2020-03-08       3 -5.02 0.0197  -1.75    0.7    0.0514
#> # … with 200 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   60    61.2  8.38     49    52    61    74
#>  2 2020-03-04    2020-03-06       2   71.5  68.8  7.15     60    71    79    79
#>  3 2020-03-04    2020-03-07       3   82    83    9.07     72    78    90   100
#>  4 2020-03-04    2020-03-08       4   93    93.6 10.6      73    87    96   107
#>  5 2020-03-04    2020-03-09       5  103   102.  10.7      80    94   106   113
#>  6 2020-03-04    2020-03-10       6  118   117.  15.7      98   112   133   138
#>  7 2020-03-04    2020-03-11       7  145   141.  19.4     121   134   150   186
#>  8 2020-03-05    2020-03-06       1   66.5  68.0  8.76     56    58    68    81
#>  9 2020-03-05    2020-03-07       2   77.5  77.8  6.92     70    73    80    92
#> 10 2020-03-05    2020-03-08       3   91    91.8 12.0      73    78    95   109
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.37  3.46  3.47 0.3        5.19
#>  2 1      2020-03-04    2020-03-06       2  3.70  1.81  2.93 0.3        8.15
#>  3 1      2020-03-04    2020-03-07       3  4.13  2.49  3.11 0.200      6.67
#>  4 1      2020-03-04    2020-03-08       4  6.02  7.56  4.76 0.100      4.45
#>  5 1      2020-03-04    2020-03-09       5  7.80 10.2   4.31 0          8.15
#>  6 1      2020-03-04    2020-03-10       6  8.38 18.1   4.68 0         17.8 
#>  7 1      2020-03-04    2020-03-11       7  9.54 24.4   6.89 0         13.3 
#>  8 1      2020-03-05    2020-03-06       1  4.72  3.8   3.26 0.200      8.15
#>  9 1      2020-03-05    2020-03-07       2  5.26  5.21  3.62 0.100      5.19
#> 10 1      2020-03-05    2020-03-08       3  5.32  4.64  3.66 0.200     11.1 
#> # … with 200 more rows
#>