Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean      sd bottom lower upper
#>  * <chr>         <date>       <int>  <dbl> <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.30 0.00755   2.29  2.29  2.30
#>  2 2020-03-04    2020-03-06       2   2.23  2.25 0.0458    2.21  2.22  2.24
#>  3 2020-03-04    2020-03-07       3   2.19  2.20 0.0516    2.15  2.16  2.19
#>  4 2020-03-04    2020-03-08       4   2.12  2.14 0.0705    2.09  2.09  2.13
#>  5 2020-03-04    2020-03-09       5   2.06  2.07 0.0903    1.97  2.05  2.09
#>  6 2020-03-04    2020-03-10       6   2.01  2.02 0.115     1.86  1.95  2.03
#>  7 2020-03-04    2020-03-11       7   1.97  1.96 0.139     1.75  1.89  1.99
#>  8 2020-03-05    2020-03-06       1   2.22  2.23 0.0245    2.20  2.21  2.23
#>  9 2020-03-05    2020-03-07       2   2.19  2.18 0.0365    2.12  2.18  2.22
#> 10 2020-03-05    2020-03-08       3   2.14  2.13 0.0441    2.06  2.10  2.15
#> # … with 123 more rows, and 1 more variable: top <dbl>
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss   crps   logs   bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl>  <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1  8.28 0.0271 31.6    1       0.00506
#>  2 2020-03-04    2020-03-06       2 -4.30 0.0394 -0.789  1       0.0161 
#>  3 2020-03-04    2020-03-07       3 -5.01 0.0252 -2.03   1       0.0325 
#>  4 2020-03-04    2020-03-08       4 -5.14 0.0133 -2.13   0       0.0378 
#>  5 2020-03-04    2020-03-09       5 -4.90 0.0106 -2.25   0       0.0338 
#>  6 2020-03-04    2020-03-10       6 -4.42 0.0204 -1.60  -0.200   0.0677 
#>  7 2020-03-04    2020-03-11       7 -3.89 0.0433 -1.17  -0.6     0.102  
#>  8 2020-03-05    2020-03-06       1 -5.21 0.0225 -1.99   1       0.0200 
#>  9 2020-03-05    2020-03-07       2 -5.76 0.0239 -1.76   0.6     0.0329 
#> 10 2020-03-05    2020-03-08       3 -5.91 0.0188 -1.82   0.4     0.0447 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>  * <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   68    69.8  8.63     60    60    69    84
#>  2 2020-03-04    2020-03-06       2   81.5  85.3 12.5      66    66    83   105
#>  3 2020-03-04    2020-03-07       3  101   100.   6.87     87    98   106   109
#>  4 2020-03-04    2020-03-08       4  120.  124.  15.9     108   108   121   157
#>  5 2020-03-04    2020-03-09       5  146.  148.  14.6     124   139   159   171
#>  6 2020-03-04    2020-03-10       6  165   164.  24.6     114   158   185   195
#>  7 2020-03-04    2020-03-11       7  194.  197.  31.1     146   190   220   257
#>  8 2020-03-05    2020-03-06       1   85.5  85.6  7.99     74    78    89    99
#>  9 2020-03-05    2020-03-07       2  106.  101.  10.6      87   105   116   116
#> 10 2020-03-05    2020-03-08       3  119   118.  10.6     101   118   128   134
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>  * <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.89  3.52  3.14   0.6      7.41
#>  2 1      2020-03-04    2020-03-06       2  6.02  7.07  3.83   0.7     17.0 
#>  3 1      2020-03-04    2020-03-07       3  7.20  8.69  4.10   0.8      6.67
#>  4 1      2020-03-04    2020-03-08       4  7.47 13.5   4.38   1       13.3 
#>  5 1      2020-03-04    2020-03-09       5 10.5  24.0   5.44   1       19.3 
#>  6 1      2020-03-04    2020-03-10       6  7.22 16.3   5.04   0.6     20.0 
#>  7 1      2020-03-04    2020-03-11       7  7.80 17.8   4.82   0.8     31.9 
#>  8 1      2020-03-05    2020-03-06       1  6.81  8.28  3.89   1        9.64
#>  9 1      2020-03-05    2020-03-07       2  6.34  7.84  3.88   0.6     10.4 
#> 10 1      2020-03-05    2020-03-08       3  7.18 10.7   4.17   0.8     11.9 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>  * <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.31  2.31 0.0466   2.25  2.28  2.34  2.40
#>  2 2020-03-04    2020-03-06       2   2.28  2.27 0.0644   2.18  2.23  2.32  2.39
#>  3 2020-03-04    2020-03-07       3   2.24  2.25 0.0914   2.14  2.20  2.31  2.40
#>  4 2020-03-04    2020-03-08       4   2.19  2.20 0.101    2.09  2.09  2.22  2.35
#>  5 2020-03-04    2020-03-09       5   2.13  2.16 0.119    2.06  2.06  2.17  2.35
#>  6 2020-03-04    2020-03-10       6   2.07  2.10 0.203    1.80  2.02  2.24  2.41
#>  7 2020-03-04    2020-03-11       7   2.06  2.07 0.188    1.86  1.97  2.14  2.35
#>  8 2020-03-05    2020-03-06       1   2.22  2.23 0.0511   2.12  2.20  2.25  2.30
#>  9 2020-03-05    2020-03-07       2   2.19  2.18 0.0722   2.10  2.12  2.20  2.35
#> 10 2020-03-05    2020-03-08       3   2.14  2.14 0.105    2.06  2.13  2.19  2.32
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps   logs  bias sharpness
#>  * <chr>  <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -4.69 0.0345 -1.51  0.8      0.0435
#>  2 1      2020-03-04    2020-03-06       2 -3.39 0.0569 -1.10  1        0.0862
#>  3 1      2020-03-04    2020-03-07       3 -3.34 0.0605 -1.05  0.8      0.112 
#>  4 1      2020-03-04    2020-03-08       4 -3.52 0.0536 -1.15  1        0.101 
#>  5 1      2020-03-04    2020-03-09       5 -3.18 0.0636 -0.973 0.8      0.121 
#>  6 1      2020-03-04    2020-03-10       6 -3.08 0.0588 -0.877 0.8      0.0960
#>  7 1      2020-03-04    2020-03-11       7 -3.36 0.0431 -0.902 0.200    0.0951
#>  8 1      2020-03-05    2020-03-06       1 -5.98 0.0232 -1.65  0.8      0.0294
#>  9 1      2020-03-05    2020-03-07       2 -6.07 0.0176 -1.81  0.200    0.0696
#> 10 1      2020-03-05    2020-03-08       3 -6.29 0.0169 -1.86  0.4      0.0290
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>  * <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   71    71.7  8.73     58    61    73    85
#>  2 2020-03-04    2020-03-06       2   87    86.3 10.9      65    83    97    99
#>  3 2020-03-04    2020-03-07       3  105   105.  15.6      72   100   115   121
#>  4 2020-03-04    2020-03-08       4  124.  127.  18.2     100   119   138   151
#>  5 2020-03-04    2020-03-09       5  154   156.  19.6     129   139   166   188
#>  6 2020-03-04    2020-03-10       6  178   184.  35.0     136   150   187   234
#>  7 2020-03-04    2020-03-11       7  210.  222.  44.3     168   186   233   299
#>  8 2020-03-05    2020-03-06       1   86    85.4  8.52     73    85    98    98
#>  9 2020-03-05    2020-03-07       2   98.5  99.8  9.91     84    93   104   116
#> 10 2020-03-05    2020-03-08       3  116.  119.  14.4      98    98   116   140
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>  * <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.98  3.89  3.28   0.7      8.15
#>  2 1      2020-03-04    2020-03-06       2  7.14 12.6   4.69   0.8      8.90
#>  3 1      2020-03-04    2020-03-07       3  7.30 13.3   4.83   0.8     10.4 
#>  4 1      2020-03-04    2020-03-08       4  8.42 18.0   4.65   1       18.5 
#>  5 1      2020-03-04    2020-03-09       5 11.2  29.7   6.18   1       17.0 
#>  6 1      2020-03-04    2020-03-10       6  9.08 28.6   5.02   1       51.9 
#>  7 1      2020-03-04    2020-03-11       7  9.44 35.2   5.29   1       29.7 
#>  8 1      2020-03-05    2020-03-06       1  7.01  9.15  3.96   1       10.4 
#>  9 1      2020-03-05    2020-03-07       2  5.87  5.82  3.54   0.7      8.90
#> 10 1      2020-03-05    2020-03-08       3  6.75  9.91  4.12   0.7     17.0 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>