Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
An uninitialised bsts model

horizon
Numeric, the time horizon over which to predict

samples
Numeric, the number of samples to from the posterior of the model fit. Note that twice this number of MCMC steps will be taken with half used as burn in.

timeout
Numeric, the number of seconds to allow before terminating model fitting. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0219   2.26  2.28  2.31  2.34
#>  2 2020-03-04    2020-03-06       2   2.24  2.24 0.0444   2.13  2.22  2.24  2.30
#>  3 2020-03-04    2020-03-07       3   2.19  2.18 0.0521   2.04  2.19  2.21  2.22
#>  4 2020-03-04    2020-03-08       4   2.14  2.13 0.0618   1.97  2.14  2.16  2.20
#>  5 2020-03-04    2020-03-09       5   2.10  2.07 0.0829   1.84  2.10  2.11  2.12
#>  6 2020-03-04    2020-03-10       6   2.04  2.01 0.103    1.72  2.04  2.06  2.06
#>  7 2020-03-04    2020-03-11       7   1.98  1.95 0.111    1.65  1.98  2.02  2.02
#>  8 2020-03-05    2020-03-06       1   2.21  2.22 0.0552   2.13  2.20  2.26  2.32
#>  9 2020-03-05    2020-03-07       2   2.23  2.21 0.0982   2.06  2.21  2.31  2.36
#> 10 2020-03-05    2020-03-08       3   2.16  2.16 0.115    1.98  2.09  2.25  2.34
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 8
#>    forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -6.15 0.0152 -2.40  0.9      0.0191
#>  2 2020-03-04    2020-03-06       2 -5.27 0.0363  2.87  0.9      0.0169
#>  3 2020-03-04    2020-03-07       3 -5.60 0.0318 -0.837 0.9      0.0231
#>  4 2020-03-04    2020-03-08       4 -5.53 0.0239 -1.41  0.8      0.0237
#>  5 2020-03-04    2020-03-09       5 -5.06 0.0278 -1.17  0.9      0.0206
#>  6 2020-03-04    2020-03-10       6 -4.61 0.0123 -2.26  0.7      0.0299
#>  7 2020-03-04    2020-03-11       7 -4.14 0.0214 -1.98  0.200    0.0472
#>  8 2020-03-05    2020-03-06       1 -5.60 0.0181 -1.87  0.8      0.0517
#>  9 2020-03-05    2020-03-07       2 -4.29 0.0480 -0.777 0.7      0.0862
#> 10 2020-03-05    2020-03-08       3 -4.15 0.0377 -0.984 0.6      0.115 
#> # … with 95 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   55.5  56.1  7.28     44    51    56    69
#>  2 2020-03-04    2020-03-06       2   63    65.4  9.16     54    54    64    81
#>  3 2020-03-04    2020-03-07       3   82.5  82.3 10.3      67    74    86   100
#>  4 2020-03-04    2020-03-08       4   89.5  91.6 11.7      74    85    97   115
#>  5 2020-03-04    2020-03-09       5  102.  101   12.5      81    94   108   119
#>  6 2020-03-04    2020-03-10       6  109   109.  14.3      89    89   110   130
#>  7 2020-03-04    2020-03-11       7  128.  130.  17.8     101   125   137   169
#>  8 2020-03-05    2020-03-06       1   67    65.4  6.13     55    66    73    73
#>  9 2020-03-05    2020-03-07       2   84    83.1  6.15     71    82    88    92
#> 10 2020-03-05    2020-03-08       3   96    94.9 10.3      78    90   105   109
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.86  4.69  3.62 0.200      5.93
#>  2 1      2020-03-04    2020-03-06       2  5.09  5.32  3.76 0.200      8.90
#>  3 1      2020-03-04    2020-03-07       3  4.90  4.01  3.62 0.200     11.9 
#>  4 1      2020-03-04    2020-03-08       4  5.69  7.08  4.07 0.200      9.64
#>  5 1      2020-03-04    2020-03-09       5  6.55  9.08  4.02 0.200     10.4 
#>  6 1      2020-03-04    2020-03-10       6 10.8  24.5   5.55 0         17.8 
#>  7 1      2020-03-04    2020-03-11       7 10.6  29.1   5.19 0.100     11.1 
#>  8 1      2020-03-05    2020-03-06       1  5.23  4.32  3.25 0          8.90
#>  9 1      2020-03-05    2020-03-07       2  4.23  2.45  2.96 0.100      5.93
#> 10 1      2020-03-05    2020-03-08       3  5.09  3.93  3.51 0.3       11.9 
#> # … with 95 more rows
#> 
#> $raw_rt_forecast
#> # A tibble: 1,330 x 6
#>    obs_sample forecast_date sample date          rt horizon
#>    <chr>      <chr>          <int> <date>     <dbl>   <int>
#>  1 1          2020-03-04         1 2020-03-05  2.30       1
#>  2 1          2020-03-04         2 2020-03-05  2.30       1
#>  3 1          2020-03-04         3 2020-03-05  2.28       1
#>  4 1          2020-03-04         4 2020-03-05  2.29       1
#>  5 1          2020-03-04         5 2020-03-05  2.31       1
#>  6 1          2020-03-04         6 2020-03-05  2.30       1
#>  7 1          2020-03-04         7 2020-03-05  2.28       1
#>  8 1          2020-03-04         8 2020-03-05  2.26       1
#>  9 1          2020-03-04         9 2020-03-05  2.34       1
#> 10 1          2020-03-04        10 2020-03-05  2.27       1
#> # … with 1,320 more rows
#> 
#> $raw_case_forecast
#> # A tibble: 1,050 x 6
#>    obs_sample forecast_date sample date       cases horizon
#>    <chr>      <chr>          <dbl> <date>     <int>   <int>
#>  1 1          2020-03-04         1 2020-03-05    52       1
#>  2 1          2020-03-04         1 2020-03-06    54       2
#>  3 1          2020-03-04         1 2020-03-07    86       3
#>  4 1          2020-03-04         1 2020-03-08    85       4
#>  5 1          2020-03-04         1 2020-03-09   103       5
#>  6 1          2020-03-04         1 2020-03-10   124       6
#>  7 1          2020-03-04         1 2020-03-11   121       7
#>  8 1          2020-03-04         2 2020-03-05    44       1
#>  9 1          2020-03-04         2 2020-03-06    59       2
#> 10 1          2020-03-04         2 2020-03-07    86       3
#> # … with 1,040 more rows
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.31  2.31 0.0707   2.16  2.29  2.35  2.37
#>  2 2020-03-04    2020-03-06       2   2.26  2.26 0.0824   2.16  2.21  2.29  2.43
#>  3 2020-03-04    2020-03-07       3   2.20  2.19 0.122    2.02  2.11  2.27  2.42
#>  4 2020-03-04    2020-03-08       4   2.18  2.15 0.162    1.92  2.07  2.26  2.38
#>  5 2020-03-04    2020-03-09       5   2.13  2.09 0.233    1.80  2.02  2.20  2.50
#>  6 2020-03-04    2020-03-10       6   2.10  2.06 0.291    1.61  1.83  2.17  2.63
#>  7 2020-03-04    2020-03-11       7   2.03  2.01 0.309    1.32  1.98  2.24  2.41
#>  8 2020-03-05    2020-03-06       1   2.19  2.20 0.0505   2.13  2.13  2.19  2.27
#>  9 2020-03-05    2020-03-07       2   2.15  2.15 0.0768   2.04  2.09  2.20  2.28
#> 10 2020-03-05    2020-03-08       3   2.11  2.10 0.113    1.93  1.99  2.17  2.26
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss   crps    logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>   <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -4.63 0.0206 -1.51   0.5      0.0647
#>  2 1      2020-03-04    2020-03-06       2 -4.43 0.0343 -1.15   0.6      0.0962
#>  3 1      2020-03-04    2020-03-07       3 -3.89 0.0384 -0.904  0.5      0.121 
#>  4 1      2020-03-04    2020-03-08       4 -3.31 0.0525 -0.498  0.6      0.209 
#>  5 1      2020-03-04    2020-03-09       5 -2.51 0.0620 -0.344  0.5      0.242 
#>  6 1      2020-03-04    2020-03-10       6 -2.18 0.0790 -0.0605 0.5      0.326 
#>  7 1      2020-03-04    2020-03-11       7 -2.10 0.0770 -0.0746 0.5      0.313 
#>  8 1      2020-03-05    2020-03-06       1 -6.16 0.0206 -1.67   0.3      0.0214
#>  9 1      2020-03-05    2020-03-07       2 -5.51 0.0280 -1.28   0.200    0.0381
#> 10 1      2020-03-05    2020-03-08       3 -4.69 0.0402 -0.976  0.3      0.0827
#> # … with 200 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   62    62.4  8.04     47    58    66    76
#>  2 2020-03-04    2020-03-06       2   69.5  70.1 10.5      55    63    74    89
#>  3 2020-03-04    2020-03-07       3   80.5  79.8 11.8      62    73    88   102
#>  4 2020-03-04    2020-03-08       4   96.5  94.7 11.8      74    86    99   113
#>  5 2020-03-04    2020-03-09       5  111   111.  20.4      72   108   132   132
#>  6 2020-03-04    2020-03-10       6  127   126.  22.3      99   115   138   169
#>  7 2020-03-04    2020-03-11       7  142.  141.  35.0      80   124   156   181
#>  8 2020-03-05    2020-03-06       1   66    66.7  7.44     56    64    69    80
#>  9 2020-03-05    2020-03-07       2   79    79.4 10.0      60    78    89    94
#> 10 2020-03-05    2020-03-08       3   96    94.0 12.2      71    90   103   112
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.21  2.46  3.28 0.4        8.90
#>  2 1      2020-03-04    2020-03-06       2  5.04  2.9   3.53 0.3       10.4 
#>  3 1      2020-03-04    2020-03-07       3  5.59  6.04  3.82 0.100     12.6 
#>  4 1      2020-03-04    2020-03-08       4  5.59  5.81  3.77 0.100     11.1 
#>  5 1      2020-03-04    2020-03-09       5  6.46  8.02  4.37 0.4       24.5 
#>  6 1      2020-03-04    2020-03-10       6  6.84 11.3   4.67 0.200     19.3 
#>  7 1      2020-03-04    2020-03-11       7  7.91 18.7   5.12 0.3       42.3 
#>  8 1      2020-03-05    2020-03-06       1  4.45  3.76  3.36 0.200      5.19
#>  9 1      2020-03-05    2020-03-07       2  5.22  4.28  3.33 0.3        8.90
#> 10 1      2020-03-05    2020-03-08       3  5.55  4.01  3.50 0.3       11.1 
#> # … with 200 more rows
#> 
#> $raw_rt_forecast
#> # A tibble: 2,660 x 6
#>    obs_sample forecast_date sample date          rt horizon
#>    <chr>      <chr>          <int> <date>     <dbl>   <int>
#>  1 1          2020-03-04         1 2020-03-05  2.52       1
#>  2 1          2020-03-04         2 2020-03-05  2.35       1
#>  3 1          2020-03-04         3 2020-03-05  2.37       1
#>  4 1          2020-03-04         4 2020-03-05  2.32       1
#>  5 1          2020-03-04         5 2020-03-05  2.26       1
#>  6 1          2020-03-04         6 2020-03-05  2.29       1
#>  7 1          2020-03-04         7 2020-03-05  2.16       1
#>  8 1          2020-03-04         8 2020-03-05  2.25       1
#>  9 1          2020-03-04         9 2020-03-05  2.25       1
#> 10 1          2020-03-04        10 2020-03-05  2.23       1
#> # … with 2,650 more rows
#> 
#> $raw_case_forecast
#> # A tibble: 2,100 x 6
#>    obs_sample forecast_date sample date       cases horizon
#>    <chr>      <chr>          <dbl> <date>     <int>   <int>
#>  1 1          2020-03-04         1 2020-03-05    69       1
#>  2 1          2020-03-04         1 2020-03-06    80       2
#>  3 1          2020-03-04         1 2020-03-07   100       3
#>  4 1          2020-03-04         1 2020-03-08   120       4
#>  5 1          2020-03-04         1 2020-03-09   160       5
#>  6 1          2020-03-04         1 2020-03-10   169       6
#>  7 1          2020-03-04         1 2020-03-11   239       7
#>  8 1          2020-03-04         2 2020-03-05    54       1
#>  9 1          2020-03-04         2 2020-03-06    48       2
#> 10 1          2020-03-04         2 2020-03-07    86       3
#> # … with 2,090 more rows
#>