Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model etc.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.28 0.0200   2.26  2.26  2.29  2.32
#>  2 2020-03-04    2020-03-06       2   2.23  2.23 0.0519   2.13  2.19  2.23  2.33
#>  3 2020-03-04    2020-03-07       3   2.18  2.18 0.0661   2.04  2.15  2.20  2.30
#>  4 2020-03-04    2020-03-08       4   2.13  2.14 0.103    2.00  2.12  2.14  2.40
#>  5 2020-03-04    2020-03-09       5   2.08  2.09 0.155    1.91  2.05  2.10  2.50
#>  6 2020-03-04    2020-03-10       6   2.03  2.06 0.203    1.85  2.02  2.06  2.61
#>  7 2020-03-04    2020-03-11       7   1.98  2.01 0.225    1.82  1.95  2.03  2.62
#>  8 2020-03-05    2020-03-06       1   2.20  2.20 0.0655   2.08  2.17  2.26  2.29
#>  9 2020-03-05    2020-03-07       2   2.15  2.13 0.114    1.89  2.12  2.25  2.28
#> 10 2020-03-05    2020-03-08       3   2.10  2.08 0.136    1.76  2.06  2.17  2.26
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 8
#>    forecast_date date       horizon   dss   crps  logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl> <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -7.02 0.0103 -2.58 0.7      0.0259
#>  2 2020-03-04    2020-03-06       2 -5.58 0.0195 -1.90 0.8      0.0392
#>  3 2020-03-04    2020-03-07       3 -5.31 0.0221 -1.62 0.8      0.0348
#>  4 2020-03-04    2020-03-08       4 -4.55 0.0175 -1.94 0.7      0.0187
#>  5 2020-03-04    2020-03-09       5 -3.79 0.0171 -2.00 0.6      0.0355
#>  6 2020-03-04    2020-03-10       6 -3.27 0.0173 -2.01 0.6      0.0313
#>  7 2020-03-04    2020-03-11       7 -3.09 0.0357 -1.43 0.200    0.0583
#>  8 2020-03-05    2020-03-06       1 -5.55 0.0172 -1.64 0.5      0.0620
#>  9 2020-03-05    2020-03-07       2 -4.43 0.0235 -1.24 0.5      0.104 
#> 10 2020-03-05    2020-03-08       3 -4.05 0.0232 -1.37 0.5      0.0864
#> # … with 95 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   61    60.4  7.69     49    55    63    76
#>  2 2020-03-04    2020-03-06       2   65    64.3  4.60     58    60    67    71
#>  3 2020-03-04    2020-03-07       3   84.5  82.2 12.3      65    83    97    97
#>  4 2020-03-04    2020-03-08       4   94    93.8  6.36     84    90    95   105
#>  5 2020-03-04    2020-03-09       5  106.  104.  10.3      86    98   111   119
#>  6 2020-03-04    2020-03-10       6  124   122.  21.4      95   111   136   165
#>  7 2020-03-04    2020-03-11       7  135   135.  13.6     115   132   145   164
#>  8 2020-03-05    2020-03-06       1   63    60.8  5.53     49    63    68    68
#>  9 2020-03-05    2020-03-07       2   75    80.2 10.3      69    69    76    98
#> 10 2020-03-05    2020-03-08       3   95.5  96    8.76     79    90    97   112
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.10  2     2.94 0.200      6.67
#>  2 1      2020-03-04    2020-03-06       2  6.93  6.23  3.80 0          5.19
#>  3 1      2020-03-04    2020-03-07       3  5.16  4.04  3.65 0.4       14.1 
#>  4 1      2020-03-04    2020-03-08       4  5.44  5.72  3.75 0.200      5.19
#>  5 1      2020-03-04    2020-03-09       5  5.97  6.66  3.71 0.100     10.4 
#>  6 1      2020-03-04    2020-03-10       6  6.96 13.3   4.53 0.100     18.5 
#>  7 1      2020-03-04    2020-03-11       7 11.2  24.8   5.12 0         10.4 
#>  8 1      2020-03-05    2020-03-06       1  8.72  9.4   5.18 0          5.19
#>  9 1      2020-03-05    2020-03-07       2  5.20  6.34  3.98 0.3        6.67
#> 10 1      2020-03-05    2020-03-08       3  4.76  3.92  3.51 0.200      5.93
#> # … with 95 more rows
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.28  2.28 0.0417   2.20  2.27  2.28  2.34
#>  2 2020-03-04    2020-03-06       2   2.23  2.24 0.0603   2.13  2.22  2.26  2.29
#>  3 2020-03-04    2020-03-07       3   2.20  2.19 0.0733   2.14  2.14  2.21  2.30
#>  4 2020-03-04    2020-03-08       4   2.13  2.13 0.0986   2.05  2.09  2.16  2.28
#>  5 2020-03-04    2020-03-09       5   2.09  2.09 0.153    2.01  2.03  2.11  2.30
#>  6 2020-03-04    2020-03-10       6   2.03  2.03 0.183    1.92  1.98  2.06  2.35
#>  7 2020-03-04    2020-03-11       7   2.00  1.99 0.207    1.86  1.92  2.02  2.29
#>  8 2020-03-05    2020-03-06       1   2.21  2.20 0.0243   2.16  2.20  2.22  2.24
#>  9 2020-03-05    2020-03-07       2   2.15  2.14 0.0287   2.11  2.14  2.17  2.18
#> 10 2020-03-05    2020-03-08       3   2.09  2.08 0.0474   2.01  2.06  2.11  2.17
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss    crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>   <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -5.83 0.0139  -2.18    0.8    0.0363
#>  2 1      2020-03-04    2020-03-06       2 -4.73 0.0243  -1.61    0.7    0.0503
#>  3 1      2020-03-04    2020-03-07       3 -4.45 0.0467  -0.730   0.9    0.0426
#>  4 1      2020-03-04    2020-03-08       4 -4.05 0.0317  -1.32    0.8    0.0872
#>  5 1      2020-03-04    2020-03-09       5 -3.14 0.0462  -0.954   0.9    0.0770
#>  6 1      2020-03-04    2020-03-10       6 -2.81 0.0392  -0.854   0.5    0.126 
#>  7 1      2020-03-04    2020-03-11       7 -2.57 0.0408  -0.719   0.5    0.111 
#>  8 1      2020-03-05    2020-03-06       1 -7.00 0.0130  -2.04    0.9    0.0104
#>  9 1      2020-03-05    2020-03-07       2 -7.58 0.00616 -2.68    0.5    0.0212
#> 10 1      2020-03-05    2020-03-08       3 -6.53 0.00809 -2.53    0.4    0.0244
#> # … with 200 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   61    61.2  6.95     49    56    66    70
#>  2 2020-03-04    2020-03-06       2   65    65.3  5.96     55    59    66    73
#>  3 2020-03-04    2020-03-07       3   81.5  80.4 10.2      60    79    92    93
#>  4 2020-03-04    2020-03-08       4   91    90    9.89     77    88   100   108
#>  5 2020-03-04    2020-03-09       5  104.   99.8 15.3      71    98   111   120
#>  6 2020-03-04    2020-03-10       6  128   121.  19.1      97   128   149   149
#>  7 2020-03-04    2020-03-11       7  144   138.  27.8     100   115   149   177
#>  8 2020-03-05    2020-03-06       1   70.5  69.4 11.7      49    65    79    87
#>  9 2020-03-05    2020-03-07       2   84.5  81.6  9.65     67    77    86    98
#> 10 2020-03-05    2020-03-08       3   86    90.6 13.7      77    80    90   117
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  3.92  2.92  3.36 0.4        3.71
#>  2 1      2020-03-04    2020-03-06       2  6.76  5.9   3.67 0          4.45
#>  3 1      2020-03-04    2020-03-07       3  5.70  6.01  3.70 0.200     11.9 
#>  4 1      2020-03-04    2020-03-08       4  5.72  5.43  3.45 0         10.4 
#>  5 1      2020-03-04    2020-03-09       5  6.65  7.74  3.74 0.100      9.64
#>  6 1      2020-03-04    2020-03-10       6  7.13 11.4   4.38 0.200     28.2 
#>  7 1      2020-03-04    2020-03-11       7  8.20 20.2   4.81 0.100     31.9 
#>  8 1      2020-03-05    2020-03-06       1  4.20  2.27  3.22 0.3        8.15
#>  9 1      2020-03-05    2020-03-07       2  4.91  3.63  3.11 0.100      2.22
#> 10 1      2020-03-05    2020-03-08       3  5.48  7.32  4.22 0.3        6.67
#> # … with 200 more rows
#>