Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.29 0.0341   2.25  2.27  2.31  2.36
#>  2 2020-03-04    2020-03-06       2   2.24  2.24 0.0550   2.13  2.21  2.25  2.34
#>  3 2020-03-04    2020-03-07       3   2.17  2.18 0.0900   2.02  2.13  2.22  2.31
#>  4 2020-03-04    2020-03-08       4   2.16  2.14 0.111    1.94  2.14  2.27  2.28
#>  5 2020-03-04    2020-03-09       5   2.12  2.09 0.164    1.81  2.05  2.21  2.36
#>  6 2020-03-04    2020-03-10       6   2.04  2.02 0.196    1.72  2.01  2.19  2.35
#>  7 2020-03-04    2020-03-11       7   1.98  1.99 0.220    1.70  1.94  2.19  2.37
#>  8 2020-03-05    2020-03-06       1   2.23  2.24 0.0549   2.18  2.19  2.23  2.33
#>  9 2020-03-05    2020-03-07       2   2.23  2.22 0.0755   2.08  2.22  2.32  2.32
#> 10 2020-03-05    2020-03-08       3   2.14  2.12 0.109    1.90  2.11  2.25  2.25
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 8
#>    forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -6.34 0.0145 -2.12    0.7    0.0344
#>  2 2020-03-04    2020-03-06       2 -5.23 0.0281 -1.45    0.9    0.0313
#>  3 2020-03-04    2020-03-07       3 -4.79 0.0234 -1.44    0.7    0.0688
#>  4 2020-03-04    2020-03-08       4 -4.40 0.0403 -0.909   0.7    0.0965
#>  5 2020-03-04    2020-03-09       5 -3.69 0.0457 -0.680   0.6    0.136 
#>  6 2020-03-04    2020-03-10       6 -3.36 0.0425 -0.618   0.5    0.190 
#>  7 2020-03-04    2020-03-11       7 -3.11 0.0584 -0.364   0.4    0.267 
#>  8 2020-03-05    2020-03-06       1 -5.02 0.0249 -1.83    0.8    0.0455
#>  9 2020-03-05    2020-03-07       2 -4.31 0.0430 -1.14    0.8    0.0918
#> 10 2020-03-05    2020-03-08       3 -4.53 0.0320 -1.06    0.6    0.131 
#> # … with 95 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   59    59   10.1      43    55    70    72
#>  2 2020-03-04    2020-03-06       2   72.5  72.3  6.15     64    71    80    80
#>  3 2020-03-04    2020-03-07       3   87    82.5 15.3      62    69    91   109
#>  4 2020-03-04    2020-03-08       4   95.5  94.4 12.4      74    91   107   109
#>  5 2020-03-04    2020-03-09       5  106.  108.  12.0      93    93   106   131
#>  6 2020-03-04    2020-03-10       6  122   118.  18.1      93   119   141   141
#>  7 2020-03-04    2020-03-11       7  136   141.  26.6     102   102   141   178
#>  8 2020-03-05    2020-03-06       1   62    64.5 11.5      46    56    74    80
#>  9 2020-03-05    2020-03-07       2   87    86.3  9.74     69    82    93   101
#> 10 2020-03-05    2020-03-08       3   92.5  92.8  5.88     85    91    95   107
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.69  3.34  3.50 0.4       11.9 
#>  2 1      2020-03-04    2020-03-06       2  3.54  1.99  3.10 0.5        9.64
#>  3 1      2020-03-04    2020-03-07       3  5.49  4.01  3.85 0.5       16.3 
#>  4 1      2020-03-04    2020-03-08       4  5.34  4.38  3.65 0.4       15.6 
#>  5 1      2020-03-04    2020-03-09       5  5.40  5.41  3.86 0.3       13.3 
#>  6 1      2020-03-04    2020-03-10       6  7.51 13.5   4.32 0         25.9 
#>  7 1      2020-03-04    2020-03-11       7  7.54 14.6   4.62 0.200     37.1 
#>  8 1      2020-03-05    2020-03-06       1  5.39  5.15  3.75 0.4       14.8 
#>  9 1      2020-03-05    2020-03-07       2  4.48  2.53  3.30 0.5        8.90
#> 10 1      2020-03-05    2020-03-08       3  6.16  7.38  6.38 0.100      2.97
#> # … with 95 more rows
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.28  2.29 0.0584   2.22  2.27  2.29  2.40
#>  2 2020-03-04    2020-03-06       2   2.21  2.22 0.0646   2.10  2.15  2.21  2.29
#>  3 2020-03-04    2020-03-07       3   2.15  2.17 0.0952   1.97  2.10  2.17  2.29
#>  4 2020-03-04    2020-03-08       4   2.09  2.11 0.101    2.00  2.04  2.11  2.31
#>  5 2020-03-04    2020-03-09       5   2.02  2.03 0.108    1.83  1.96  2.08  2.22
#>  6 2020-03-04    2020-03-10       6   1.97  1.97 0.112    1.84  1.92  2.04  2.22
#>  7 2020-03-04    2020-03-11       7   1.90  1.92 0.118    1.75  1.86  2.00  2.04
#>  8 2020-03-05    2020-03-06       1   2.22  2.22 0.0446   2.13  2.18  2.24  2.28
#>  9 2020-03-05    2020-03-07       2   2.15  2.15 0.0588   2.08  2.14  2.17  2.26
#> 10 2020-03-05    2020-03-08       3   2.10  2.10 0.0681   2.02  2.07  2.14  2.25
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss    crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>   <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -5.01 0.0167  -2.01    0.8    0.0228
#>  2 1      2020-03-04    2020-03-06       2 -4.84 0.0231  -1.38    0.6    0.0824
#>  3 1      2020-03-04    2020-03-07       3 -4.21 0.0331  -1.06    0.5    0.108 
#>  4 1      2020-03-04    2020-03-08       4 -4.41 0.0228  -1.41    0.5    0.0791
#>  5 1      2020-03-04    2020-03-09       5 -4.33 0.0250  -1.36    0.5    0.0892
#>  6 1      2020-03-04    2020-03-10       6 -3.94 0.0346  -1.17    0.3    0.0839
#>  7 1      2020-03-04    2020-03-11       7 -3.83 0.0521  -0.769   0.3    0.105 
#>  8 1      2020-03-05    2020-03-06       1 -5.45 0.0203  -2.15    1      0.0322
#>  9 1      2020-03-05    2020-03-07       2 -7.67 0.00524 -3.01    0.7    0.0186
#> 10 1      2020-03-05    2020-03-08       3 -6.95 0.00834 -2.44    0.6    0.0288
#> # … with 200 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   59    58.0  6.13     48    54    61    69
#>  2 2020-03-04    2020-03-06       2   69.5  69.6  8.28     57    67    77    80
#>  3 2020-03-04    2020-03-07       3   82.5  82.6 10.2      63    82    90    94
#>  4 2020-03-04    2020-03-08       4   93    90.2 10.1      75    87    98   107
#>  5 2020-03-04    2020-03-09       5  102.  102.  10.8      86    92   105   115
#>  6 2020-03-04    2020-03-10       6  114   117.  18.6      95   105   124   156
#>  7 2020-03-04    2020-03-11       7  122.  127.  19.3     110   110   125   170
#>  8 2020-03-05    2020-03-06       1   65    66.0  8.46     54    60    68    80
#>  9 2020-03-05    2020-03-07       2   78.5  78.4  8.49     65    77    88    91
#> 10 2020-03-05    2020-03-08       3   88.5  89.5 10.9      70    78    91   103
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.23  2.81  2.79 0.100      4.45
#>  2 1      2020-03-04    2020-03-06       2  4.21  1.95  3.15 0.6        7.41
#>  3 1      2020-03-04    2020-03-07       3  5.07  3.33  3.26 0.3        7.41
#>  4 1      2020-03-04    2020-03-08       4  5.15  5.11  4.09 0.200      5.19
#>  5 1      2020-03-04    2020-03-09       5  5.97  9.05  4.82 0.100      8.90
#>  6 1      2020-03-04    2020-03-10       6  7.61 17.5   4.82 0.100     12.6 
#>  7 1      2020-03-04    2020-03-11       7  8.34 19.4   4.71 0.100     21.5 
#>  8 1      2020-03-05    2020-03-06       1  4.99  4.86  3.83 0.3        9.64
#>  9 1      2020-03-05    2020-03-07       2  5.37  4.41  3.45 0.100     10.4 
#> 10 1      2020-03-05    2020-03-08       3  6.10  8.42  4.12 0.100     11.9 
#> # … with 200 more rows
#>