Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.31 0.0630   2.23  2.23  2.30  2.42
#>  2 2020-03-04    2020-03-06       2   2.25  2.25 0.0663   2.12  2.23  2.27  2.33
#>  3 2020-03-04    2020-03-07       3   2.21  2.22 0.0752   2.11  2.20  2.30  2.33
#>  4 2020-03-04    2020-03-08       4   2.20  2.19 0.0894   2.01  2.17  2.26  2.33
#>  5 2020-03-04    2020-03-09       5   2.14  2.15 0.0993   2.05  2.05  2.18  2.32
#>  6 2020-03-04    2020-03-10       6   2.15  2.15 0.0873   2.06  2.09  2.22  2.32
#>  7 2020-03-04    2020-03-11       7   2.22  2.17 0.136    1.99  2.19  2.34  2.34
#>  8 2020-03-05    2020-03-06       1   2.21  2.22 0.0254   2.19  2.19  2.22  2.26
#>  9 2020-03-05    2020-03-07       2   2.15  2.17 0.0268   2.14  2.14  2.16  2.21
#> 10 2020-03-05    2020-03-08       3   2.09  2.09 0.0259   2.06  2.08  2.10  2.15
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss    crps   logs   bias sharpness
#>    <chr>         <date>       <int> <dbl>   <dbl>  <dbl>  <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -5.10 0.0235  -1.71   0.6      0.0637
#>  2 2020-03-04    2020-03-06       2 -4.82 0.0402  -0.273  0.6      0.0337
#>  3 2020-03-04    2020-03-07       3 -4.17 0.0466  -1.19   0.6      0.107 
#>  4 2020-03-04    2020-03-08       4 -3.91 0.0573  -0.755  0.8      0.0747
#>  5 2020-03-04    2020-03-09       5 -3.73 0.0459  -1.12   0.4      0.124 
#>  6 2020-03-04    2020-03-10       6 -2.64 0.0817  -0.851  1        0.0927
#>  7 2020-03-04    2020-03-11       7 -2.68 0.0894  -0.576  0.4      0.158 
#>  8 2020-03-05    2020-03-06       1 -6.13 0.0151  -2.43   0.6      0.0265
#>  9 2020-03-05    2020-03-07       2 -6.71 0.0100  -2.53   0.6      0.0259
#> 10 2020-03-05    2020-03-08       3 -7.21 0.00924 -2.48  -0.400    0.0201
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   76    73.6  6.92     63    75    82    82
#>  2 2020-03-04    2020-03-06       2   85.5  85.1  7.31     72    84    92    95
#>  3 2020-03-04    2020-03-07       3  106   106.   8.18     96    96   110   118
#>  4 2020-03-04    2020-03-08       4  124.  120.  11.1      99   123   127   131
#>  5 2020-03-04    2020-03-09       5  155   157.  17.9     133   133   159   186
#>  6 2020-03-04    2020-03-10       6  184   181.  11.2     161   176   189   200
#>  7 2020-03-04    2020-03-11       7  217   223.  26.1     190   190   228   261
#>  8 2020-03-05    2020-03-06       1   79    79.2  7.22     68    76    83    92
#>  9 2020-03-05    2020-03-07       2   95.5  93.9  9.87     76    93   102   109
#> 10 2020-03-05    2020-03-08       3  110.  112.  14.8      94    99   117   136
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  6.37  6.92  3.65   0.9      7.41
#>  2 1      2020-03-04    2020-03-06       2  6.92  8.43  3.93   0.8      6.67
#>  3 1      2020-03-04    2020-03-07       3  9.78 14.1   5.16   1       11.1 
#>  4 1      2020-03-04    2020-03-08       4  7.51 12.8   4.23   0.7      3.71
#>  5 1      2020-03-04    2020-03-09       5 11.4  30.9   6.00   1       23.7 
#>  6 1      2020-03-04    2020-03-10       6 18.8  34.0  10.3    1       10.4 
#>  7 1      2020-03-04    2020-03-11       7 11.6  42.4   6.14   1       34.1 
#>  8 1      2020-03-05    2020-03-06       1  4.67  3.96  3.49   0.6      5.19
#>  9 1      2020-03-05    2020-03-07       2  4.87  4.47  3.71   0.4      8.15
#> 10 1      2020-03-05    2020-03-08       3  5.83  5.14  3.83   0.3     14.8 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.31 0.0581   2.23  2.26  2.31  2.43
#>  2 2020-03-04    2020-03-06       2   2.29  2.29 0.0594   2.19  2.25  2.32  2.37
#>  3 2020-03-04    2020-03-07       3   2.22  2.24 0.0797   2.13  2.18  2.26  2.38
#>  4 2020-03-04    2020-03-08       4   2.18  2.22 0.104    2.08  2.08  2.19  2.37
#>  5 2020-03-04    2020-03-09       5   2.13  2.18 0.131    2.01  2.03  2.15  2.41
#>  6 2020-03-04    2020-03-10       6   2.03  2.13 0.154    1.94  1.94  2.05  2.40
#>  7 2020-03-04    2020-03-11       7   2.04  2.11 0.159    1.95  1.85  2.05  2.40
#>  8 2020-03-05    2020-03-06       1   2.22  2.23 0.0379   2.19  2.20  2.22  2.31
#>  9 2020-03-05    2020-03-07       2   2.16  2.15 0.0509   2.04  2.12  2.17  2.22
#> 10 2020-03-05    2020-03-08       3   2.10  2.10 0.0598   2.01  2.09  2.12  2.17
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps   logs   bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>  <dbl>  <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -4.76 0.0299 -1.78   0.8      0.0178
#>  2 1      2020-03-04    2020-03-06       2 -1.22 0.0790 -0.261  1        0.0559
#>  3 1      2020-03-04    2020-03-07       3 -3.29 0.0548 -0.971  1        0.0541
#>  4 1      2020-03-04    2020-03-08       4 -3.80 0.0480 -1.25   0.8      0.0653
#>  5 1      2020-03-04    2020-03-09       5 -3.69 0.0456 -1.22   0.6      0.0713
#>  6 1      2020-03-04    2020-03-10       6 -3.87 0.0260 -1.21  -0.200    0.0344
#>  7 1      2020-03-04    2020-03-11       7 -4.02 0.0267 -1.18   0        0.0688
#>  8 1      2020-03-05    2020-03-06       1 -4.92 0.0295 -1.67   0.8      0.0691
#>  9 1      2020-03-05    2020-03-07       2 -5.37 0.0247 -1.34   0        0.0716
#> 10 1      2020-03-05    2020-03-08       3 -5.05 0.0168 -1.57  -0.200    0.0762
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   66.5  67    9.16     51    59    72    78
#>  2 2020-03-04    2020-03-06       2   85.5  84.3 10.7      67    81    95    96
#>  3 2020-03-04    2020-03-07       3   98.5 101.  12.7      82    97   110   120
#>  4 2020-03-04    2020-03-08       4  123   124.  14.1      96   118   135   139
#>  5 2020-03-04    2020-03-09       5  140.  143.  17.1     114   134   153   167
#>  6 2020-03-04    2020-03-10       6  180   181.  20.8     151   174   196   215
#>  7 2020-03-04    2020-03-11       7  211   209.  26.1     177   177   216   256
#>  8 2020-03-05    2020-03-06       1   82.5  83.6 10.1      67    77    90    97
#>  9 2020-03-05    2020-03-07       2   91.5  95.3 12.8      80    85    98   118
#> 10 2020-03-05    2020-03-08       3  118   118.  13.6     100   107   121   146
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.38  2.98  3.45 0.200      9.64
#>  2 1      2020-03-04    2020-03-06       2  6.20  7.73  3.92 0.8       13.3 
#>  3 1      2020-03-04    2020-03-07       3  5.26  3.98  3.52 0.4       13.3 
#>  4 1      2020-03-04    2020-03-08       4  7.41 13.3   4.57 0.8        9.64
#>  5 1      2020-03-04    2020-03-09       5  8.05 16.2   4.59 0.9       17.0 
#>  6 1      2020-03-04    2020-03-10       6  8.72 24.7   5.94 1         15.6 
#>  7 1      2020-03-04    2020-03-11       7  8.83 23.3   4.90 1         25.2 
#>  8 1      2020-03-05    2020-03-06       1  6.42  7.93  3.80 1          9.64
#>  9 1      2020-03-05    2020-03-07       2  5.07  2.97  3.40 0.5        8.90
#> 10 1      2020-03-05    2020-03-08       3  7.48 12.7   4.36 0.8       15.6 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>