Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0468   2.20  2.28  2.32  2.39
#>  2 2020-03-04    2020-03-06       2   2.25  2.24 0.0480   2.17  2.23  2.28  2.32
#>  3 2020-03-04    2020-03-07       3   2.20  2.21 0.0819   2.10  2.17  2.24  2.39
#>  4 2020-03-04    2020-03-08       4   2.15  2.17 0.102    2.04  2.13  2.21  2.42
#>  5 2020-03-04    2020-03-09       5   2.13  2.13 0.109    1.96  2.10  2.19  2.34
#>  6 2020-03-04    2020-03-10       6   2.10  2.09 0.132    1.91  2.06  2.20  2.35
#>  7 2020-03-04    2020-03-11       7   2.06  2.06 0.152    1.85  2.05  2.19  2.35
#>  8 2020-03-05    2020-03-06       1   2.23  2.22 0.0335   2.16  2.20  2.24  2.27
#>  9 2020-03-05    2020-03-07       2   2.16  2.14 0.0517   2.04  2.15  2.20  2.20
#> 10 2020-03-05    2020-03-08       3   2.08  2.06 0.0617   1.94  2.07  2.12  2.13
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 8
#>    forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -5.85 0.0171 -2.05  0.8      0.0305
#>  2 2020-03-04    2020-03-06       2 -5.03 0.0295 -1.59  0.8      0.0487
#>  3 2020-03-04    2020-03-07       3 -4.47 0.0329 -1.47  0.8      0.0649
#>  4 2020-03-04    2020-03-08       4 -4.20 0.0344 -1.38  0.8      0.0680
#>  5 2020-03-04    2020-03-09       5 -4.13 0.0433 -0.951 0.8      0.0849
#>  6 2020-03-04    2020-03-10       6 -3.90 0.0467 -0.828 0.7      0.107 
#>  7 2020-03-04    2020-03-11       7 -3.76 0.0430 -0.858 0.7      0.141 
#>  8 2020-03-05    2020-03-06       1 -6.27 0.0165 -2.02  0.7      0.0289
#>  9 2020-03-05    2020-03-07       2 -6.00 0.0132 -1.95  0.6      0.0468
#> 10 2020-03-05    2020-03-08       3 -5.18 0.0189 -2.05  0.200    0.0461
#> # … with 95 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   59    58.8  6.18     48    58    64    70
#>  2 2020-03-04    2020-03-06       2   70.5  72.8  8.53     56    69    77    85
#>  3 2020-03-04    2020-03-07       3   87.5  85.5  8.59     72    87    97    97
#>  4 2020-03-04    2020-03-08       4   92    92.5  6.43     81    86    95   101
#>  5 2020-03-04    2020-03-09       5  109   110.  16.6      90    90   111   132
#>  6 2020-03-04    2020-03-10       6  130.  128.  10.6     114   130   142   142
#>  7 2020-03-04    2020-03-11       7  143   141.  16.9     113   135   157   166
#>  8 2020-03-05    2020-03-06       1   69    70.3  8.08     59    62    72    85
#>  9 2020-03-05    2020-03-07       2   78    79.4  5.34     72    72    78    87
#> 10 2020-03-05    2020-03-08       3   87.5  87.2 10.6      72    76    92   104
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.05  2.54  2.99 0.200      5.93
#>  2 1      2020-03-04    2020-03-06       2  4.18  2.24  3.11 0.4        5.93
#>  3 1      2020-03-04    2020-03-07       3  4.29  1.95  3.20 0.4        9.64
#>  4 1      2020-03-04    2020-03-08       4  6.04  6.05  3.55 0          6.67
#>  5 1      2020-03-04    2020-03-09       5  5.63  6.23  4.14 0.4       23.0 
#>  6 1      2020-03-04    2020-03-10       6  6.37  7.74  3.82 0.100     12.6 
#>  7 1      2020-03-04    2020-03-11       7  8.21 17.1   4.63 0         17.8 
#>  8 1      2020-03-05    2020-03-06       1  4.20  2.77  3.34 0.3        8.15
#>  9 1      2020-03-05    2020-03-07       2  6.13  5.74  3.48 0          5.93
#> 10 1      2020-03-05    2020-03-08       3  6.78  9.44  4.01 0.100     12.6 
#> # … with 95 more rows
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.29 0.0410   2.22  2.28  2.32  2.37
#>  2 2020-03-04    2020-03-06       2   2.25  2.24 0.0575   2.18  2.23  2.27  2.36
#>  3 2020-03-04    2020-03-07       3   2.18  2.19 0.0845   2.09  2.13  2.20  2.38
#>  4 2020-03-04    2020-03-08       4   2.14  2.15 0.103    1.95  2.10  2.17  2.30
#>  5 2020-03-04    2020-03-09       5   2.07  2.09 0.105    1.94  2.03  2.14  2.27
#>  6 2020-03-04    2020-03-10       6   2.01  2.05 0.131    1.81  1.97  2.10  2.24
#>  7 2020-03-04    2020-03-11       7   1.95  2.00 0.147    1.80  1.87  2.03  2.31
#>  8 2020-03-05    2020-03-06       1   2.20  2.20 0.0170   2.17  2.20  2.21  2.23
#>  9 2020-03-05    2020-03-07       2   2.15  2.15 0.0266   2.12  2.13  2.16  2.21
#> 10 2020-03-05    2020-03-08       3   2.09  2.09 0.0342   2.03  2.05  2.10  2.14
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss    crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>   <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -5.67 0.0226  -1.72  0.8      0.0450
#>  2 1      2020-03-04    2020-03-06       2 -2.32 0.0461  -0.692 1        0.0230
#>  3 1      2020-03-04    2020-03-07       3 -5.11 0.0197  -1.69  0.7      0.0516
#>  4 1      2020-03-04    2020-03-08       4 -4.61 0.0308  -1.77  0.9      0.0319
#>  5 1      2020-03-04    2020-03-09       5 -5.05 0.0188  -1.65  0.5      0.0601
#>  6 1      2020-03-04    2020-03-10       6 -4.94 0.0188  -1.52  0.4      0.0584
#>  7 1      2020-03-04    2020-03-11       7 -4.37 0.0347  -1.03  0.4      0.0820
#>  8 1      2020-03-05    2020-03-06       1 -7.67 0.00763 -2.83  0.8      0.0121
#>  9 1      2020-03-05    2020-03-07       2 -7.97 0.00632 -2.66  0.5      0.0247
#> 10 1      2020-03-05    2020-03-08       3 -6.15 0.0213  -1.09  0.200    0.0220
#> # … with 200 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   60    59.4  7.20     48    52    61    71
#>  2 2020-03-04    2020-03-06       2   65    67.6  8.37     54    61    69    81
#>  3 2020-03-04    2020-03-07       3   82.5  82.0  7.76     66    78    86    89
#>  4 2020-03-04    2020-03-08       4   89.5  91.0 10.1      78    78    90   109
#>  5 2020-03-04    2020-03-09       5  107   110.  14.3      93    93   111   139
#>  6 2020-03-04    2020-03-10       6  126   126.  21.7      95   109   138   160
#>  7 2020-03-04    2020-03-11       7  134.  132.  16.1     110   112   135   158
#>  8 2020-03-05    2020-03-06       1   67    67.4 11.4      40    67    78    78
#>  9 2020-03-05    2020-03-07       2   79.5  80.7  9.24     73    78    85    98
#> 10 2020-03-05    2020-03-08       3   97    93.6 10.7      78    94   107   107
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.33  3.6   3.36 0.200      6.67
#>  2 1      2020-03-04    2020-03-06       2  4.89  4.71  3.68 0.3       10.4 
#>  3 1      2020-03-04    2020-03-07       3  5.71  5.72  3.82 0.100      4.45
#>  4 1      2020-03-04    2020-03-08       4  5.77  5.87  3.57 0.100     11.9 
#>  5 1      2020-03-04    2020-03-09       5  5.45  6.05  4.07 0.3       12.6 
#>  6 1      2020-03-04    2020-03-10       6  7.25 12.0   4.17 0         21.5 
#>  7 1      2020-03-04    2020-03-11       7 10.6  28.4   5.35 0         17.8 
#>  8 1      2020-03-05    2020-03-06       1  4.64  2.74  3.18 0.3        8.90
#>  9 1      2020-03-05    2020-03-07       2  5.21  6.12  4.36 0.100      5.19
#> 10 1      2020-03-05    2020-03-08       3  5.21  4.52  3.60 0.200     13.3 
#> # … with 200 more rows
#>