Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
An uninitialised bsts model

horizon
Numeric, the time horizon over which to predict

samples
Numeric, the number of samples to from the posterior of the model fit. Note that twice this number of MCMC steps will be taken with half used as burn in.

timeout
Numeric, the number of seconds to allow before terminating model fitting. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0168   2.27  2.27  2.29  2.32
#>  2 2020-03-04    2020-03-06       2   2.23  2.23 0.0182   2.18  2.22  2.24  2.25
#>  3 2020-03-04    2020-03-07       3   2.18  2.18 0.0206   2.15  2.17  2.19  2.22
#>  4 2020-03-04    2020-03-08       4   2.13  2.13 0.0297   2.08  2.12  2.14  2.19
#>  5 2020-03-04    2020-03-09       5   2.08  2.08 0.0208   2.04  2.07  2.09  2.12
#>  6 2020-03-04    2020-03-10       6   2.02  2.03 0.0271   1.99  2.01  2.05  2.07
#>  7 2020-03-04    2020-03-11       7   1.97  1.97 0.0238   1.93  1.96  1.99  2.01
#>  8 2020-03-05    2020-03-06       1   2.21  2.23 0.103    2.14  2.17  2.23  2.47
#>  9 2020-03-05    2020-03-07       2   2.15  2.16 0.0852   2.05  2.13  2.17  2.37
#> 10 2020-03-05    2020-03-08       3   2.07  2.08 0.120    1.91  2.03  2.14  2.35
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 8
#>    forecast_date date       horizon   dss    crps    logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>   <dbl>   <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -6.34 0.0134  -2.61     1     0.0153 
#>  2 2020-03-04    2020-03-06       2 -4.11 0.0279  -0.0399   0.9   0.00820
#>  3 2020-03-04    2020-03-07       3 -4.56 0.0246  -1.82     1     0.0206 
#>  4 2020-03-04    2020-03-08       4 -6.30 0.0170  -1.86     0.8   0.0232 
#>  5 2020-03-04    2020-03-09       5 -7.00 0.0119  -2.13     0.8   0.0117 
#>  6 2020-03-04    2020-03-10       6 -7.32 0.00808 -2.40     0.5   0.0343 
#>  7 2020-03-04    2020-03-11       7 -3.62 0.0321  -1.49     0     0.0262 
#>  8 2020-03-05    2020-03-06       1 -4.48 0.0183  -1.93     0.7   0.0468 
#>  9 2020-03-05    2020-03-07       2 -5.01 0.0114  -2.33     0.6   0.0318 
#> 10 2020-03-05    2020-03-08       3 -4.28 0.0293  -1.20     0.4   0.0974 
#> # … with 95 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   57.5  59.4  6.75     48    55    62    72
#>  2 2020-03-04    2020-03-06       2   70.5  68.7  9.45     53    67    76    82
#>  3 2020-03-04    2020-03-07       3   83.5  83.5  6.85     74    80    88    97
#>  4 2020-03-04    2020-03-08       4   91.5  88.4 11.3      68    86    95   101
#>  5 2020-03-04    2020-03-09       5   98.5  98.6 10.9      79    90   102   114
#>  6 2020-03-04    2020-03-10       6  120.  123.  16.7      93   112   125   151
#>  7 2020-03-04    2020-03-11       7  131   136   12.0     124   124   132   161
#>  8 2020-03-05    2020-03-06       1   74    72.2  7.21     63    74    84    84
#>  9 2020-03-05    2020-03-07       2   78    79.5  6.65     73    73    79    92
#> 10 2020-03-05    2020-03-08       3   97.5  96.8 10.3      82    96   104   118
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.03  2.68  3.18 0.3        5.19
#>  2 1      2020-03-04    2020-03-06       2  4.62  2.67  3.34 0.4        8.15
#>  3 1      2020-03-04    2020-03-07       3  4.22  2.69  3.20 0.100      6.67
#>  4 1      2020-03-04    2020-03-08       4  6.35  7.86  3.77 0          7.41
#>  5 1      2020-03-04    2020-03-09       5  7.51 11.6   4.23 0         11.9 
#>  6 1      2020-03-04    2020-03-10       6  6.88 12.6   4.74 0.200     11.1 
#>  7 1      2020-03-04    2020-03-11       7 12.3  24.9   5.49 0          9.64
#>  8 1      2020-03-05    2020-03-06       1  3.86  2.2   3.20 0.6        8.90
#>  9 1      2020-03-05    2020-03-07       2  5.50  5.83  3.69 0.100      6.67
#> 10 1      2020-03-05    2020-03-08       3  4.84  3.52  3.34 0.200      6.67
#> # … with 95 more rows
#> 
#> $raw_rt_forecast
#> # A tibble: 1,330 x 6
#>    obs_sample forecast_date sample date          rt horizon
#>    <chr>      <chr>          <int> <date>     <dbl>   <int>
#>  1 1          2020-03-04         1 2020-03-05  2.29       1
#>  2 1          2020-03-04         2 2020-03-05  2.32       1
#>  3 1          2020-03-04         3 2020-03-05  2.29       1
#>  4 1          2020-03-04         4 2020-03-05  2.28       1
#>  5 1          2020-03-04         5 2020-03-05  2.30       1
#>  6 1          2020-03-04         6 2020-03-05  2.28       1
#>  7 1          2020-03-04         7 2020-03-05  2.27       1
#>  8 1          2020-03-04         8 2020-03-05  2.27       1
#>  9 1          2020-03-04         9 2020-03-05  2.31       1
#> 10 1          2020-03-04        10 2020-03-05  2.28       1
#> # … with 1,320 more rows
#> 
#> $raw_case_forecast
#> # A tibble: 1,050 x 6
#>    obs_sample forecast_date sample date       cases horizon
#>    <chr>      <chr>          <dbl> <date>     <int>   <int>
#>  1 1          2020-03-04         1 2020-03-05    65       1
#>  2 1          2020-03-04         1 2020-03-06    75       2
#>  3 1          2020-03-04         1 2020-03-07    83       3
#>  4 1          2020-03-04         1 2020-03-08   101       4
#>  5 1          2020-03-04         1 2020-03-09   102       5
#>  6 1          2020-03-04         1 2020-03-10   146       6
#>  7 1          2020-03-04         1 2020-03-11   130       7
#>  8 1          2020-03-04         2 2020-03-05    48       1
#>  9 1          2020-03-04         2 2020-03-06    76       2
#> 10 1          2020-03-04         2 2020-03-07    74       3
#> # … with 1,040 more rows
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.31 0.0544   2.26  2.26  2.30  2.42
#>  2 2020-03-04    2020-03-06       2   2.24  2.28 0.121    2.17  2.22  2.26  2.58
#>  3 2020-03-04    2020-03-07       3   2.21  2.24 0.120    2.11  2.16  2.23  2.40
#>  4 2020-03-04    2020-03-08       4   2.15  2.19 0.143    2.04  2.07  2.18  2.43
#>  5 2020-03-04    2020-03-09       5   2.10  2.16 0.176    1.98  2.06  2.14  2.59
#>  6 2020-03-04    2020-03-10       6   2.07  2.11 0.192    1.88  2.03  2.15  2.33
#>  7 2020-03-04    2020-03-11       7   2.03  2.08 0.225    1.81  1.92  2.05  2.43
#>  8 2020-03-05    2020-03-06       1   2.21  2.21 0.0231   2.18  2.19  2.21  2.25
#>  9 2020-03-05    2020-03-07       2   2.17  2.17 0.0406   2.13  2.13  2.18  2.25
#> 10 2020-03-05    2020-03-08       3   2.12  2.12 0.0642   2.05  2.11  2.15  2.25
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss    crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>   <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -6.95 0.0113  -2.47  0.9      0.0182
#>  2 1      2020-03-04    2020-03-06       2 -3.99 0.0304  -1.53  0.9      0.0224
#>  3 1      2020-03-04    2020-03-07       3 -5.63 0.0283  -1.34  0.8      0.0366
#>  4 1      2020-03-04    2020-03-08       4 -5.88 0.0260  -0.702 0.8      0.0260
#>  5 1      2020-03-04    2020-03-09       5 -5.78 0.0219  -1.60  0.8      0.0403
#>  6 1      2020-03-04    2020-03-10       6 -6.01 0.0115  -2.14  0.8      0.0492
#>  7 1      2020-03-04    2020-03-11       7 -5.37 0.0179  -1.86  0.200    0.0500
#>  8 1      2020-03-05    2020-03-06       1 -6.53 0.00902 -2.81  0.9      0.0118
#>  9 1      2020-03-05    2020-03-07       2 -5.59 0.0191  -1.73  0.6      0.0630
#> 10 1      2020-03-05    2020-03-08       3 -5.17 0.0203  -1.53  0.7      0.0435
#> # … with 200 more rows
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   60.5  61.4  6.06     51    58    65    68
#>  2 2020-03-04    2020-03-06       2   76    75.4 11.1      58    70    79    91
#>  3 2020-03-04    2020-03-07       3   82.5  82.8  8.81     65    81    89    96
#>  4 2020-03-04    2020-03-08       4   94    95.0 14.9      74    74    95   120
#>  5 2020-03-04    2020-03-09       5  108.  114   16.8      87   104   121   144
#>  6 2020-03-04    2020-03-10       6  120.  123.  13.3     109   109   123   152
#>  7 2020-03-04    2020-03-11       7  148.  148.  28.8     105   137   163   190
#>  8 2020-03-05    2020-03-06       1   69.5  70.1  7.21     59    60    70    82
#>  9 2020-03-05    2020-03-07       2   81    80.2  9.57     70    72    82   103
#> 10 2020-03-05    2020-03-08       3   95.5  94.1  9.70     81    94   104   113
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 9
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  3.86  2.06  3.01 0.4        8.15
#>  2 1      2020-03-04    2020-03-06       2  4.92  3.17  3.22 0.7        8.15
#>  3 1      2020-03-04    2020-03-07       3  5.22  4.14  3.33 0.100      9.64
#>  4 1      2020-03-04    2020-03-08       4  5.86  7.6   4.14 0.4       17.8 
#>  5 1      2020-03-04    2020-03-09       5  5.56  5.48  4.09 0.3        8.90
#>  6 1      2020-03-04    2020-03-10       6 10.5  18.1   5.70 0          9.64
#>  7 1      2020-03-04    2020-03-11       7  9.36 23.6   5.20 0         20.8 
#>  8 1      2020-03-05    2020-03-06       1  3.98  2.1   3.24 0.4        7.41
#>  9 1      2020-03-05    2020-03-07       2  5.31  5.21  3.65 0.200     12.6 
#> 10 1      2020-03-05    2020-03-08       3  4.65  3.42  3.31 0.200      9.64
#> # … with 200 more rows
#> 
#> $raw_rt_forecast
#> # A tibble: 2,660 x 6
#>    obs_sample forecast_date sample date          rt horizon
#>    <chr>      <chr>          <int> <date>     <dbl>   <int>
#>  1 1          2020-03-04         1 2020-03-05  2.30       1
#>  2 1          2020-03-04         2 2020-03-05  2.29       1
#>  3 1          2020-03-04         3 2020-03-05  2.33       1
#>  4 1          2020-03-04         4 2020-03-05  2.29       1
#>  5 1          2020-03-04         5 2020-03-05  2.27       1
#>  6 1          2020-03-04         6 2020-03-05  2.30       1
#>  7 1          2020-03-04         7 2020-03-05  2.25       1
#>  8 1          2020-03-04         8 2020-03-05  2.27       1
#>  9 1          2020-03-04         9 2020-03-05  2.27       1
#> 10 1          2020-03-04        10 2020-03-05  2.29       1
#> # … with 2,650 more rows
#> 
#> $raw_case_forecast
#> # A tibble: 2,100 x 6
#>    obs_sample forecast_date sample date       cases horizon
#>    <chr>      <chr>          <dbl> <date>     <int>   <int>
#>  1 1          2020-03-04         1 2020-03-05    59       1
#>  2 1          2020-03-04         1 2020-03-06    89       2
#>  3 1          2020-03-04         1 2020-03-07    78       3
#>  4 1          2020-03-04         1 2020-03-08    80       4
#>  5 1          2020-03-04         1 2020-03-09   117       5
#>  6 1          2020-03-04         1 2020-03-10   120       6
#>  7 1          2020-03-04         1 2020-03-11   110       7
#>  8 1          2020-03-04         2 2020-03-05    65       1
#>  9 1          2020-03-04         2 2020-03-06    71       2
#> 10 1          2020-03-04         2 2020-03-07    82       3
#> # … with 2,090 more rows
#>