Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  observations = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE
)

Arguments

observations
A dataframe of observations against which to score. Should contain a date and rt column.

model
An uninitialised bsts model

horizon
Numeric, the time horizon over which to predict

samples
Numeric, the number of samples to from the posterior of the model fit. Note that twice this number of MCMC steps will be taken with half used as burn in.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

Value

Examples

## Observed data
observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Evaluate a model
evaluate_model(observations,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10)

#> $forecasts
#> # A tibble: 126 x 10
#>    forecast_date date       horizon bottom lower median  mean upper   top    sd
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 2020-01-04    2020-01-05       1   3.52  3.91   4.06  4.04  4.17  4.61 0.342
#>  2 2020-01-04    2020-01-06       2   4.07  4.51   5.15  5.01  5.51  5.72 0.626
#>  3 2020-01-04    2020-01-07       3   5.24  5.89   6.11  6.05  6.45  6.60 0.486
#>  4 2020-01-04    2020-01-08       4   5.91  6.98   7.35  7.21  7.72  8.15 0.783
#>  5 2020-01-04    2020-01-09       5   6.45  7.86   8.64  8.43  9.22  9.85 1.13 
#>  6 2020-01-04    2020-01-10       6   7.11  8.67   9.57  9.31 10.2  10.4  1.16 
#>  7 2020-01-04    2020-01-11       7   8.16  9.35  10.4  10.2  11.1  11.5  1.22 
#>  8 2020-01-05    2020-01-06       1   4.41  4.96   5.04  5.22  5.26  6.76 0.745
#>  9 2020-01-05    2020-01-07       2   4.15  5.84   5.95  6.09  6.33  8.27 1.22 
#> 10 2020-01-05    2020-01-08       3   3.48  6.78   7.15  6.90  7.51  8.98 1.70 
#> # … with 116 more rows
#> 
#> $scores
#> # A tibble: 98 x 8
#>    forecast_date date       horizon    dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int>  <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-01-04    2020-01-05       1 -2.24  0.0686 -0.353   0.6     0.189
#>  2 2020-01-04    2020-01-06       2 -1.04  0.184   0.750   0.6     0.690
#>  3 2020-01-04    2020-01-07       3 -1.54  0.129   0.346   0.7     0.535
#>  4 2020-01-04    2020-01-08       4 -0.512 0.255   0.932   0.7     0.620
#>  5 2020-01-04    2020-01-09       5  0.293 0.402   1.36    0.7     1.08 
#>  6 2020-01-04    2020-01-10       6  0.275 0.411   1.34    0.7     1.07 
#>  7 2020-01-04    2020-01-11       7  0.312 0.412   1.46    0.7     1.18 
#>  8 2020-01-05    2020-01-06       1 -0.600 0.0960 -0.154   0.6     0.327
#>  9 2020-01-05    2020-01-07       2  0.298 0.149   0.278   0.4     0.520
#> 10 2020-01-05    2020-01-08       3  0.965 0.235   0.730   0.5     0.563
#> # … with 88 more rows
#>