Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.35  2.35 0.0621   2.26  2.30  2.38  2.43
#>  2 2020-03-04    2020-03-06       2   2.31  2.31 0.0626   2.20  2.25  2.33  2.43
#>  3 2020-03-04    2020-03-07       3   2.27  2.29 0.0840   2.20  2.20  2.29  2.47
#>  4 2020-03-04    2020-03-08       4   2.26  2.26 0.119    2.04  2.21  2.28  2.49
#>  5 2020-03-04    2020-03-09       5   2.19  2.21 0.131    2.07  2.07  2.20  2.50
#>  6 2020-03-04    2020-03-10       6   2.17  2.21 0.148    1.97  2.10  2.22  2.50
#>  7 2020-03-04    2020-03-11       7   2.11  2.15 0.201    1.91  1.99  2.13  2.62
#>  8 2020-03-05    2020-03-06       1   2.26  2.25 0.0382   2.17  2.24  2.26  2.33
#>  9 2020-03-05    2020-03-07       2   2.19  2.19 0.0417   2.11  2.17  2.21  2.25
#> 10 2020-03-05    2020-03-08       3   2.16  2.17 0.0728   2.08  2.13  2.17  2.34
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -3.45 0.0564 -1.12    0.8    0.0797
#>  2 2020-03-04    2020-03-06       2 -2.01 0.0802 -0.490   1      0.0593
#>  3 2020-03-04    2020-03-07       3 -2.00 0.0960 -0.431   1      0.0855
#>  4 2020-03-04    2020-03-08       4 -2.37 0.113   1.58    0.8    0.0575
#>  5 2020-03-04    2020-03-09       5 -2.63 0.0871 -0.667   1      0.139 
#>  6 2020-03-04    2020-03-10       6 -2.26 0.116  -0.295   0.8    0.0900
#>  7 2020-03-04    2020-03-11       7 -2.79 0.0648 -0.743   0.6    0.145 
#>  8 2020-03-05    2020-03-06       1 -3.66 0.0488  0.834   0.8    0.0178
#>  9 2020-03-05    2020-03-07       2 -5.45 0.0286 -1.17    0.6    0.0315
#> 10 2020-03-05    2020-03-08       3 -4.52 0.0346 -1.40    0.6    0.0363
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1    71   71.7  4.97     65    68    72    80
#>  2 2020-03-04    2020-03-06       2    87   86.7  7.21     76    81    88   101
#>  3 2020-03-04    2020-03-07       3   107  103.  11.7      85    94   109   123
#>  4 2020-03-04    2020-03-08       4   128. 128.  13.8     108   121   137   156
#>  5 2020-03-04    2020-03-09       5   144. 152.  24.6     120   120   145   195
#>  6 2020-03-04    2020-03-10       6   192  185.  26.1     142   179   206   225
#>  7 2020-03-04    2020-03-11       7   230  234.  47.3     178   204   248   345
#>  8 2020-03-05    2020-03-06       1    81   83.8 11.2      71    71    82   108
#>  9 2020-03-05    2020-03-07       2    99   97.2  8.60     82    98   108   108
#> 10 2020-03-05    2020-03-08       3   117  116.  10.7      96   108   119   133
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  6.51  6.15  4.16   1        3.71
#>  2 1      2020-03-04    2020-03-06       2  7.86  9.99  4.62   1        5.19
#>  3 1      2020-03-04    2020-03-07       3  6.74  9.8   4.02   0.8     14.1 
#>  4 1      2020-03-04    2020-03-08       4  8.96 18.3   4.98   1       12.6 
#>  5 1      2020-03-04    2020-03-09       5  8.63 22.5   4.86   1       31.9 
#>  6 1      2020-03-04    2020-03-10       6  9.46 29.4   5.14   1       20.0 
#>  7 1      2020-03-04    2020-03-11       7  9.84 43.6   5.69   1       32.6 
#>  8 1      2020-03-05    2020-03-06       1  5.76  5.44  3.58   0.7     11.9 
#>  9 1      2020-03-05    2020-03-07       2  5.47  6.22  3.81   0.6      8.90
#> 10 1      2020-03-05    2020-03-08       3  6.46  9.25  4.18   0.8     11.9 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.31  2.31 0.0383   2.25  2.28  2.32  2.36
#>  2 2020-03-04    2020-03-06       2   2.26  2.26 0.0650   2.17  2.24  2.29  2.33
#>  3 2020-03-04    2020-03-07       3   2.22  2.25 0.0988   2.09  2.19  2.22  2.43
#>  4 2020-03-04    2020-03-08       4   2.18  2.22 0.126    2.00  2.14  2.19  2.42
#>  5 2020-03-04    2020-03-09       5   2.14  2.18 0.146    1.93  2.08  2.16  2.40
#>  6 2020-03-04    2020-03-10       6   2.09  2.12 0.175    1.79  2.04  2.12  2.37
#>  7 2020-03-04    2020-03-11       7   2.06  2.08 0.178    1.72  2.03  2.13  2.36
#>  8 2020-03-05    2020-03-06       1   2.22  2.21 0.0360   2.14  2.21  2.25  2.26
#>  9 2020-03-05    2020-03-07       2   2.18  2.18 0.0685   2.06  2.17  2.23  2.29
#> 10 2020-03-05    2020-03-08       3   2.15  2.14 0.0944   1.98  2.07  2.18  2.31
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -5.31 0.0264 -1.77  0.6     0.0391 
#>  2 1      2020-03-04    2020-03-06       2 -4.76 0.0390 -0.735 0.6     0.0340 
#>  3 1      2020-03-04    2020-03-07       3 -4.22 0.0481 -1.64  0.8     0.0143 
#>  4 1      2020-03-04    2020-03-08       4 -3.94 0.0609 44.7   0.8     0.00916
#>  5 1      2020-03-04    2020-03-09       5 -3.98 0.0590  0.263 0.8     0.0261 
#>  6 1      2020-03-04    2020-03-10       6 -3.74 0.0527 -0.785 0.8     0.0571 
#>  7 1      2020-03-04    2020-03-11       7 -3.75 0.0316 -1.55  0.8     0.0534 
#>  8 1      2020-03-05    2020-03-06       1 -6.31 0.0110 -2.04  0.200   0.0483 
#>  9 1      2020-03-05    2020-03-07       2 -6.01 0.0173 -1.94  0.4     0.0313 
#> 10 1      2020-03-05    2020-03-08       3 -5.72 0.0195 -1.65  0.4     0.0420 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   71.5  69.4  8.76     57    68    76    84
#>  2 2020-03-04    2020-03-06       2   85    84.4  8.67     68    80    90    94
#>  3 2020-03-04    2020-03-07       3  102   102.  10.1      85    94   106   114
#>  4 2020-03-04    2020-03-08       4  120.  122.  12.5      99   113   120   144
#>  5 2020-03-04    2020-03-09       5  142.  148.  17.5     132   132   144   187
#>  6 2020-03-04    2020-03-10       6  172   175.  21.4     144   160   180   216
#>  7 2020-03-04    2020-03-11       7  212   210.  25.0     157   201   222   236
#>  8 2020-03-05    2020-03-06       1   86    82.7  9.96     68    85    96    96
#>  9 2020-03-05    2020-03-07       2   94    94.8 12.4      70    86   101   109
#> 10 2020-03-05    2020-03-08       3  120.  116.  14.6      91   108   124   137
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.07  5.57  4.26   0.6      5.93
#>  2 1      2020-03-04    2020-03-06       2  6.08  7.81  4.46   0.8      4.45
#>  3 1      2020-03-04    2020-03-07       3  6.59  9.11  3.98   0.8     14.8 
#>  4 1      2020-03-04    2020-03-08       4  7.01 11.7   4.30   0.6     11.1 
#>  5 1      2020-03-04    2020-03-09       5  9.79 21.2   5.69   1        9.64
#>  6 1      2020-03-04    2020-03-10       6  7.81 17.0   4.64   0.8     17.8 
#>  7 1      2020-03-04    2020-03-11       7 10.1  34.6   6.47   0.8     10.4 
#>  8 1      2020-03-05    2020-03-06       1  5.17  4.54  3.61   0.7     11.1 
#>  9 1      2020-03-05    2020-03-07       2  5.58  5.06  3.87   0.5     19.3 
#> 10 1      2020-03-05    2020-03-08       3  5.58  6.98  4.05   0.4     14.8 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>