Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0217   2.26  2.28  2.30  2.34
#>  2 2020-03-04    2020-03-06       2   2.23  2.23 0.0173   2.20  2.23  2.24  2.26
#>  3 2020-03-04    2020-03-07       3   2.18  2.18 0.0225   2.15  2.15  2.18  2.22
#>  4 2020-03-04    2020-03-08       4   2.14  2.13 0.0207   2.10  2.13  2.16  2.16
#>  5 2020-03-04    2020-03-09       5   2.09  2.07 0.0399   2.00  2.07  2.10  2.13
#>  6 2020-03-04    2020-03-10       6   2.03  2.04 0.0507   1.94  2.01  2.06  2.13
#>  7 2020-03-04    2020-03-11       7   1.98  1.99 0.0554   1.88  1.96  1.99  2.10
#>  8 2020-03-05    2020-03-06       1   2.21  2.21 0.0174   2.18  2.20  2.22  2.24
#>  9 2020-03-05    2020-03-07       2   2.15  2.15 0.0312   2.10  2.13  2.16  2.20
#> 10 2020-03-05    2020-03-08       3   2.10  2.10 0.0461   2.04  2.06  2.12  2.17
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss    crps  logs   bias sharpness
#>    <chr>         <date>       <int> <dbl>   <dbl> <dbl>  <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -6.57 0.0131  -2.50  0.8      0.0157
#>  2 2020-03-04    2020-03-06       2 -2.45 0.0302  -1.09  1        0.0139
#>  3 2020-03-04    2020-03-07       3 -4.45 0.0265  -1.78  1        0.0291
#>  4 2020-03-04    2020-03-08       4 -5.57 0.0202  -2.10  0.8      0.0169
#>  5 2020-03-04    2020-03-09       5 -6.40 0.0156  -1.89  0.4      0.0252
#>  6 2020-03-04    2020-03-10       6 -5.97 0.0112  -2.16  0.4      0.0336
#>  7 2020-03-04    2020-03-11       7 -5.62 0.0186  -1.93 -0.6      0.0330
#>  8 2020-03-05    2020-03-06       1 -6.85 0.0119  -2.54  0.8      0.0165
#>  9 2020-03-05    2020-03-07       2 -6.96 0.00831 -2.39  0.200    0.0338
#> 10 2020-03-05    2020-03-08       3 -6.26 0.0156  -1.80  0        0.0539
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   71    70.8  6.66     62    64    76    78
#>  2 2020-03-04    2020-03-06       2   87    86.1  9.12     71    79    91   103
#>  3 2020-03-04    2020-03-07       3  104.  103.   9.34     89    98   107   118
#>  4 2020-03-04    2020-03-08       4  116.  116.  10.3      95   112   125   129
#>  5 2020-03-04    2020-03-09       5  142   144.  12.3     117   137   149   160
#>  6 2020-03-04    2020-03-10       6  166.  168.  14.0     151   154   169   196
#>  7 2020-03-04    2020-03-11       7  190.  193.  12.8     172   182   196   218
#>  8 2020-03-05    2020-03-06       1   87    84.1  9.06     69    82    92    97
#>  9 2020-03-05    2020-03-07       2   90.5  91.3 10.4      80    84    95   113
#> 10 2020-03-05    2020-03-08       3  116.  119.  10.2     105   105   118   137
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.21  4.58  3.21 0.8        8.90
#>  2 1      2020-03-04    2020-03-06       2  6.61  8.71  3.98 0.8        9.64
#>  3 1      2020-03-04    2020-03-07       3  7.39 10.5   4.02 1          6.67
#>  4 1      2020-03-04    2020-03-08       4  6.52  9.71  4.19 0.8       11.1 
#>  5 1      2020-03-04    2020-03-09       5 10.6  21.5   5.08 1          8.90
#>  6 1      2020-03-04    2020-03-10       6  9.27 19.6   5.07 1         17.8 
#>  7 1      2020-03-04    2020-03-11       7  9.46 19.0   5.24 1         10.4 
#>  8 1      2020-03-05    2020-03-06       1  5.97  7.51  3.91 0.6        7.41
#>  9 1      2020-03-05    2020-03-07       2  4.69  2.71  3.31 0.200      9.64
#> 10 1      2020-03-05    2020-03-08       3  7.55 11.4   4.23 1         12.6 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.30 0.0623   2.23  2.26  2.31  2.42
#>  2 2020-03-04    2020-03-06       2   2.27  2.26 0.0719   2.16  2.20  2.28  2.39
#>  3 2020-03-04    2020-03-07       3   2.22  2.22 0.113    2.01  2.11  2.26  2.41
#>  4 2020-03-04    2020-03-08       4   2.19  2.18 0.131    2.03  2.12  2.26  2.42
#>  5 2020-03-04    2020-03-09       5   2.16  2.14 0.181    1.96  2.02  2.21  2.42
#>  6 2020-03-04    2020-03-10       6   2.12  2.10 0.172    1.77  1.99  2.22  2.32
#>  7 2020-03-04    2020-03-11       7   2.06  2.07 0.182    1.85  1.93  2.18  2.43
#>  8 2020-03-05    2020-03-06       1   2.23  2.23 0.0373   2.18  2.23  2.26  2.29
#>  9 2020-03-05    2020-03-07       2   2.19  2.20 0.0843   2.06  2.15  2.23  2.33
#> 10 2020-03-05    2020-03-08       3   2.13  2.15 0.117    1.88  2.09  2.21  2.28
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -4.73 0.0294 -1.28  0.200    0.0944
#>  2 1      2020-03-04    2020-03-06       2 -3.70 0.0723 -0.268 0.8      0.0710
#>  3 1      2020-03-04    2020-03-07       3 -3.32 0.0763 -0.538 0.6      0.148 
#>  4 1      2020-03-04    2020-03-08       4 -3.16 0.0821 -0.373 0.6      0.189 
#>  5 1      2020-03-04    2020-03-09       5 -2.70 0.0982 -0.139 0.4      0.220 
#>  6 1      2020-03-04    2020-03-10       6 -2.84 0.0995 -0.128 0.6      0.215 
#>  7 1      2020-03-04    2020-03-11       7 -2.80 0.0933 -0.137 0.4      0.222 
#>  8 1      2020-03-05    2020-03-06       1 -5.34 0.0183 -1.96  0.4      0.0464
#>  9 1      2020-03-05    2020-03-07       2 -4.12 0.0360 -1.03  0.200    0.121 
#> 10 1      2020-03-05    2020-03-08       3 -3.65 0.0393 -0.722 0        0.169 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1     69  68.2 10.2      55    55    70    84
#>  2 2020-03-04    2020-03-06       2     85  83.7 10.3      67    82    89   100
#>  3 2020-03-04    2020-03-07       3    100 102.  14.4      83    87   106   128
#>  4 2020-03-04    2020-03-08       4    121 123.  13.9     102   116   129   147
#>  5 2020-03-04    2020-03-09       5    138 142.  23.3     114   127   146   198
#>  6 2020-03-04    2020-03-10       6    174 178.  28.5     143   156   184   230
#>  7 2020-03-04    2020-03-11       7    194 204.  38.5     154   164   202   283
#>  8 2020-03-05    2020-03-06       1     81  80    9.44     68    77    85   100
#>  9 2020-03-05    2020-03-07       2     99 101.  10.4      86    86    99   116
#> 10 2020-03-05    2020-03-08       3    117 116.  14.0     102   114   125   149
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.89  4.04  3.60 0.200     15.6 
#>  2 1      2020-03-04    2020-03-06       2  6.71 10.4   6.76 0.8        4.45
#>  3 1      2020-03-04    2020-03-07       3  6.48  9.63  4.27 0.4       16.3 
#>  4 1      2020-03-04    2020-03-08       4  8.53 17.4   4.77 1          9.64
#>  5 1      2020-03-04    2020-03-09       5  8.09 19.9   4.76 0.8       23.7 
#>  6 1      2020-03-04    2020-03-10       6  8.74 29.3   5.20 0.8       40.8 
#>  7 1      2020-03-04    2020-03-11       7  8.81 27.8   5.19 0.6       51.9 
#>  8 1      2020-03-05    2020-03-06       1  5.53  5.69  3.96 0.8        5.19
#>  9 1      2020-03-05    2020-03-07       2  6.68  8.77  3.94 0.8       10.4 
#> 10 1      2020-03-05    2020-03-08       3  5.89  5.99  3.90 0.6       11.9 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>