Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.33 0.0827   2.26  2.26  2.30  2.54
#>  2 2020-03-04    2020-03-06       2   2.27  2.29 0.117    2.18  2.25  2.30  2.60
#>  3 2020-03-04    2020-03-07       3   2.23  2.26 0.145    2.14  2.20  2.26  2.65
#>  4 2020-03-04    2020-03-08       4   2.18  2.22 0.168    2.08  2.12  2.22  2.65
#>  5 2020-03-04    2020-03-09       5   2.20  2.20 0.173    2.01  2.12  2.27  2.62
#>  6 2020-03-04    2020-03-10       6   2.12  2.15 0.187    1.97  2.00  2.17  2.61
#>  7 2020-03-04    2020-03-11       7   2.12  2.13 0.198    1.92  2.01  2.20  2.60
#>  8 2020-03-05    2020-03-06       1   2.23  2.24 0.0407   2.19  2.19  2.23  2.32
#>  9 2020-03-05    2020-03-07       2   2.22  2.22 0.0633   2.09  2.20  2.27  2.31
#> 10 2020-03-05    2020-03-08       3   2.25  2.25 0.108    2.08  2.15  2.28  2.41
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -4.51 0.0258 -1.95  0.8      0.0385
#>  2 2020-03-04    2020-03-06       2 -3.56 0.0512 -1.28  0.8      0.0377
#>  3 2020-03-04    2020-03-07       3 -3.24 0.0558 -1.19  0.8      0.0508
#>  4 2020-03-04    2020-03-08       4 -3.12 0.0521 -1.20  0.6      0.0719
#>  5 2020-03-04    2020-03-09       5 -2.84 0.0744 -0.732 0.6      0.109 
#>  6 2020-03-04    2020-03-10       6 -2.96 0.0572 -0.846 0.4      0.155 
#>  7 2020-03-04    2020-03-11       7 -2.99 0.0596 -0.689 0.200    0.141 
#>  8 2020-03-05    2020-03-06       1 -4.83 0.0296 -1.77  0.8      0.0415
#>  9 2020-03-05    2020-03-07       2 -4.05 0.0540 -0.713 0.8      0.0643
#> 10 2020-03-05    2020-03-08       3 -2.54 0.0920 -0.610 0.8      0.148 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   68.5  72.9 11.3      56    56    69    89
#>  2 2020-03-04    2020-03-06       2   82    80.4 14.3      50    79    91   100
#>  3 2020-03-04    2020-03-07       3  104.  108.  16.2      91    91   106   139
#>  4 2020-03-04    2020-03-08       4  128.  130.  14.0     102   122   139   147
#>  5 2020-03-04    2020-03-09       5  155   154.  22.1     116   143   158   203
#>  6 2020-03-04    2020-03-10       6  190.  195.  34.9     152   181   215   275
#>  7 2020-03-04    2020-03-11       7  229   226.  40.2     163   213   244   314
#>  8 2020-03-05    2020-03-06       1   80.5  82.8  9.73     71    77    83   106
#>  9 2020-03-05    2020-03-07       2  104   102.  12.1      80    97   112   122
#> 10 2020-03-05    2020-03-08       3  125   128.  19.2     111   116   131   178
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.60  5.75  3.65   0.6     14.8 
#>  2 1      2020-03-04    2020-03-06       2  5.52  6.26  4.03   0.6     11.1 
#>  3 1      2020-03-04    2020-03-07       3  7.18 11.7   4.18   1       17.8 
#>  4 1      2020-03-04    2020-03-08       4  9.77 21.3   5.21   0.9     12.6 
#>  5 1      2020-03-04    2020-03-09       5  9.31 27.0   5.23   0.9     16.3 
#>  6 1      2020-03-04    2020-03-10       6  9.58 35.6   5.53   1       28.2 
#>  7 1      2020-03-04    2020-03-11       7  9.72 40.2   5.59   0.8     23.0 
#>  8 1      2020-03-05    2020-03-06       1  5.57  5.5   3.47   0.8      5.19
#>  9 1      2020-03-05    2020-03-07       2  6.43  9.43  4.17   0.8     11.1 
#> 10 1      2020-03-05    2020-03-08       3  7.89 17.7   5.26   1       11.9 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.27  2.26 0.0491   2.16  2.26  2.29  2.33
#>  2 2020-03-04    2020-03-06       2   2.20  2.21 0.0697   2.08  2.16  2.21  2.31
#>  3 2020-03-04    2020-03-07       3   2.14  2.14 0.105    1.91  2.09  2.19  2.25
#>  4 2020-03-04    2020-03-08       4   2.06  2.08 0.119    1.90  1.96  2.10  2.21
#>  5 2020-03-04    2020-03-09       5   2.01  2.01 0.134    1.75  1.88  2.04  2.18
#>  6 2020-03-04    2020-03-10       6   1.96  1.95 0.126    1.75  1.86  2.02  2.13
#>  7 2020-03-04    2020-03-11       7   1.91  1.91 0.131    1.69  1.88  2.01  2.11
#>  8 2020-03-05    2020-03-06       1   2.23  2.23 0.0694   2.17  2.19  2.25  2.37
#>  9 2020-03-05    2020-03-07       2   2.14  2.15 0.0772   2.09  2.09  2.16  2.30
#> 10 2020-03-05    2020-03-08       3   2.10  2.11 0.104    1.98  2.02  2.13  2.26
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps    logs   bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>   <dbl>  <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -5.47 0.0137 -1.93   -0.400    0.0511
#>  2 1      2020-03-04    2020-03-06       2 -4.94 0.0157 -2.05   -0.200    0.0496
#>  3 1      2020-03-04    2020-03-07       3 -4.01 0.0401 -1.07   -0.6      0.0535
#>  4 1      2020-03-04    2020-03-08       4 -3.69 0.0667  0.250  -0.6      0.0937
#>  5 1      2020-03-04    2020-03-09       5 -3.29 0.0907  0.272  -0.6      0.0672
#>  6 1      2020-03-04    2020-03-10       6 -3.03 0.104  -0.0670 -0.8      0.108 
#>  7 1      2020-03-04    2020-03-11       7 -2.78 0.120   0.167  -0.8      0.121 
#>  8 1      2020-03-05    2020-03-06       1 -5.24 0.0176 -1.88    0.200    0.0399
#>  9 1      2020-03-05    2020-03-07       2 -5.12 0.0179 -1.62    0        0.0686
#> 10 1      2020-03-05    2020-03-08       3 -4.35 0.0270 -1.30   -0.400    0.0804
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   69.5  69.0  6.47     57    62    71    79
#>  2 2020-03-04    2020-03-06       2   85.5  82.8  9.92     61    84    90    93
#>  3 2020-03-04    2020-03-07       3   97.5  97.4 10.3      77    93   104   112
#>  4 2020-03-04    2020-03-08       4  114   116.  13.1      96   108   125   141
#>  5 2020-03-04    2020-03-09       5  138.  139.  17.2     116   134   159   165
#>  6 2020-03-04    2020-03-10       6  155   157.  17.6     129   149   174   177
#>  7 2020-03-04    2020-03-11       7  176.  180.  22.2     148   157   180   213
#>  8 2020-03-05    2020-03-06       1   82    81.8  8.60     71    75    84   100
#>  9 2020-03-05    2020-03-07       2   96    95.8 10.3      79    87   102   110
#> 10 2020-03-05    2020-03-08       3  118.  116.  14.8      97   108   123   145
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.11  3.3   3.37 0.5        6.67
#>  2 1      2020-03-04    2020-03-06       2  5.73  8.33  4.71 0.8        5.93
#>  3 1      2020-03-04    2020-03-07       3  5.52  5.26  3.79 0.6        9.64
#>  4 1      2020-03-04    2020-03-08       4  5.84  7.23  4.00 0.6        8.90
#>  5 1      2020-03-04    2020-03-09       5  7.15 12.9   4.44 0.7       17.8 
#>  6 1      2020-03-04    2020-03-10       6  5.79  5.16  3.76 0.200     18.5 
#>  7 1      2020-03-04    2020-03-11       7  6.40  4.94  3.99 0         17.0 
#>  8 1      2020-03-05    2020-03-06       1  4.66  4.12  3.43 0.8        5.93
#>  9 1      2020-03-05    2020-03-07       2  4.82  3.09  3.23 0.7        5.93
#> 10 1      2020-03-05    2020-03-08       3  5.82  8.25  4.34 0.6       10.4 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>