Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0285   2.24  2.29  2.32  2.32
#>  2 2020-03-04    2020-03-06       2   2.21  2.21 0.0578   2.11  2.17  2.25  2.30
#>  3 2020-03-04    2020-03-07       3   2.16  2.17 0.0659   2.06  2.12  2.19  2.29
#>  4 2020-03-04    2020-03-08       4   2.06  2.08 0.0503   2.01  2.01  2.07  2.18
#>  5 2020-03-04    2020-03-09       5   2.05  2.05 0.0636   1.96  2.02  2.11  2.14
#>  6 2020-03-04    2020-03-10       6   1.99  1.99 0.0842   1.83  1.98  2.09  2.09
#>  7 2020-03-04    2020-03-11       7   1.96  1.95 0.0997   1.77  1.96  2.09  2.09
#>  8 2020-03-05    2020-03-06       1   2.20  2.20 0.0164   2.17  2.19  2.21  2.22
#>  9 2020-03-05    2020-03-07       2   2.14  2.13 0.0462   2.03  2.12  2.15  2.20
#> 10 2020-03-05    2020-03-08       3   2.06  2.06 0.0466   1.96  2.06  2.11  2.11
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss    crps  logs   bias sharpness
#>    <chr>         <date>       <int> <dbl>   <dbl> <dbl>  <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -6.74 0.0148  -2.07  0.4      0.0323
#>  2 2020-03-04    2020-03-06       2 -5.72 0.0160  -1.70  0.200    0.0588
#>  3 2020-03-04    2020-03-07       3 -5.41 0.0167  -1.81  0.200    0.0536
#>  4 2020-03-04    2020-03-08       4 -5.80 0.0195  -1.71 -0.200    0.0562
#>  5 2020-03-04    2020-03-09       5 -5.58 0.0192  -1.52 -0.200    0.0942
#>  6 2020-03-04    2020-03-10       6 -4.83 0.0285  -1.31 -0.200    0.0991
#>  7 2020-03-04    2020-03-11       7 -4.28 0.0389  -1.13 -0.400    0.122 
#>  8 2020-03-05    2020-03-06       1 -8.29 0.00491 -2.97  0.200    0.0122
#>  9 2020-03-05    2020-03-07       2 -6.10 0.00862 -2.36 -0.200    0.0251
#> 10 2020-03-05    2020-03-08       3 -5.14 0.0239  -1.85 -0.6      0.0507
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   68    68.8  2.70     66    68    69    76
#>  2 2020-03-04    2020-03-06       2   84    85.6  5.62     78    78    84    96
#>  3 2020-03-04    2020-03-07       3   99.5  96.1  9.83     76    96   104   107
#>  4 2020-03-04    2020-03-08       4  116   116.  10.9     100   110   119   138
#>  5 2020-03-04    2020-03-09       5  128.  130.  13.5     113   113   129   153
#>  6 2020-03-04    2020-03-10       6  158   159.  19.5     131   153   165   201
#>  7 2020-03-04    2020-03-11       7  198   196.  21.1     163   185   207   239
#>  8 2020-03-05    2020-03-06       1   80.5  81.7 10.2      68    68    81   102
#>  9 2020-03-05    2020-03-07       2   90    92.5  8.91     84    84    90   109
#> 10 2020-03-05    2020-03-08       3  116.  113.   9.63     96   116   125   125
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  7.01  4.68 20.6  1          1.48
#>  2 1      2020-03-04    2020-03-06       2  8.93  9.66  5.52 1          5.19
#>  3 1      2020-03-04    2020-03-07       3  5.22  6.53  4.25 0.6        5.93
#>  4 1      2020-03-04    2020-03-08       4  6.48  8.73  3.86 0.7        6.67
#>  5 1      2020-03-04    2020-03-09       5  6.31  8.03  3.98 0.6       17.0 
#>  6 1      2020-03-04    2020-03-10       6  6.76 11.2   5.04 0.6        8.90
#>  7 1      2020-03-04    2020-03-11       7  8.11 19.0   4.77 0.8       16.3 
#>  8 1      2020-03-05    2020-03-06       1  5.35  4.51  3.52 0.6       10.4 
#>  9 1      2020-03-05    2020-03-07       2  4.55  2.67  3.23 0.200      8.15
#> 10 1      2020-03-05    2020-03-08       3  5.85  7.01  3.85 0.8       10.4 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.33  2.34 0.0716   2.19  2.29  2.37  2.45
#>  2 2020-03-04    2020-03-06       2   2.27  2.27 0.0912   2.10  2.20  2.29  2.40
#>  3 2020-03-04    2020-03-07       3   2.27  2.27 0.102    1.98  2.26  2.32  2.33
#>  4 2020-03-04    2020-03-08       4   2.23  2.22 0.151    2.05  2.21  2.30  2.52
#>  5 2020-03-04    2020-03-09       5   2.23  2.19 0.176    1.93  2.18  2.28  2.57
#>  6 2020-03-04    2020-03-10       6   2.18  2.15 0.195    1.88  2.13  2.27  2.59
#>  7 2020-03-04    2020-03-11       7   2.13  2.09 0.232    1.72  2.11  2.26  2.58
#>  8 2020-03-05    2020-03-06       1   2.23  2.22 0.0245   2.18  2.21  2.24  2.26
#>  9 2020-03-05    2020-03-07       2   2.16  2.17 0.0294   2.14  2.15  2.17  2.24
#> 10 2020-03-05    2020-03-08       3   2.10  2.12 0.0373   2.08  2.09  2.11  2.19
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon    dss   crps   logs   bias sharpness
#>    <chr>  <chr>         <date>       <int>  <dbl>  <dbl>  <dbl>  <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -3.16  0.0623 -1.03   1        0.0507
#>  2 1      2020-03-04    2020-03-06       2 -3.81  0.0637 -0.328  0.8      0.0482
#>  3 1      2020-03-04    2020-03-07       3 10.5   0.114   6.86   1        0.0350
#>  4 1      2020-03-04    2020-03-08       4 -2.59  0.0950 -0.235  0.8      0.0519
#>  5 1      2020-03-04    2020-03-09       5  0.363 0.140  -0.169  1        0.0583
#>  6 1      2020-03-04    2020-03-10       6 -0.662 0.120  -0.268  1        0.0649
#>  7 1      2020-03-04    2020-03-11       7 -0.573 0.107  -0.167  1        0.0842
#>  8 1      2020-03-05    2020-03-06       1 -6.06  0.0169 -2.24   0.8      0.0206
#>  9 1      2020-03-05    2020-03-07       2 -6.62  0.0106 -2.59   0.6      0.0160
#> 10 1      2020-03-05    2020-03-08       3 -6.26  0.0121 -2.15  -0.200    0.0261
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   68    69.7  8.42     60    60    71    87
#>  2 2020-03-04    2020-03-06       2   88    86.2  9.25     68    82    91    97
#>  3 2020-03-04    2020-03-07       3  100   102.  16.9      71    91   106   126
#>  4 2020-03-04    2020-03-08       4  128   128.  14.3     105   124   135   153
#>  5 2020-03-04    2020-03-09       5  150.  149.  23.1      95   136   155   171
#>  6 2020-03-04    2020-03-10       6  180.  182.  32.9     109   161   197   224
#>  7 2020-03-04    2020-03-11       7  228.  216.  41.4     105   222   243   246
#>  8 2020-03-05    2020-03-06       1   82    83    8.23     75    75    82    98
#>  9 2020-03-05    2020-03-07       2   96.5  96.0 11.5      81    85   100   115
#> 10 2020-03-05    2020-03-08       3  114   116.  12.6     101    97   116   138
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  4.88  3.24  3.22   0.6      7.41
#>  2 1      2020-03-04    2020-03-06       2  7.90 11.3   4.77   0.8      6.67
#>  3 1      2020-03-04    2020-03-07       3  5.69  6.82  3.98   0.8      9.64
#>  4 1      2020-03-04    2020-03-08       4 11.4  19.2   5.84   1        9.64
#>  5 1      2020-03-04    2020-03-09       5 13.9  30.2   7.41   1       17.0 
#>  6 1      2020-03-04    2020-03-10       6 10.1  29.8   5.32   1       28.9 
#>  7 1      2020-03-04    2020-03-11       7 17.5  47.0   8.50   1        7.41
#>  8 1      2020-03-05    2020-03-06       1  5.37  6.33  4.27   0.8      5.93
#>  9 1      2020-03-05    2020-03-07       2  5.53  5.49  3.74   0.4     17.0 
#> 10 1      2020-03-05    2020-03-08       3  6.97 10.4   4.06   0.8     12.6 
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>