Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.32  2.32 0.0317   2.27  2.30  2.33  2.38
#>  2 2020-03-04    2020-03-06       2   2.25  2.27 0.0411   2.23  2.23  2.25  2.34
#>  3 2020-03-04    2020-03-07       3   2.18  2.21 0.0578   2.15  2.15  2.18  2.33
#>  4 2020-03-04    2020-03-08       4   2.16  2.18 0.0647   2.12  2.14  2.16  2.34
#>  5 2020-03-04    2020-03-09       5   2.10  2.12 0.0768   2.03  2.06  2.11  2.30
#>  6 2020-03-04    2020-03-10       6   2.08  2.09 0.0865   1.98  2.06  2.13  2.29
#>  7 2020-03-04    2020-03-11       7   2.04  2.06 0.111    1.94  1.97  2.06  2.28
#>  8 2020-03-05    2020-03-06       1   2.20  2.20 0.0485   2.13  2.18  2.24  2.29
#>  9 2020-03-05    2020-03-07       2   2.17  2.16 0.0443   2.07  2.15  2.19  2.22
#> 10 2020-03-05    2020-03-08       3   2.11  2.09 0.0571   1.97  2.10  2.13  2.15
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 9
#>    forecast_date date       horizon   dss   crps   logs  bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -4.21 0.0338 -1.44  1        0.0269
#>  2 2020-03-04    2020-03-06       2 -2.38 0.0584 -0.557 1        0.0233
#>  3 2020-03-04    2020-03-07       3 -4.55 0.0325 -1.69  1        0.0405
#>  4 2020-03-04    2020-03-08       4 -4.08 0.0448 -1.38  1        0.0250
#>  5 2020-03-04    2020-03-09       5 -4.60 0.0276 -1.74  0.6      0.0460
#>  6 2020-03-04    2020-03-10       6 -4.43 0.0329 -1.45  0.6      0.0613
#>  7 2020-03-04    2020-03-11       7 -4.31 0.0274 -1.20  0        0.103 
#>  8 2020-03-05    2020-03-06       1 -6.08 0.0131 -1.89  0.200    0.0487
#>  9 2020-03-05    2020-03-07       2 -6.26 0.0151 -1.89  0.4      0.0312
#> 10 2020-03-05    2020-03-08       3 -5.74 0.0114 -2.09  0        0.0247
#> # … with 95 more rows, and 1 more variable: calibration <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   75    73.5  8.75     51    71    78    82
#>  2 2020-03-04    2020-03-06       2   84    84.4  8.30     75    82    87   105
#>  3 2020-03-04    2020-03-07       3   97.5  99.5  9.05     87    87   100   116
#>  4 2020-03-04    2020-03-08       4  130   128.  10.3     106   126   136   140
#>  5 2020-03-04    2020-03-09       5  146   144.  11.6     122   144   154   158
#>  6 2020-03-04    2020-03-10       6  176.  176.  12.4     157   173   187   195
#>  7 2020-03-04    2020-03-11       7  202.  203.  12.6     181   195   212   224
#>  8 2020-03-05    2020-03-06       1   83    83.7  9.98     61    79    89    95
#>  9 2020-03-05    2020-03-07       2  100    99.1  8.61     87    92   106   111
#> 10 2020-03-05    2020-03-08       3  110   111.  11.3      93   105   111   134
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 10
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.83  8.93  6.53   0.8      5.19
#>  2 1      2020-03-04    2020-03-06       2  6.22  7.52  3.92   1        3.71
#>  3 1      2020-03-04    2020-03-07       3  6.09  6.89  3.66   0.8      9.64
#>  4 1      2020-03-04    2020-03-08       4 11.8  21.0   5.18   1        8.90
#>  5 1      2020-03-04    2020-03-09       5 11.2  21.7   5.26   1       11.1 
#>  6 1      2020-03-04    2020-03-10       6 13.6  28.0   6.75   1       12.6 
#>  7 1      2020-03-04    2020-03-11       7 13.8  28.9   6.79   1       12.6 
#>  8 1      2020-03-05    2020-03-06       1  5.77  8.11  4.26   0.8      7.41
#>  9 1      2020-03-05    2020-03-07       2  6.05  6.71  3.61   0.8     11.9 
#> 10 1      2020-03-05    2020-03-08       3  5.50  5.41  3.67   0.8      6.67
#> # … with 95 more rows, and 1 more variable: calibration <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.30 0.0387   2.26  2.29  2.31  2.34
#>  2 2020-03-04    2020-03-06       2   2.24  2.25 0.0314   2.20  2.23  2.26  2.31
#>  3 2020-03-04    2020-03-07       3   2.20  2.20 0.0370   2.15  2.17  2.22  2.25
#>  4 2020-03-04    2020-03-08       4   2.15  2.16 0.0527   2.07  2.11  2.15  2.20
#>  5 2020-03-04    2020-03-09       5   2.10  2.11 0.0556   2.01  2.08  2.12  2.21
#>  6 2020-03-04    2020-03-10       6   2.05  2.05 0.0521   1.98  2.04  2.07  2.14
#>  7 2020-03-04    2020-03-11       7   2.00  2.01 0.0692   1.88  1.98  2.03  2.11
#>  8 2020-03-05    2020-03-06       1   2.23  2.22 0.0392   2.14  2.21  2.25  2.27
#>  9 2020-03-05    2020-03-07       2   2.16  2.17 0.0584   2.02  2.15  2.19  2.24
#> 10 2020-03-05    2020-03-08       3   2.13  2.12 0.0746   1.98  2.09  2.16  2.19
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 10
#>    sample forecast_date date       horizon     dss   crps   logs  bias sharpness
#>    <chr>  <chr>         <date>       <int>   <dbl>  <dbl>  <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  3.99   0.0203  2.06    1     0.00875
#>  2 1      2020-03-04    2020-03-06       2  0.0774 0.0386 -0.835   1     0.0110 
#>  3 1      2020-03-04    2020-03-07       3 -2.89   0.0346 -1.27    1     0.0200 
#>  4 1      2020-03-04    2020-03-08       4 -5.59   0.0281  5.99    0.8   0.0112 
#>  5 1      2020-03-04    2020-03-09       5 -6.18   0.0214 -1.26    0.6   0.0283 
#>  6 1      2020-03-04    2020-03-10       6 -6.28   0.0153  1.78    0.6   0.0110 
#>  7 1      2020-03-04    2020-03-11       7 -5.86   0.0149 -2.36   -0.6   0.0176 
#>  8 1      2020-03-05    2020-03-06       1  8.61   0.0347 11.1     1     0.00702
#>  9 1      2020-03-05    2020-03-07       2 -5.00   0.0236 -2.07    1     0.0223 
#> 10 1      2020-03-05    2020-03-08       3 -5.09   0.0260 -1.81    0.8   0.0244 
#> # … with 200 more rows, and 1 more variable: calibration <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1    74   72.0  7.26     61    73    81    81
#>  2 2020-03-04    2020-03-06       2    88   88    8.60     76    84    92   104
#>  3 2020-03-04    2020-03-07       3   104. 102.  10.4      88    97   108   118
#>  4 2020-03-04    2020-03-08       4   124  124.  13.1     102   116   130   138
#>  5 2020-03-04    2020-03-09       5   142. 146.  15.7     116   139   160   166
#>  6 2020-03-04    2020-03-10       6   170  170.  13.1     158   167   174   203
#>  7 2020-03-04    2020-03-11       7   204. 204.  19.3     182   194   215   238
#>  8 2020-03-05    2020-03-06       1    82   84.8 11.0      74    75    86   109
#>  9 2020-03-05    2020-03-07       2    96   98.4 11.7      82    82    97   118
#> 10 2020-03-05    2020-03-08       3   120. 122.  15.8      98   105   122   148
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 10
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.33  4.91  3.32   0.8      8.90
#>  2 1      2020-03-04    2020-03-06       2 10.2  12.3   5.63   1        5.19
#>  3 1      2020-03-04    2020-03-07       3  8.58 11.0   4.65   1        7.41
#>  4 1      2020-03-04    2020-03-08       4  8.94 14.6   4.67   0.9      8.90
#>  5 1      2020-03-04    2020-03-09       5  8.64 18.1   4.80   0.9     19.3 
#>  6 1      2020-03-04    2020-03-10       6 29.6  25.6  58.1    1        2.97
#>  7 1      2020-03-04    2020-03-11       7 19.2  28.8  10.8    1        8.90
#>  8 1      2020-03-05    2020-03-06       1  5.86  7.61  4.94   0.8      5.19
#>  9 1      2020-03-05    2020-03-07       2  6.13  8.11  3.97   0.6     14.1 
#> 10 1      2020-03-05    2020-03-08       3  8.00 15.6   4.86   1        8.90
#> # … with 200 more rows, and 1 more variable: calibration <dbl>
#>