Evaluate a Model for Forecasting Rts

Evaluate a Model for Forecasting Rts

evaluate_model(
  obs_rts = NULL,
  obs_cases = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3,
  serial_interval = NULL,
  rdist = NULL,
  return_raw = FALSE
)

Arguments

obs_rts
Dataframe of Rt observations to forecast with and score against. Should contain a date and rt column. If multiple samples are included this should be denoted using a numeric sample variable.

obs_cases
Dataframe of case observations to use for case prediction and scoring. Should contain a date and cases column. If multiple samples are included this should be denoted using a numeric sample variable.

model
A model object in the format of bsts_model or fable_model. See the corresponding help files for details.

horizon
Numeric, the time horizon over which to predict.

samples
Numeric, number of samples to take.

timeout
Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.

bound_rt
Logical, defaults to TRUE. Should Rt values be bounded to be greater than or equal to 0.

min_points
Numeric, defaults to 3. The minimum number of time points at which to begin iteratively evaluating the forecast.

serial_interval
A numeric vector describing the probability distribution the serial interval. See EpiNow::covid_serial_interval for an example of the format.

rdist
A function to be used to sample the number of cases. Must take two arguments with the first specfying the number of samples and the second the mean. Defaults to rpois if not supplied

return_raw
Logical, should raw cases and rt forecasts be returned. Defaults to FALSE.

Value

Examples

## Evaluate a model based on a single sample of input cases
evaluate_model(EpiSoon::example_obs_rts,
               EpiSoon::example_obs_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.30  2.28 0.0544   2.19  2.30  2.35  2.35
#>  2 2020-03-04    2020-03-06       2   2.27  2.25 0.0619   2.13  2.24  2.31  2.33
#>  3 2020-03-04    2020-03-07       3   2.19  2.19 0.0730   2.06  2.17  2.25  2.31
#>  4 2020-03-04    2020-03-08       4   2.18  2.16 0.0654   2.07  2.10  2.19  2.26
#>  5 2020-03-04    2020-03-09       5   2.11  2.10 0.0860   1.95  2.09  2.17  2.23
#>  6 2020-03-04    2020-03-10       6   2.05  2.04 0.104    1.92  2.03  2.14  2.23
#>  7 2020-03-04    2020-03-11       7   2.01  1.99 0.0927   1.83  1.92  2.03  2.12
#>  8 2020-03-05    2020-03-06       1   2.24  2.24 0.0400   2.19  2.19  2.25  2.29
#>  9 2020-03-05    2020-03-07       2   2.19  2.20 0.0656   2.11  2.14  2.21  2.30
#> 10 2020-03-05    2020-03-08       3   2.17  2.15 0.0754   2.03  2.14  2.25  2.25
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 105 x 12
#>    forecast_date date       horizon   dss   crps  logs   bias sharpness
#>    <chr>         <date>       <int> <dbl>  <dbl> <dbl>  <dbl>     <dbl>
#>  1 2020-03-04    2020-03-05       1 -5.89 0.0217 -1.53  0.200    0.0418
#>  2 2020-03-04    2020-03-06       2 -4.59 0.0431 -1.07  0.6      0.0566
#>  3 2020-03-04    2020-03-07       3 -4.98 0.0300 -1.33  0.4      0.0732
#>  4 2020-03-04    2020-03-08       4 -4.74 0.0350 -1.34  0.4      0.0752
#>  5 2020-03-04    2020-03-09       5 -4.72 0.0353 -1.15  0.4      0.101 
#>  6 2020-03-04    2020-03-10       6 -4.60 0.0270 -1.11  0.4      0.132 
#>  7 2020-03-04    2020-03-11       7 -4.79 0.0175 -1.53 -0.200    0.0820
#>  8 2020-03-05    2020-03-06       1 -5.15 0.0252 -1.83  0.6      0.0564
#>  9 2020-03-05    2020-03-07       2 -4.85 0.0263 -1.57  0.6      0.0653
#> 10 2020-03-05    2020-03-08       3 -4.92 0.0323 -1.24  0.200    0.0893
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   70.5  70.7  9.13     55    65    77    84
#>  2 2020-03-04    2020-03-06       2   85.5  82   12.0      56    80    88    96
#>  3 2020-03-04    2020-03-07       3   94    95    7.72     83    93   104   104
#>  4 2020-03-04    2020-03-08       4  120   121.  16.1     101   112   136   148
#>  5 2020-03-04    2020-03-09       5  144.  144.  19.2     111   135   158   176
#>  6 2020-03-04    2020-03-10       6  164   168.  17.5     145   145   164   196
#>  7 2020-03-04    2020-03-11       7  208   206.  21.7     176   201   231   235
#>  8 2020-03-05    2020-03-06       1   86    83    7.69     71    84    91    91
#>  9 2020-03-05    2020-03-07       2  100.   97.1  7.95     83   100   106   106
#> 10 2020-03-05    2020-03-08       3  123   120.  12.9     101   120   139   139
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 105 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.11  4.77  3.56   0.6      8.90
#>  2 1      2020-03-04    2020-03-06       2  5.49  7.4   4.71   0.6      8.15
#>  3 1      2020-03-04    2020-03-07       3  4.90  4.26  3.39   0.6     11.1 
#>  4 1      2020-03-04    2020-03-08       4  7.07 11.    4.17   0.7     19.3 
#>  5 1      2020-03-04    2020-03-09       5  8.18 18.9   4.81   0.8     17.0 
#>  6 1      2020-03-04    2020-03-10       6  8.21 17.3   4.61   1       23.0 
#>  7 1      2020-03-04    2020-03-11       7  9.72 27.7   5.11   1       28.2 
#>  8 1      2020-03-05    2020-03-06       1  5.85  6.4   3.56   0.8      6.67
#>  9 1      2020-03-05    2020-03-07       2  5.50  6.15  3.74   0.6      7.41
#> 10 1      2020-03-05    2020-03-08       3  7.14 11.2   4.15   0.8     16.3 
#> # … with 95 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 

## Samples of observed data
sampled_obs <- EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_rts %>%
   dplyr::mutate(sample = 2))

sampled_cases <- EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 1) %>%
   dplyr::bind_rows(EpiSoon::example_obs_cases %>%
   dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               sampled_cases,
               model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
               horizon = 7, samples = 10,
               serial_interval = EpiSoon::example_serial_interval)

#> $forecast_rts
#> # A tibble: 133 x 10
#>    forecast_date date       horizon median  mean     sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   2.29  2.29 0.0183   2.27  2.29  2.31  2.31
#>  2 2020-03-04    2020-03-06       2   2.23  2.23 0.0314   2.17  2.22  2.24  2.28
#>  3 2020-03-04    2020-03-07       3   2.18  2.18 0.0289   2.15  2.17  2.20  2.24
#>  4 2020-03-04    2020-03-08       4   2.13  2.13 0.0352   2.08  2.10  2.14  2.19
#>  5 2020-03-04    2020-03-09       5   2.08  2.09 0.0369   2.04  2.05  2.08  2.15
#>  6 2020-03-04    2020-03-10       6   2.04  2.04 0.0400   1.99  2.00  2.04  2.09
#>  7 2020-03-04    2020-03-11       7   1.98  1.99 0.0544   1.93  1.95  1.99  2.05
#>  8 2020-03-05    2020-03-06       1   2.23  2.22 0.0363   2.15  2.21  2.25  2.27
#>  9 2020-03-05    2020-03-07       2   2.16  2.18 0.0638   2.10  2.12  2.18  2.26
#> 10 2020-03-05    2020-03-08       3   2.14  2.15 0.0621   2.07  2.07  2.15  2.22
#> # … with 123 more rows
#> 
#> $rt_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss   crps   logs   bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl>  <dbl>  <dbl>  <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1 -6.41 0.0126 -2.59   1        0.0179
#>  2 1      2020-03-04    2020-03-06       2 -4.99 0.0324 -0.472  0.8      0.0242
#>  3 1      2020-03-04    2020-03-07       3 -5.40 0.0267 -1.63   0.6      0.0410
#>  4 1      2020-03-04    2020-03-08       4 -5.40 0.0210 -2.14   0.6      0.0250
#>  5 1      2020-03-04    2020-03-09       5 -5.25 0.0240 -1.86   0.6      0.0475
#>  6 1      2020-03-04    2020-03-10       6 -5.83 0.0175 -1.83   0.200    0.0588
#>  7 1      2020-03-04    2020-03-11       7 -5.48 0.0150 -1.89  -0.200    0.0549
#>  8 1      2020-03-05    2020-03-06       1 -5.49 0.0263 -1.68   0.8      0.0334
#>  9 1      2020-03-05    2020-03-07       2 -4.47 0.0361 -1.31   0.6      0.0565
#> 10 1      2020-03-05    2020-03-08       3 -4.53 0.0354 -1.34   0.4      0.0731
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#> 
#> $forecast_cases
#> # A tibble: 105 x 10
#>    forecast_date date       horizon median  mean    sd bottom lower upper   top
#>    <chr>         <date>       <int>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 2020-03-04    2020-03-05       1   70.5  70.1  7.29     58    66    73    83
#>  2 2020-03-04    2020-03-06       2   89    86.4  7.36     72    88    93    94
#>  3 2020-03-04    2020-03-07       3   98.5  99.4 14.5      80    85   103   128
#>  4 2020-03-04    2020-03-08       4  121   122.  12.3     100   115   128   143
#>  5 2020-03-04    2020-03-09       5  140   144   15.8     128   128   142   180
#>  6 2020-03-04    2020-03-10       6  170   169.  16.4     143   169   183   196
#>  7 2020-03-04    2020-03-11       7  188   192.  23.0     152   178   201   223
#>  8 2020-03-05    2020-03-06       1   82.5  80.8  9.66     66    79    91    95
#>  9 2020-03-05    2020-03-07       2   93.5  94.1 10.8      74    89    96   108
#> 10 2020-03-05    2020-03-08       3  111   112.  10.8      99   104   115   137
#> # … with 95 more rows
#> 
#> $case_scores
#> # A tibble: 210 x 13
#>    sample forecast_date date       horizon   dss  crps  logs  bias sharpness
#>    <chr>  <chr>         <date>       <int> <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 1      2020-03-04    2020-03-05       1  5.56  5.69  3.56   0.8      7.41
#>  2 1      2020-03-04    2020-03-06       2 10.7  12.6   4.26   1        1.48
#>  3 1      2020-03-04    2020-03-07       3  6.10  7.29  4.09   0.6     17.0 
#>  4 1      2020-03-04    2020-03-08       4  7.85 15.0   5.27   0.8      6.67
#>  5 1      2020-03-04    2020-03-09       5  9.80 22.1  11.9    1        5.93
#>  6 1      2020-03-04    2020-03-10       6  8.27 22.4   5.47   0.8      9.64
#>  7 1      2020-03-04    2020-03-11       7  7.47 13.0   4.32   0.7     18.5 
#>  8 1      2020-03-05    2020-03-06       1  4.75  4.7   3.75   0.4      9.64
#>  9 1      2020-03-05    2020-03-07       2  5.75  6.79  3.84   0.8      7.41
#> 10 1      2020-03-05    2020-03-08       3  5.52  4.68  3.53   0.6      9.64
#> # … with 200 more rows, and 4 more variables: calibration <dbl>, median <dbl>,
#> #   iqr <dbl>, ci <dbl>
#>