Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 18 x 9
#>    score  model       bottom    lower   median     mean   upper     top       sd
#>    <chr>  <chr>        <dbl>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1 bias   AR 3      -1.00e+0 -8.00e-1 -6.00e-1  -0.555  -0.400    0       0.268 
#>  2 bias   Semi-lo…  -1.00e+0 -8.00e-1 -2.00e-1  -0.123   0.6      1       0.736 
#>  3 calib… AR 3       8.57e-5  8.57e-5  5.85e-3   0.0287  0.0149   0.184   0.0699
#>  4 calib… Semi-lo…   8.57e-5  8.57e-5  9.29e-5   0.146   0.205    0.906   0.258 
#>  5 ci     AR 3       1.87e+0  2.94e+0  3.74e+0   6.02    5.35    29.4     8.15  
#>  6 ci     Semi-lo…   5.52e-2  1.77e-1  3.01e-1   0.997   1.30     4.55    1.32  
#>  7 crps   AR 3       4.79e-1  6.57e-1  8.43e-1   0.890   1.05     1.46    0.282 
#>  8 crps   Semi-lo…   7.13e-3  2.44e-2  4.49e-2   0.0587  0.0806   0.168   0.0438
#>  9 dss    AR 3       4.98e-1  9.94e-1  1.31e+0   3.66    2.04    15.2    15.9   
#> 10 dss    Semi-lo…  -7.62e+0 -5.17e+0 -3.81e+0  -1.84   -0.997   18.0     6.69  
#> 11 iqr    AR 3       1.77e+0  2.34e+0  3.24e+0   3.85    5.24     7.55    1.77  
#> 12 iqr    Semi-lo…   2.93e-2  9.40e-2  2.03e-1   0.259   0.392    0.729   0.196 
#> 13 logs   AR 3       1.46e+0  1.72e+0  1.99e+0 Inf       2.46   Inf     Inf     
#> 14 logs   Semi-lo…  -2.70e+0 -1.58e+0 -9.58e-1   1.07   -0.0372  24.5     9.68  
#> 15 median AR 3       5.57e-1  2.38e+0  3.16e+0   2.91    3.63     4.10    0.965 
#> 16 median Semi-lo…   9.48e-3  6.37e-2  1.28e-1   0.154   0.224    0.415   0.111 
#> 17 sharp… AR 3       0.       0.       2.70e-1   0.552   0.943    2.17    0.657 
#> 18 sharp… Semi-lo…   1.50e-2  2.89e-2  4.70e-2   0.0602  0.0720   0.191   0.0466

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model          bottom  lower median   mean  upper    top     sd
#>      <int> <chr> <chr>           <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3          0.491   0.627  0.820  0.851  1.05   1.27   0.253 
#>  2       1 crps  Semi-local l… 0.00519 0.0104 0.0220 0.0261 0.0320 0.0804 0.0235
#>  3       2 crps  AR 3          0.496   0.735  0.840  0.947  1.17   1.63   0.338 
#>  4       2 crps  Semi-local l… 0.00729 0.0182 0.0286 0.0404 0.0562 0.123  0.0340
#>  5       3 crps  AR 3          0.436   0.637  0.823  0.840  0.986  1.42   0.275 
#>  6       3 crps  Semi-local l… 0.0161  0.0226 0.0411 0.0574 0.0791 0.174  0.0446
#>  7       4 crps  AR 3          0.547   0.768  0.943  0.944  1.04   1.46   0.270 
#>  8       4 crps  Semi-local l… 0.0116  0.0308 0.0758 0.0709 0.0976 0.138  0.0399
#>  9       5 crps  AR 3          0.500   0.676  0.921  0.921  1.09   1.45   0.278 
#> 10       5 crps  Semi-local l… 0.0309  0.0535 0.0683 0.0788 0.102  0.153  0.0368
#> 11       6 crps  AR 3          0.512   0.624  0.797  0.838  1.01   1.33   0.253 
#> 12       6 crps  Semi-local l… 0.0170  0.0444 0.0604 0.0774 0.108  0.173  0.0481
#> 13       7 crps  AR 3          0.454   0.679  0.855  0.892  1.06   1.47   0.299 
#> 14       7 crps  Semi-local l… 0.0233  0.0425 0.0511 0.0760 0.111  0.178  0.0506

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model            bottom lower median   mean upper   top    sd
#>   <chr>      <chr> <chr>             <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 3               5.58  6.58   7.11 Inf     8.79 Inf   Inf  
#> 2 Region 1   logs  Semi-local line…   3.42  4.14   4.93   9.63  8.75  51.1  12.5
#> 3 Region 2   logs  AR 3               5.54  6.64   7.27 Inf     8.80 Inf   Inf  
#> 4 Region 2   logs  Semi-local line…   3.42  4.16   5.16  10.7   8.45  57.2  18.2