Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Dummy data
observations <- data.frame(rt = 1:20,
                            date = as.Date("2020-01-01")
                           + lubridate::days(1:20))

timeseries <- list(observations, observations)
names(timeseries) <- c("Region 1", "Region 2")

## List of forecasting bsts models wrapped in functions.
models <- list("Sparse AR" = function(ss, y){bsts::AddAutoAr(ss, y = y, lags = 7)},
               "Semi-local linear trend" = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)})

## Compare models
evaluations <- compare_timeseries(timeseries, models,
                                  horizon = 7, samples = 10)

scores <- evaluations$scores

## Score across the default groups
summarise_scores(scores)

#> # A tibble: 10 x 9
#>    score    model             bottom   lower median    mean upper    top      sd
#>    <chr>    <chr>              <dbl>   <dbl>  <dbl>   <dbl> <dbl>  <dbl>   <dbl>
#>  1 bias     Semi-local line…  0.200   0.3     0.45    0.45  0.5    0.8     0.158
#>  2 bias     Sparse AR         0       0       0.100   0.115 0.200  0.5     0.136
#>  3 crps     Semi-local line…  0.0469  0.103   0.163   0.215 0.249  0.797   0.179
#>  4 crps     Sparse AR         0.376   1.34    2.89    3.60  5.56   8.85    2.63 
#>  5 dss      Semi-local line… -3.34   -2.10   -1.07   -0.822 0.249  2.44    1.55 
#>  6 dss      Sparse AR         0.383   2.72    4.47   10.8   7.94  67.5    25.9  
#>  7 logs     Semi-local line… -0.659  -0.0862  0.463   0.497 0.952  1.99    0.739
#>  8 logs     Sparse AR         1.19    2.34    3.08  Inf     4.06  80.3   Inf    
#>  9 sharpne… Semi-local line…  0.129   0.262   0.436   0.646 0.803  2.32    0.572
#> 10 sharpne… Sparse AR         0       1.20    2.10    2.44  3.27   6.67    1.75 

## Also summarise across time horizon
summarise_scores(scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model            bottom  lower median  mean upper    top     sd
#>      <int> <chr> <chr>             <dbl>  <dbl>  <dbl> <dbl> <dbl>  <dbl>  <dbl>
#>  1       1 crps  Semi-local line… 0.0386 0.0749  0.110 0.125 0.142  0.345 0.0807
#>  2       1 crps  Sparse AR        0.239  0.497   0.772 1.04  1.50   2.74  0.770 
#>  3       2 crps  Semi-local line… 0.0511 0.0840  0.122 0.142 0.160  0.362 0.0908
#>  4       2 crps  Sparse AR        0.480  0.864   1.39  1.78  2.40   4.37  1.21  
#>  5       3 crps  Semi-local line… 0.0493 0.102   0.168 0.200 0.232  0.535 0.162 
#>  6       3 crps  Sparse AR        0.803  1.52    2.33  2.69  3.84   5.59  1.49  
#>  7       4 crps  Semi-local line… 0.0575 0.125   0.168 0.234 0.251  0.756 0.187 
#>  8       4 crps  Sparse AR        1.10   2.23    3.76  3.60  4.68   6.11  1.54  
#>  9       5 crps  Semi-local line… 0.0578 0.143   0.184 0.256 0.281  0.736 0.200 
#> 10       5 crps  Sparse AR        1.70   2.99    4.81  4.74  6.12   8.03  1.99  
#> 11       6 crps  Semi-local line… 0.0599 0.157   0.228 0.287 0.314  0.776 0.211 
#> 12       6 crps  Sparse AR        2.15   4.29    6.58  6.06  7.65   8.98  2.19  
#> 13       7 crps  Semi-local line… 0.110  0.171   0.253 0.330 0.351  0.825 0.234 
#> 14       7 crps  Sparse AR        3.01   6.37    7.93  7.43  8.28  10.7   2.09  

## Instead summarise across region
summarise_scores(scores, "region", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   region  score model         bottom   lower median    mean upper    top      sd
#>   <chr>   <chr> <chr>          <dbl>   <dbl>  <dbl>   <dbl> <dbl>  <dbl>   <dbl>
#> 1 Region… logs  Semi-local l… -0.569 -0.0949  0.506   0.540 1.15    2.06   0.761
#> 2 Region… logs  Sparse AR      1.33   2.21    2.90  Inf     3.75  151.   Inf    
#> 3 Region… logs  Semi-local l… -0.680 -0.0664  0.424   0.455 0.875   1.93   0.717
#> 4 Region… logs  Sparse AR      1.19   2.46    3.19    6.89  4.40   46.2   15.4