Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 4" =
                    function(ss, y){bsts::AddAutoAr(ss, y = y, lags = 4)},
               "Semi-local linear trend" =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 10 x 9
#>    score   model            bottom   lower  median    mean   upper    top     sd
#>    <chr>   <chr>             <dbl>   <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>
#>  1 bias    AR 4            0        0.200   0.4     0.45    0.7     1     0.315 
#>  2 bias    Semi-local li…  0        0.100   0.35    0.42    0.7     1     0.357 
#>  3 crps    AR 4            0.0118   0.0295  0.0498  0.0569  0.0780  0.132 0.0340
#>  4 crps    Semi-local li…  0.00926  0.0211  0.0471  0.0610  0.0886  0.171 0.0474
#>  5 dss     AR 4           -6.03    -4.55   -3.71   -3.72   -2.89   -0.902 1.40  
#>  6 dss     Semi-local li… -7.04    -5.28   -3.81   -2.13   -0.682  13.6   5.51  
#>  7 logs    AR 4           -2.04    -1.32   -0.879  -0.808  -0.468   1.27  0.913 
#>  8 logs    Semi-local li… -2.49    -1.66   -0.856   0.813   0.155  13.8   9.84  
#>  9 sharpn… AR 4            0.0280   0.0569  0.0807  0.0908  0.111   0.205 0.0463
#> 10 sharpn… Semi-local li…  0.0136   0.0317  0.0490  0.0605  0.0780  0.180 0.0411

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model          bottom  lower median   mean  upper    top     sd
#>      <int> <chr> <chr>           <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 4          0.00904 0.0154 0.0203 0.0279 0.0304 0.0799 0.0202
#>  2       1 crps  Semi-local l… 0.00575 0.0121 0.0171 0.0280 0.0365 0.0918 0.0252
#>  3       2 crps  AR 4          0.0128  0.0218 0.0342 0.0393 0.0475 0.0947 0.0236
#>  4       2 crps  Semi-local l… 0.00963 0.0156 0.0348 0.0438 0.0693 0.105  0.0321
#>  5       3 crps  AR 4          0.0208  0.0369 0.0520 0.0532 0.0635 0.0951 0.0216
#>  6       3 crps  Semi-local l… 0.0124  0.0220 0.0460 0.0638 0.0953 0.180  0.0497
#>  7       4 crps  AR 4          0.0259  0.0453 0.0706 0.0679 0.0895 0.104  0.0256
#>  8       4 crps  Semi-local l… 0.0115  0.0223 0.0788 0.0761 0.115  0.190  0.0565
#>  9       5 crps  AR 4          0.0166  0.0448 0.0597 0.0686 0.0970 0.128  0.0342
#> 10       5 crps  Semi-local l… 0.0194  0.0428 0.0713 0.0798 0.116  0.159  0.0471
#> 11       6 crps  AR 4          0.0263  0.0520 0.0680 0.0790 0.110  0.142  0.0365
#> 12       6 crps  Semi-local l… 0.0170  0.0272 0.0698 0.0729 0.108  0.153  0.0459
#> 13       7 crps  AR 4          0.0350  0.0444 0.0676 0.0793 0.0936 0.160  0.0406
#> 14       7 crps  Semi-local l… 0.0269  0.0443 0.0628 0.0772 0.0885 0.194  0.0494

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model             bottom lower median  mean upper   top    sd
#>   <chr>      <chr> <chr>              <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 4                3.48  4.51   5.07  7.83  6.22  37.6 10.6 
#> 2 Region 1   logs  Semi-local linea…   3.35  4.43   5.12 13.1   6.58  76.3 42.4 
#> 3 Region 2   logs  AR 4                3.29  4.30   4.82  6.68  6.20  35.5  6.98
#> 4 Region 2   logs  Semi-local linea…   3.32  4.31   5.10 10.6   7.64  51.0 20.7