Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 18 x 9
#>    score  model      bottom    lower   median     mean   upper      top       sd
#>    <chr>  <chr>       <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>    <dbl>
#>  1 bias   AR 3     -1.00e+0 -8.00e-1 -6.00e-1  -0.578  -0.400   -0.0450   0.236 
#>  2 bias   Semi-lo… -1.00e+0 -8.00e-1  0.       -0.0752  0.6      1        0.728 
#>  3 calib… AR 3      8.57e-5  8.57e-5  1.99e-3   0.0302  0.0151   0.241    0.0741
#>  4 calib… Semi-lo…  8.57e-5  8.57e-5  9.29e-5   0.0783  0.0108   0.812    0.199 
#>  5 ci     AR 3      1.98e+0  2.87e+0  3.69e+0   5.92    4.91    28.6      8.18  
#>  6 ci     Semi-lo…  7.07e-2  2.11e-1  3.74e-1   0.901   0.931    4.86     1.24  
#>  7 crps   AR 3      4.71e-1  6.93e-1  8.91e-1   0.919   1.11     1.61     0.303 
#>  8 crps   Semi-lo…  8.64e-3  2.43e-2  5.06e-2   0.0602  0.0839   0.172    0.0443
#>  9 dss    AR 3      5.38e-1  9.51e-1  1.30e+0   3.15    1.87    15.4      9.40  
#> 10 dss    Semi-lo… -7.30e+0 -4.83e+0 -3.62e+0  -2.77   -2.09     7.74     3.88  
#> 11 iqr    AR 3      1.73e+0  2.47e+0  3.40e+0   3.97    5.42     8.11     1.87  
#> 12 iqr    Semi-lo…  3.69e-2  1.15e-1  2.10e-1   0.266   0.367    0.791    0.203 
#> 13 logs   AR 3      1.38e+0  1.75e+0  2.07e+0 Inf       3.11   Inf      Inf     
#> 14 logs   Semi-lo… -2.53e+0 -1.57e+0 -8.90e-1  -0.0301 -0.387   13.8      4.44  
#> 15 median AR 3      9.87e-1  2.52e+0  3.25e+0   3.00    3.66     4.21     0.923 
#> 16 median Semi-lo…  4.62e-3  5.67e-2  1.42e-1   0.160   0.246    0.430    0.122 
#> 17 sharp… AR 3      0.       0.       2.71e-1   0.506   0.846    2.01     0.606 
#> 18 sharp… Semi-lo…  1.70e-2  3.56e-2  5.72e-2   0.0687  0.0872   0.202    0.0448

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model           bottom  lower median   mean  upper   top     sd
#>      <int> <chr> <chr>            <dbl>  <dbl>  <dbl>  <dbl>  <dbl> <dbl>  <dbl>
#>  1       1 crps  AR 3           0.437   0.697  0.926  0.961  1.11   1.80  0.361 
#>  2       1 crps  Semi-local li… 0.00479 0.0107 0.0192 0.0290 0.0324 0.123 0.0299
#>  3       2 crps  AR 3           0.528   0.717  0.922  0.964  1.17   1.65  0.323 
#>  4       2 crps  Semi-local li… 0.0119  0.0186 0.0321 0.0402 0.0516 0.116 0.0291
#>  5       3 crps  AR 3           0.519   0.695  0.876  0.891  1.07   1.35  0.258 
#>  6       3 crps  Semi-local li… 0.0134  0.0295 0.0478 0.0574 0.0777 0.135 0.0384
#>  7       4 crps  AR 3           0.538   0.736  0.885  0.926  1.03   1.62  0.294 
#>  8       4 crps  Semi-local li… 0.0137  0.0430 0.0644 0.0740 0.103  0.174 0.0465
#>  9       5 crps  AR 3           0.435   0.623  0.951  0.914  1.11   1.42  0.288 
#> 10       5 crps  Semi-local li… 0.0190  0.0508 0.0703 0.0797 0.107  0.148 0.0402
#> 11       6 crps  AR 3           0.487   0.604  0.771  0.843  1.02   1.51  0.304 
#> 12       6 crps  Semi-local li… 0.0182  0.0487 0.0706 0.0797 0.103  0.182 0.0477
#> 13       7 crps  AR 3           0.495   0.746  0.861  0.905  1.06   1.45  0.275 
#> 14       7 crps  Semi-local li… 0.0221  0.0387 0.0646 0.0781 0.0903 0.187 0.0505

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model           bottom lower median   mean upper   top     sd
#>   <chr>      <chr> <chr>            <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl>  <dbl>
#> 1 Region 1   logs  AR 3              5.98  6.69   7.30 Inf    10.7  Inf   Inf   
#> 2 Region 1   logs  Semi-local lin…   3.41  4.20   4.97   9.12  8.17  30.4  15.5 
#> 3 Region 2   logs  AR 3              5.73  6.75   7.23 Inf    11.3  Inf   Inf   
#> 4 Region 2   logs  Semi-local lin…   3.56  4.23   5.10   8.54  8.03  33.4   9.40