Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 10 x 9
#>    score   model        bottom   lower  median     mean   upper     top       sd
#>    <chr>   <chr>         <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1 bias    AR 3        0        0       0.3      0.403   0.9      1       0.408 
#>  2 bias    Semi-loca…  0        0.100   0.3      0.405   0.7      1       0.349 
#>  3 crps    AR 3        0.00816  0.0251  0.0526   0.0829  0.104    0.437   0.102 
#>  4 crps    Semi-loca…  0.00800  0.0235  0.0482   0.0590  0.0876   0.153   0.0426
#>  5 dss     AR 3       -7.40    -5.09   -2.14    12.5     6.30   118.     70.8   
#>  6 dss     Semi-loca… -7.39    -5.26   -3.66    -2.98   -2.02     5.63    4.48  
#>  7 logs    AR 3       -2.65    -1.58   -0.402  Inf       3.92   159.    Inf     
#>  8 logs    Semi-loca… -2.71    -1.68   -0.988   -0.199  -0.242    5.03    5.88  
#>  9 sharpn… AR 3        0.00848  0.0201  0.0307   0.0642  0.0517   0.577   0.163 
#> 10 sharpn… Semi-loca…  0.0140   0.0352  0.0578   0.0659  0.0880   0.169   0.0417

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model          bottom  lower median   mean  upper    top     sd
#>      <int> <chr> <chr>           <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3          0.00683 0.0148 0.0266 0.0367 0.0485 0.109  0.0317
#>  2       1 crps  Semi-local l… 0.00554 0.0105 0.0141 0.0250 0.0232 0.117  0.0277
#>  3       2 crps  AR 3          0.00862 0.0211 0.0401 0.0541 0.0691 0.192  0.0500
#>  4       2 crps  Semi-local l… 0.00749 0.0147 0.0294 0.0379 0.0548 0.0962 0.0294
#>  5       3 crps  AR 3          0.0126  0.0305 0.0556 0.0778 0.103  0.304  0.0757
#>  6       3 crps  Semi-local l… 0.0101  0.0254 0.0376 0.0570 0.0778 0.190  0.0468
#>  7       4 crps  AR 3          0.0177  0.0396 0.0686 0.0899 0.108  0.329  0.0851
#>  8       4 crps  Semi-local l… 0.0167  0.0343 0.0696 0.0739 0.102  0.157  0.0426
#>  9       5 crps  AR 3          0.0194  0.0318 0.0752 0.101  0.122  0.403  0.107 
#> 10       5 crps  Semi-local l… 0.0267  0.0512 0.0795 0.0820 0.111  0.139  0.0361
#> 11       6 crps  AR 3          0.0129  0.0376 0.0820 0.122  0.154  0.553  0.144 
#> 12       6 crps  Semi-local l… 0.0259  0.0467 0.0856 0.0777 0.100  0.137  0.0354
#> 13       7 crps  AR 3          0.00529 0.0260 0.0809 0.130  0.160  0.618  0.171 
#> 14       7 crps  Semi-local l… 0.0293  0.0444 0.0672 0.0765 0.106  0.163  0.0424

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model             bottom lower median  mean upper   top    sd
#>   <chr>      <chr> <chr>              <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 3                3.16  4.32   4.93 12.9   8.06  74.5  29.6
#> 2 Region 1   logs  Semi-local linea…   3.37  4.28   5.30 11.1   6.18  65.5  25.8
#> 3 Region 2   logs  AR 3                3.34  4.33   5.28 13.1  10.8   78.1  24.2
#> 4 Region 2   logs  Semi-local linea…   3.42  4.40   5.00  8.76  6.48  42.8  13.8