Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 18 x 9
#>    score  model      bottom    lower   median     mean   upper      top       sd
#>    <chr>  <chr>       <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>    <dbl>
#>  1 bias   AR 3     -1.00e+0 -8.00e-1 -6.00e-1  -0.574  -0.400   -0.0450   0.238 
#>  2 bias   Semi-lo… -1.00e+0 -8.00e-1 -4.00e-1  -0.18    0.4      1        0.713 
#>  3 calib… AR 3      8.57e-5  8.57e-5  2.86e-3   0.0261  0.0151   0.238    0.0627
#>  4 calib… Semi-lo…  8.57e-5  8.57e-5  1.20e-4   0.109   0.0223   0.985    0.258 
#>  5 ci     AR 3      1.77e+0  2.81e+0  3.49e+0   5.11    4.66    23.2      6.97  
#>  6 ci     Semi-lo…  6.71e-2  1.72e-1  3.09e-1   0.865   1.15     4.37     1.16  
#>  7 crps   AR 3      4.25e-1  7.08e-1  8.91e-1   0.898   1.07     1.53     0.280 
#>  8 crps   Semi-lo…  7.68e-3  2.34e-2  4.47e-2   0.0575  0.0883   0.163    0.0434
#>  9 dss    AR 3      4.54e-1  1.03e+0  1.28e+0   2.22    1.85     7.51     5.26  
#> 10 dss    Semi-lo… -7.19e+0 -5.21e+0 -3.81e+0  -2.74   -2.04     8.86     4.07  
#> 11 iqr    AR 3      1.71e+0  2.38e+0  3.40e+0   3.86    4.98     7.79     1.73  
#> 12 iqr    Semi-lo…  2.63e-2  9.02e-2  1.96e-1   0.258   0.390    0.731    0.206 
#> 13 logs   AR 3      1.33e+0  1.70e+0  1.98e+0 Inf       2.44   Inf      Inf     
#> 14 logs   Semi-lo… -2.62e+0 -1.67e+0 -1.02e+0  -0.363  -0.133    7.58     2.31  
#> 15 median AR 3      6.53e-1  2.56e+0  3.28e+0   3.03    3.80     4.21     0.938 
#> 16 median Semi-lo…  5.33e-3  5.69e-2  1.27e-1   0.153   0.236    0.388    0.115 
#> 17 sharp… AR 3      0.       0.       2.41e-1   0.479   0.834    1.78     0.568 
#> 18 sharp… Semi-lo…  1.60e-2  3.06e-2  4.81e-2   0.0557  0.0695   0.140    0.0337

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model          bottom  lower median   mean  upper    top     sd
#>      <int> <chr> <chr>           <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3          0.447   0.760  0.930  0.933  1.10   1.55   0.300 
#>  2       1 crps  Semi-local l… 0.00495 0.0109 0.0165 0.0243 0.0303 0.0768 0.0208
#>  3       2 crps  AR 3          0.410   0.696  0.900  0.889  1.10   1.45   0.289 
#>  4       2 crps  Semi-local l… 0.0103  0.0178 0.0262 0.0388 0.0577 0.103  0.0291
#>  5       3 crps  AR 3          0.467   0.654  0.775  0.777  0.923  1.10   0.187 
#>  6       3 crps  Semi-local l… 0.00763 0.0270 0.0425 0.0587 0.0896 0.165  0.0427
#>  7       4 crps  AR 3          0.584   0.715  0.886  0.901  1.04   1.33   0.225 
#>  8       4 crps  Semi-local l… 0.0117  0.0381 0.0685 0.0753 0.0999 0.213  0.0515
#>  9       5 crps  AR 3          0.496   0.658  0.825  0.879  1.05   1.45   0.289 
#> 10       5 crps  Semi-local l… 0.0210  0.0540 0.0831 0.0815 0.104  0.157  0.0395
#> 11       6 crps  AR 3          0.367   0.906  1.02   1.01   1.17   1.50   0.280 
#> 12       6 crps  Semi-local l… 0.0195  0.0362 0.0679 0.0707 0.0988 0.139  0.0397
#> 13       7 crps  AR 3          0.437   0.674  0.866  0.916  1.11   1.66   0.356 
#> 14       7 crps  Semi-local l… 0.0162  0.0370 0.0531 0.0678 0.0834 0.160  0.0472

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model            bottom lower median   mean upper   top    sd
#>   <chr>      <chr> <chr>             <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 3               5.70  6.53   6.98 Inf     7.91 Inf   Inf  
#> 2 Region 1   logs  Semi-local line…   3.37  4.17   4.93  10.5   6.68  28.1  27.8
#> 3 Region 2   logs  AR 3               5.66  6.74   7.31 Inf     9.35 Inf   Inf  
#> 4 Region 2   logs  Semi-local line…   3.48  4.26   5.01   9.02  7.85  41.3  10.8