Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> Warning: attributes are not identical across measure variables;
#> they will be dropped

#> # A tibble: 18 x 9
#>    score  model      bottom    lower   median     mean    upper     top       sd
#>    <chr>  <chr>       <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>
#>  1 bias   AR 3     -1.00e+0 -1.00e+0 -6.00e-1  -0.320   2.00e-1   1       0.725 
#>  2 bias   Semi-lo… -1.00e+0 -1.00e+0 -6.00e-1  -0.284   2.00e-1   1       0.704 
#>  3 calib… AR 3      8.57e-5  8.57e-5  8.57e-5   0.0276  2.00e-4   0.483   0.0995
#>  4 calib… Semi-lo…  8.57e-5  8.57e-5  8.57e-5   0.0476  3.00e-4   0.900   0.168 
#>  5 ci     AR 3      7.59e-2  1.81e-1  5.79e-1   1.49    2.41e+0   5.58    1.67  
#>  6 ci     Semi-lo…  7.05e-2  1.79e-1  3.18e-1   0.820   9.93e-1   3.81    1.11  
#>  7 crps   AR 3      8.07e-3  2.81e-2  5.57e-2   0.0998  1.05e-1   0.602   0.150 
#>  8 crps   Semi-lo…  7.02e-3  2.65e-2  5.17e-2   0.0625  9.29e-2   0.174   0.0461
#>  9 dss    AR 3     -7.07e+0 -4.73e+0 -2.87e+0   9.33    9.68e-1 118.     58.4   
#> 10 dss    Semi-lo… -7.14e+0 -5.01e+0 -3.65e+0  -2.98   -2.05e+0   5.01    4.59  
#> 11 iqr    AR 3      3.00e-2  1.12e-1  2.42e-1   0.440   4.51e-1   2.81    0.703 
#> 12 iqr    Semi-lo…  2.78e-2  1.10e-1  2.28e-1   0.281   4.16e-1   0.763   0.212 
#> 13 logs   AR 3     -2.70e+0 -1.58e+0 -7.17e-1 Inf       1.63e+0 249.    NaN     
#> 14 logs   Semi-lo… -2.67e+0 -1.56e+0 -9.05e-1   0.239  -2.07e-1   5.11    9.60  
#> 15 median AR 3      5.59e-3  6.40e-2  1.65e-1   0.265   2.49e-1   1.75    0.455 
#> 16 median Semi-lo…  5.62e-3  6.95e-2  1.49e-1   0.171   2.60e-1   0.490   0.126 
#> 17 sharp… AR 3      8.23e-3  2.46e-2  4.17e-2   0.101   6.75e-2   0.885   0.220 
#> 18 sharp… Semi-lo…  1.43e-2  3.37e-2  6.03e-2   0.0661  8.36e-2   0.167   0.0409

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> Warning: attributes are not identical across measure variables;
#> they will be dropped

#> # A tibble: 14 x 10
#>    horizon score model          bottom  lower median   mean  upper    top     sd
#>      <int> <chr> <chr>           <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3          0.00740 0.0101 0.0206 0.0405 0.0364 0.305  0.0693
#>  2       1 crps  Semi-local l… 0.00494 0.0103 0.0171 0.0244 0.0284 0.0807 0.0214
#>  3       2 crps  AR 3          0.00899 0.0239 0.0371 0.0684 0.0558 0.398  0.114 
#>  4       2 crps  Semi-local l… 0.00672 0.0131 0.0324 0.0383 0.0489 0.115  0.0296
#>  5       3 crps  AR 3          0.0114  0.0256 0.0611 0.103  0.0917 0.588  0.188 
#>  6       3 crps  Semi-local l… 0.0106  0.0270 0.0578 0.0586 0.0771 0.138  0.0369
#>  7       4 crps  AR 3          0.0187  0.0541 0.0804 0.112  0.110  0.517  0.139 
#>  8       4 crps  Semi-local l… 0.0184  0.0490 0.0682 0.0768 0.0997 0.201  0.0461
#>  9       5 crps  AR 3          0.0307  0.0582 0.0910 0.140  0.132  0.705  0.192 
#> 10       5 crps  Semi-local l… 0.0350  0.0453 0.0907 0.0879 0.124  0.156  0.0415
#> 11       6 crps  AR 3          0.0144  0.0451 0.0972 0.128  0.147  0.584  0.148 
#> 12       6 crps  Semi-local l… 0.0243  0.0443 0.0791 0.0856 0.120  0.168  0.0467
#> 13       7 crps  AR 3          0.00800 0.0318 0.0743 0.135  0.162  0.603  0.165 
#> 14       7 crps  Semi-local l… 0.0242  0.0433 0.0750 0.0870 0.109  0.196  0.0546

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> Warning: attributes are not identical across measure variables;
#> they will be dropped

#> # A tibble: 4 x 10
#>   timeseries score model             bottom lower median  mean upper   top    sd
#>   <chr>      <chr> <chr>              <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 3                3.45  4.17   5.05 12.0   9.43  66.5 20.0 
#> 2 Region 1   logs  Semi-local linea…   3.45  4.08   4.66  8.89  7.83  39.2 15.3 
#> 3 Region 2   logs  AR 3                3.37  3.89   5.02  8.26  8.79  38.6  8.52
#> 4 Region 2   logs  Semi-local linea…   3.43  4.12   5.07  8.13  8.00  29.5  8.77