Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 18 x 9
#>    score  model      bottom    lower   median     mean    upper     top       sd
#>    <chr>  <chr>       <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>
#>  1 bias   AR 3     -1.00e+0 -8.00e-1 -6.00e-1  -0.568  -0.400     0       0.259 
#>  2 bias   Semi-lo… -1.00e+0 -8.00e-1 -4.00e-1  -0.228   0.4       1       0.697 
#>  3 calib… AR 3      8.57e-5  8.57e-5  9.49e-4   0.0325  0.0144    0.286   0.104 
#>  4 calib… Semi-lo…  8.57e-5  8.57e-5  9.29e-5   0.0437  0.00930   0.562   0.116 
#>  5 ci     AR 3      1.88e+0  2.82e+0  3.77e+0   5.27    5.00     25.7     5.59  
#>  6 ci     Semi-lo…  7.16e-2  1.69e-1  3.39e-1   0.770   0.897     4.11    1.00  
#>  7 crps   AR 3      4.42e-1  6.72e-1  8.35e-1   0.882   1.08      1.47    0.281 
#>  8 crps   Semi-lo…  8.06e-3  2.41e-2  5.18e-2   0.0618  0.0893    0.157   0.0433
#>  9 dss    AR 3      4.98e-1  9.62e-1  1.31e+0   2.03    1.76     10.8     3.17  
#> 10 dss    Semi-lo… -7.15e+0 -5.08e+0 -3.48e+0  -3.19   -1.83      3.62    2.74  
#> 11 iqr    AR 3      1.71e+0  2.35e+0  3.23e+0   3.75    4.82      7.95    1.72  
#> 12 iqr    Semi-lo…  3.24e-2  1.03e-1  2.32e-1   0.280   0.424     0.725   0.204 
#> 13 logs   AR 3      1.34e+0  1.72e+0  2.01e+0 Inf       2.46    Inf     Inf     
#> 14 logs   Semi-lo… -2.54e+0 -1.53e+0 -8.48e-1  -0.653  -0.191     3.17    1.62  
#> 15 median AR 3      5.98e-1  2.46e+0  3.27e+0   2.97    3.73      4.21    0.996 
#> 16 median Semi-lo…  7.81e-3  6.39e-2  1.61e-1   0.169   0.252     0.443   0.121 
#> 17 sharp… AR 3      0.       0.       2.89e-1   0.517   0.870     2.13    0.637 
#> 18 sharp… Semi-lo…  1.91e-2  3.63e-2  5.07e-2   0.0614  0.0794    0.159   0.0357

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model         bottom   lower median   mean  upper    top     sd
#>      <int> <chr> <chr>          <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3         0.413   0.712   0.944  0.921  1.15   1.40   0.314 
#>  2       1 crps  Semi-local … 0.00681 0.00991 0.0172 0.0249 0.0314 0.0863 0.0219
#>  3       2 crps  AR 3         0.495   0.668   0.910  0.919  1.04   1.47   0.293 
#>  4       2 crps  Semi-local … 0.00680 0.0173  0.0361 0.0399 0.0531 0.0899 0.0262
#>  5       3 crps  AR 3         0.470   0.740   0.864  0.872  1.07   1.24   0.232 
#>  6       3 crps  Semi-local … 0.0141  0.0232  0.0479 0.0588 0.0903 0.139  0.0384
#>  7       4 crps  AR 3         0.483   0.685   0.853  0.915  1.12   1.48   0.291 
#>  8       4 crps  Semi-local … 0.0130  0.0500  0.0695 0.0758 0.0963 0.169  0.0407
#>  9       5 crps  AR 3         0.476   0.721   0.869  0.874  0.990  1.48   0.283 
#> 10       5 crps  Semi-local … 0.0232  0.0620  0.0826 0.0864 0.114  0.151  0.0368
#> 11       6 crps  AR 3         0.440   0.647   0.731  0.788  0.841  1.46   0.278 
#> 12       6 crps  Semi-local … 0.0230  0.0379  0.0780 0.0811 0.109  0.153  0.0435
#> 13       7 crps  AR 3         0.468   0.672   0.777  0.858  1.06   1.34   0.264 
#> 14       7 crps  Semi-local … 0.0275  0.0368  0.0709 0.0851 0.122  0.182  0.0526

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model            bottom lower median   mean upper   top    sd
#>   <chr>      <chr> <chr>             <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 3               5.90  6.57   7.11 Inf     7.68 Inf   Inf  
#> 2 Region 1   logs  Semi-local line…   3.33  4.14   4.86   8.55  6.84  34.4  14.6
#> 3 Region 2   logs  AR 3               5.63  6.77   7.33 Inf     9.87 Inf   Inf  
#> 4 Region 2   logs  Semi-local line…   3.19  4.06   5.00   9.74  7.19  54.3  18.5