Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 10 x 9
#>    score   model        bottom   lower  median     mean   upper     top       sd
#>    <chr>   <chr>         <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1 bias    AR 3        0        0       0.100    0.291   0.6      1       0.355 
#>  2 bias    Semi-loca…  0        0.100   0.3      0.390   0.8      1       0.359 
#>  3 crps    AR 3        0.00803  0.0287  0.0581   0.107   0.109    0.660   0.191 
#>  4 crps    Semi-loca…  0.00949  0.0236  0.0513   0.0602  0.0936   0.150   0.0417
#>  5 dss     AR 3       -7.23    -4.68   -2.50    28.6     7.81   219.    131.    
#>  6 dss     Semi-loca… -7.09    -5.20   -3.59    -2.77   -1.83     7.75    4.58  
#>  7 logs    AR 3       -2.58    -1.58   -0.377  Inf       9.55   258.    Inf     
#>  8 logs    Semi-loca… -2.48    -1.60   -0.986    0.0966 -0.142   11.7     5.27  
#>  9 sharpn… AR 3        0.00772  0.0170  0.0343   0.0564  0.0554   0.207   0.125 
#> 10 sharpn… Semi-loca…  0.0145   0.0339  0.0523   0.0630  0.0804   0.160   0.0391

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model          bottom  lower median   mean  upper    top     sd
#>      <int> <chr> <chr>           <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3          0.00495 0.0152 0.0267 0.0435 0.0419 0.164  0.0841
#>  2       1 crps  Semi-local l… 0.00621 0.0106 0.0188 0.0262 0.0375 0.0700 0.0204
#>  3       2 crps  AR 3          0.00810 0.0204 0.0435 0.0636 0.0767 0.222  0.0965
#>  4       2 crps  Semi-local l… 0.0114  0.0174 0.0341 0.0432 0.0524 0.125  0.0322
#>  5       3 crps  AR 3          0.0227  0.0356 0.0665 0.0989 0.0932 0.398  0.186 
#>  6       3 crps  Semi-local l… 0.0150  0.0314 0.0548 0.0612 0.0800 0.159  0.0403
#>  7       4 crps  AR 3          0.0156  0.0554 0.0883 0.117  0.131  0.429  0.163 
#>  8       4 crps  Semi-local l… 0.0145  0.0411 0.0712 0.0748 0.0974 0.165  0.0417
#>  9       5 crps  AR 3          0.0148  0.0391 0.0976 0.135  0.148  0.564  0.209 
#> 10       5 crps  Semi-local l… 0.0143  0.0569 0.0825 0.0808 0.106  0.150  0.0373
#> 11       6 crps  AR 3          0.0149  0.0298 0.101  0.163  0.176  0.785  0.285 
#> 12       6 crps  Semi-local l… 0.0168  0.0282 0.0766 0.0735 0.103  0.146  0.0447
#> 13       7 crps  AR 3          0.0124  0.0368 0.112  0.171  0.170  0.791  0.271 
#> 14       7 crps  Semi-local l… 0.0186  0.0404 0.0645 0.0772 0.119  0.142  0.0445

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model           bottom lower median   mean upper   top     sd
#>   <chr>      <chr> <chr>            <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl>  <dbl>
#> 1 Region 1   logs  AR 3              3.30  4.33   4.95   9.12 11.3   31.3   8.87
#> 2 Region 1   logs  Semi-local lin…   3.26  4.42   5.31   9.59  7.88  45.8  14.9 
#> 3 Region 2   logs  AR 3              3.51  4.58   5.30 Inf     9.93  94.2 Inf   
#> 4 Region 2   logs  Semi-local lin…   3.19  4.42   5.13  13.1   7.47  87.4  40.3