Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Dummy data
observations <- data.frame(rt = 1:20,
                            date = as.Date("2020-01-01")
                           + lubridate::days(1:20))

timeseries <- list(observations, observations)
names(timeseries) <- c("Region 1", "Region 2")

## List of forecasting bsts models wrapped in functions.
models <- list("Sparse AR" = function(ss, y){bsts::AddAutoAr(ss, y = y, lags = 7)},
               "Semi-local linear trend" = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)})

## Compare models
evaluations <- compare_timeseries(timeseries, models,
                                  horizon = 7, samples = 10)

scores <- evaluations$scores

## Score across the default groups
summarise_scores(scores)

#> # A tibble: 10 x 9
#>    score    model             bottom   lower median    mean  upper   top      sd
#>    <chr>    <chr>              <dbl>   <dbl>  <dbl>   <dbl>  <dbl> <dbl>   <dbl>
#>  1 bias     Semi-local line…  0.100   0.3     0.5     0.478  0.6    0.8    0.181
#>  2 bias     Sparse AR         0       0       0.100   0.111  0.200  0.4    0.116
#>  3 crps     Semi-local line…  0.0489  0.0919  0.138   0.248  0.230  1.09   0.363
#>  4 crps     Sparse AR         0.532   1.41    2.80    3.43   5.15   9.23   2.49 
#>  5 dss      Semi-local line… -3.36   -2.22   -1.46   -1.03  -0.351  2.63   1.75 
#>  6 dss      Sparse AR         0.718   2.68    4.27    8.46   6.32  56.1   14.4  
#>  7 logs     Semi-local line… -0.650  -0.118   0.257   0.444  0.875  2.29   0.785
#>  8 logs     Sparse AR         1.42    2.32    3.00  Inf      3.78  54.5  Inf    
#>  9 sharpne… Semi-local line…  0.145   0.246   0.381   0.641  0.715  2.63   0.675
#> 10 sharpne… Sparse AR         0       1.32    2.27    2.40   3.32   5.68   1.54 

## Also summarise across time horizon
summarise_scores(scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model            bottom  lower median  mean upper    top     sd
#>      <int> <chr> <chr>             <dbl>  <dbl>  <dbl> <dbl> <dbl>  <dbl>  <dbl>
#>  1       1 crps  Semi-local line… 0.0383 0.0577 0.0912 0.113 0.127  0.309 0.0883
#>  2       1 crps  Sparse AR        0.259  0.575  0.698  0.921 0.889  2.37  0.615 
#>  3       2 crps  Semi-local line… 0.0543 0.0691 0.0876 0.150 0.146  0.527 0.164 
#>  4       2 crps  Sparse AR        0.561  1.12   1.53   1.70  1.99   3.80  0.924 
#>  5       3 crps  Semi-local line… 0.0548 0.0961 0.124  0.205 0.182  0.765 0.232 
#>  6       3 crps  Sparse AR        0.988  1.74   2.04   2.57  3.42   5.92  1.39  
#>  7       4 crps  Semi-local line… 0.0676 0.112  0.162  0.260 0.226  0.983 0.312 
#>  8       4 crps  Sparse AR        1.15   2.22   3.25   3.65  5.09   7.27  1.78  
#>  9       5 crps  Semi-local line… 0.0872 0.120  0.145  0.293 0.264  1.27  0.372 
#> 10       5 crps  Sparse AR        2.22   3.31   4.72   4.77  6.27   7.60  1.81  
#> 11       6 crps  Semi-local line… 0.0831 0.143  0.193  0.376 0.374  1.52  0.493 
#> 12       6 crps  Sparse AR        2.57   4.25   4.96   5.62  7.53   9.00  2.12  
#> 13       7 crps  Semi-local line… 0.0995 0.144  0.200  0.452 0.525  2.05  0.655 
#> 14       7 crps  Sparse AR        3.43   5.38   6.18   6.77  8.80  10.2   2.24  

## Instead summarise across region
summarise_scores(scores, "region", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   region  score model        bottom    lower median    mean upper    top      sd
#>   <chr>   <chr> <chr>         <dbl>    <dbl>  <dbl>   <dbl> <dbl>  <dbl>   <dbl>
#> 1 Region… logs  Semi-local … -0.596  0.00471  0.330   0.533 1.12    2.23   0.780
#> 2 Region… logs  Sparse AR     1.37   2.37     2.96  Inf     3.95  136.   Inf    
#> 3 Region… logs  Semi-local … -0.649 -0.163    0.163   0.355 0.679   2.30   0.784
#> 4 Region… logs  Sparse AR     1.46   2.24     3.00  Inf     3.63   46.6  Inf