Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 18 x 9
#>    score  model       bottom    lower   median     mean   upper     top       sd
#>    <chr>  <chr>        <dbl>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1 bias   AR 3      -1.00e+0 -8.00e-1 -6.00e-1  -0.564  -0.400    0       0.263 
#>  2 bias   Semi-lo…  -1.00e+0 -1.00e+0 -2.00e-1  -0.16    0.6      1       0.736 
#>  3 calib… AR 3       8.57e-5  8.57e-5  1.20e-4   0.0337  0.0291   0.269   0.0770
#>  4 calib… Semi-lo…   8.57e-5  8.57e-5  8.57e-5   0.0356  0.0152   0.605   0.113 
#>  5 ci     AR 3       2.02e+0  2.87e+0  3.78e+0   5.75    5.59    29.4     7.01  
#>  6 ci     Semi-lo…   7.60e-2  1.92e-1  3.56e-1   1.04    1.27     4.97    1.41  
#>  7 crps   AR 3       4.74e-1  6.95e-1  8.57e-1   0.921   1.14     1.60    0.303 
#>  8 crps   Semi-lo…   7.91e-3  2.55e-2  4.75e-2   0.0607  0.0884   0.170   0.0452
#>  9 dss    AR 3       5.53e-1  9.87e-1  1.38e+0   2.67    1.92    14.3     8.01  
#> 10 dss    Semi-lo…  -6.74e+0 -4.92e+0 -3.79e+0  -1.30   -1.72    17.5     8.59  
#> 11 iqr    AR 3       1.76e+0  2.32e+0  3.32e+0   3.92    5.33     7.91    1.89  
#> 12 iqr    Semi-lo…   3.58e-2  1.08e-1  2.08e-1   0.275   0.407    0.787   0.204 
#> 13 logs   AR 3       1.40e+0  1.78e+0  1.99e+0 Inf       2.40   Inf     Inf     
#> 14 logs   Semi-lo…  -2.48e+0 -1.56e+0 -9.24e-1   3.11   -0.347   39.4    20.5   
#> 15 median AR 3       6.93e-1  2.43e+0  3.24e+0   3.01    3.81     4.29    0.977 
#> 16 median Semi-lo…   5.38e-3  5.34e-2  1.20e-1   0.152   0.227    0.414   0.117 
#> 17 sharp… AR 3       0.       0.       2.30e-1   0.484   0.773    2.00    0.610 
#> 18 sharp… Semi-lo…   1.10e-2  3.01e-2  5.75e-2   0.0642  0.0890   0.154   0.0415

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model          bottom  lower median   mean  upper    top     sd
#>      <int> <chr> <chr>           <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3          0.474   0.717  0.939  1.01   1.22   1.76   0.384 
#>  2       1 crps  Semi-local l… 0.00642 0.0123 0.0185 0.0261 0.0282 0.0798 0.0221
#>  3       2 crps  AR 3          0.525   0.752  0.885  0.943  1.17   1.42   0.281 
#>  4       2 crps  Semi-local l… 0.00777 0.0162 0.0394 0.0427 0.0593 0.115  0.0311
#>  5       3 crps  AR 3          0.450   0.767  0.886  0.947  1.10   1.52   0.306 
#>  6       3 crps  Semi-local l… 0.0132  0.0303 0.0475 0.0587 0.0841 0.129  0.0379
#>  7       4 crps  AR 3          0.565   0.682  0.845  0.890  0.982  1.52   0.266 
#>  8       4 crps  Semi-local l… 0.0187  0.0368 0.0680 0.0735 0.106  0.176  0.0430
#>  9       5 crps  AR 3          0.486   0.697  0.927  0.939  1.08   1.63   0.321 
#> 10       5 crps  Semi-local l… 0.0248  0.0437 0.0800 0.0806 0.111  0.147  0.0391
#> 11       6 crps  AR 3          0.526   0.756  0.866  0.895  1.02   1.44   0.246 
#> 12       6 crps  Semi-local l… 0.0202  0.0337 0.0656 0.0793 0.118  0.170  0.0530
#> 13       7 crps  AR 3          0.429   0.603  0.749  0.762  0.881  1.23   0.226 
#> 14       7 crps  Semi-local l… 0.0249  0.0350 0.0563 0.0813 0.117  0.197  0.0583

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model            bottom lower median   mean upper   top    sd
#>   <chr>      <chr> <chr>             <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 3               5.54  6.70   7.14 Inf     8.02 Inf   Inf  
#> 2 Region 1   logs  Semi-local line…   3.24  4.10   5.09   9.27  7.42  56.8  13.8
#> 3 Region 2   logs  AR 3               6.01  6.75   7.31 Inf    10.3  Inf   Inf  
#> 4 Region 2   logs  Semi-local line…   3.40  3.99   4.79   9.93  8.61  44.9  13.7