Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 18 x 9
#>    score  model       bottom    lower   median     mean   upper     top       sd
#>    <chr>  <chr>        <dbl>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1 bias   AR 3      -1.00e+0 -8.00e-1 -6.00e-1  -0.548  -0.400    0       0.251 
#>  2 bias   Semi-lo…  -1.00e+0 -8.00e-1 -4.00e-1  -0.177   0.4      1       0.730 
#>  3 calib… AR 3       8.57e-5  8.57e-5  2.86e-3   0.0243  0.0220   0.200   0.0622
#>  4 calib… Semi-lo…   8.57e-5  8.57e-5  9.29e-5   0.0780  0.0857   0.739   0.163 
#>  5 ci     AR 3       2.09e+0  2.89e+0  3.73e+0   6.56    5.23    35.7     9.92  
#>  6 ci     Semi-lo…   7.16e-2  1.73e-1  3.66e-1   0.914   1.14     4.29    1.19  
#>  7 crps   AR 3       4.15e-1  6.60e-1  8.44e-1   0.893   1.07     1.67    0.319 
#>  8 crps   Semi-lo…   7.20e-3  2.50e-2  4.62e-2   0.0602  0.0874   0.171   0.0450
#>  9 dss    AR 3       3.34e-1  9.25e-1  1.28e+0   6.22    1.80    15.4    42.9   
#> 10 dss    Semi-lo…  -7.28e+0 -5.02e+0 -3.56e+0  -2.44   -1.62    11.2     4.58  
#> 11 iqr    AR 3       1.62e+0  2.30e+0  3.28e+0   3.81    4.95     7.79    1.87  
#> 12 iqr    Semi-lo…   3.16e-2  1.01e-1  1.86e-1   0.266   0.412    0.750   0.209 
#> 13 logs   AR 3       1.37e+0  1.76e+0  1.99e+0 Inf       2.62   Inf     Inf     
#> 14 logs   Semi-lo…  -2.64e+0 -1.59e+0 -1.00e+0   0.546  -0.313   11.7     7.73  
#> 15 median AR 3       4.14e-1  2.18e+0  3.14e+0   2.92    3.81     4.28    1.03  
#> 16 median Semi-lo…   7.82e-3  5.45e-2  1.41e-1   0.164   0.249    0.412   0.124 
#> 17 sharp… AR 3       0.       0.       2.49e-1   0.550   0.967    2.02    0.636 
#> 18 sharp… Semi-lo…   1.36e-2  3.33e-2  5.07e-2   0.0608  0.0802   0.153   0.0381

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model          bottom  lower median   mean  upper    top     sd
#>      <int> <chr> <chr>           <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3          0.461   0.693  0.798  0.943  1.16   1.78   0.377 
#>  2       1 crps  Semi-local l… 0.00464 0.0117 0.0152 0.0268 0.0317 0.0911 0.0254
#>  3       2 crps  AR 3          0.393   0.612  0.914  0.923  1.18   1.56   0.376 
#>  4       2 crps  Semi-local l… 0.00980 0.0146 0.0314 0.0408 0.0647 0.107  0.0304
#>  5       3 crps  AR 3          0.565   0.681  0.851  0.888  0.984  1.47   0.272 
#>  6       3 crps  Semi-local l… 0.0125  0.0258 0.0422 0.0587 0.0765 0.151  0.0432
#>  7       4 crps  AR 3          0.417   0.708  0.903  0.918  1.11   1.47   0.304 
#>  8       4 crps  Semi-local l… 0.0155  0.0388 0.0614 0.0758 0.0965 0.202  0.0525
#>  9       5 crps  AR 3          0.391   0.610  0.738  0.806  1.02   1.34   0.284 
#> 10       5 crps  Semi-local l… 0.0294  0.0426 0.0834 0.0843 0.124  0.151  0.0422
#> 11       6 crps  AR 3          0.425   0.659  0.828  0.874  0.953  1.71   0.344 
#> 12       6 crps  Semi-local l… 0.0199  0.0386 0.0802 0.0782 0.109  0.160  0.0428
#> 13       7 crps  AR 3          0.491   0.768  0.887  0.873  0.977  1.37   0.226 
#> 14       7 crps  Semi-local l… 0.0243  0.0373 0.0663 0.0725 0.0878 0.179  0.0469

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model           bottom lower median   mean upper   top     sd
#>   <chr>      <chr> <chr>            <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl>  <dbl>
#> 1 Region 1   logs  AR 3              5.63  6.59   7.12 Inf     7.84 Inf   Inf   
#> 2 Region 1   logs  Semi-local lin…   3.37  4.18   4.95   7.25  6.68  26.5   6.75
#> 3 Region 2   logs  AR 3              5.76  6.86   7.33 Inf     8.95 Inf   Inf   
#> 4 Region 2   logs  Semi-local lin…   3.47  4.16   5.20  10.3   7.84  71.8  17.1