Summarise model forecasting scores

Summarise model forecasting scores

summarise_scores(scores, variables = NULL, sel_scores = NULL)

Arguments

scores
A dataframe of model scores as produced by score_model

variables
A character vector of variables names to group over. By default score type and model is grouped over if present.

sel_scores
A character vector indicating which scores to return information on. Defaults to all scores

Value

A dataframe of summarised scores in a tidy format.

Examples

## Example cases
cases <- EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_cases %>%
    dplyr::mutate(timeseries = "Region 2"))

## Example Rts
rts <- EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 1") %>%
    dplyr::bind_rows(EpiSoon::example_obs_rts %>%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)}, ...)},
               "Semi-local linear trend" =
               function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)})

## Compare models
evaluations <- compare_timeseries(rts, cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = example_serial_interval)

## Score across the default groups
summarise_scores(evaluations$rt_scores)

#> # A tibble: 10 x 9
#>    score   model          bottom   lower  median    mean   upper     top      sd
#>    <chr>   <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 bias    AR 3         -1       -1      -0.8    -0.349   0.200    1      0.738 
#>  2 bias    Semi-local … -1       -0.8    -0.400  -0.166   0.6      1      0.746 
#>  3 crps    AR 3          0.00543  0.0283  0.0575  0.0667  0.103    0.180  0.0470
#>  4 crps    Semi-local …  0.00629  0.0203  0.0494  0.0599  0.0919   0.158  0.0454
#>  5 dss     AR 3         -7.50    -4.67   -2.68    6.00    3.56    83.5   26.7   
#>  6 dss     Semi-local … -7.64    -5.36   -3.82   -2.61   -1.71     9.90   5.13  
#>  7 logs    AR 3         -3.21    -1.50   -0.654  12.7     2.93   111.    52.7   
#>  8 logs    Semi-local … -2.75    -1.75   -0.930   0.256  -0.242   11.0    7.60  
#>  9 sharpn… AR 3          0.00700  0.0208  0.0346  0.0585  0.0587   0.286  0.0899
#> 10 sharpn… Semi-local …  0.0132   0.0301  0.0464  0.0498  0.0647   0.106  0.0275

## Also summarise across time horizon
summarise_scores(evaluations$rt_scores, "horizon", sel_scores = "crps")

#> # A tibble: 14 x 10
#>    horizon score model         bottom   lower median   mean  upper    top     sd
#>      <int> <chr> <chr>          <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1       1 crps  AR 3         0.00473 0.0113  0.0192 0.0288 0.0395 0.0890 0.0237
#>  2       1 crps  Semi-local … 0.00502 0.00860 0.0157 0.0249 0.0349 0.0932 0.0241
#>  3       2 crps  AR 3         0.00304 0.0237  0.0375 0.0440 0.0627 0.0964 0.0285
#>  4       2 crps  Semi-local … 0.00714 0.0132  0.0260 0.0387 0.0488 0.118  0.0342
#>  5       3 crps  AR 3         0.0105  0.0313  0.0725 0.0660 0.0933 0.133  0.0391
#>  6       3 crps  Semi-local … 0.00990 0.0219  0.0454 0.0549 0.0796 0.157  0.0432
#>  7       4 crps  AR 3         0.0175  0.0361  0.0807 0.0741 0.110  0.123  0.0378
#>  8       4 crps  Semi-local … 0.0109  0.0336  0.0818 0.0727 0.0987 0.149  0.0437
#>  9       5 crps  AR 3         0.0209  0.0446  0.0878 0.0817 0.107  0.146  0.0411
#> 10       5 crps  Semi-local … 0.0115  0.0622  0.0854 0.0817 0.109  0.146  0.0390
#> 11       6 crps  AR 3         0.0204  0.0497  0.0827 0.0941 0.140  0.182  0.0533
#> 12       6 crps  Semi-local … 0.0226  0.0486  0.0789 0.0814 0.109  0.172  0.0448
#> 13       7 crps  AR 3         0.0148  0.0539  0.0940 0.102  0.151  0.201  0.0610
#> 14       7 crps  Semi-local … 0.0253  0.0466  0.0723 0.0846 0.110  0.209  0.0518

## Instead summarise across region and summarise case scores
summarise_scores(evaluations$case_scores, "timeseries", sel_scores = "logs")

#> # A tibble: 4 x 10
#>   timeseries score model             bottom lower median  mean upper   top    sd
#>   <chr>      <chr> <chr>              <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Region 1   logs  AR 3                3.41  4.02   4.73  10.0  7.55  53.3  15.9
#> 2 Region 1   logs  Semi-local linea…   3.46  4.21   4.83  10.2  7.96  47.9  19.4
#> 3 Region 2   logs  AR 3                3.12  3.97   4.63  11.1  8.59  76.3  18.3
#> 4 Region 2   logs  Semi-local linea…   3.64  4.16   4.92  10.2  9.06  61.5  14.2